refactor: finish VeilRoot root split
