+++ apps/cocos-client/assets/scripts/VeilRoot.ts
@@ -1,302 +1,34 @@
-import { _decorator, Camera, Canvas, Color, Component, EventMouse, EventTouch, Graphics, input, Input, Label, Layers, Node, sys, UITransform, view } from "cc";
-import { getBuildingUpgradeConfig, getEquipmentDefinition, type EquipmentType } from "./project-shared/index.ts";
+import { _decorator, Component, input, Input } from "cc";
 import {
-  type BattleAction,
-  type LeaderboardEntry,
-  type MatchmakingStatusResponse,
-  type PlayerReportReason,
-  VeilCocosSession,
-  type VeilCocosSessionOptions,
-  type ConnectionEvent,
-  type HeroView,
-  type PlayerTileView,
-  type SessionUpdate,
-  type Vec2
-} from "./VeilCocosSession.ts";
-import { createCocosAccountReviewState, transitionCocosAccountReviewState, type CocosAccountReviewSection, type CocosAccountReviewState } from "./cocos-account-review.ts";
-import {
-  loadCocosCampaignSummary,
-  claimCocosDailyQuest,
-  claimAllCocosMailboxMessages,
-  claimCocosMailboxMessage,
-  submitCocosSupportTicket,
-  type CocosCampaignSummary,
-  type CocosLaunchAnnouncement,
-  type CocosMaintenanceModeSnapshot,
-  confirmCocosAccountRegistration,
-  confirmCocosPasswordRecovery,
-  createFallbackCocosPlayerAccountProfile,
-  deleteCurrentCocosPlayerAccount,
-  createCocosGuestPlayerId,
-  loadCocosAnnouncements,
-  loadCocosBattleReplayHistoryPage,
-  createCocosLobbyPreferences,
-  loadCocosLobbyRooms,
-  loadCocosMaintenanceMode,
-  loadCocosPlayerAccountProfile,
-  loadCocosPlayerAchievementProgress,
-  loadCocosPlayerEventHistory,
-  loadCocosPlayerProgressionSnapshot,
-  loginCocosGuestAuthSession,
-  loginCocosWechatAuthSession,
-  logoutCurrentCocosAuthSession,
-  postCocosPlayerReferral,
-  readPreferredCocosDisplayName,
-  rememberPreferredCocosDisplayName,
-  requestCocosWechatSubscribeConsent,
-  requestCocosAccountRegistration,
-  requestCocosPasswordRecovery,
-  resolveCocosConfigCenterUrl,
-  saveCocosLobbyPreferences,
-  submitCocosSeasonalEventProgress,
-  syncCurrentCocosAuthSession,
-  updateCocosTutorialProgress,
-  type CocosLobbyRoomSummary,
-  type CocosPlayerAccountProfile,
-  type CocosSeasonalEvent
-} from "./cocos-lobby.ts";
-import {
-  loginWithCocosProvider,
-  resolveCocosLoginProviders,
-  resolveCocosLoginRuntimeConfig,
-  type CocosLoginProviderDescriptor,
-  type CocosLoginRuntimeConfig
-} from "./cocos-login-provider.ts";
-import { predictPlayerWorldAction as predictSharedPlayerWorldAction } from "./project-shared/map.ts";
-import { type CocosWorldAction, predictPlayerWorldAction } from "./cocos-prediction.ts";
-import { VeilBattleTransition } from "./VeilBattleTransition.ts";
-import { VeilBattlePanel } from "./VeilBattlePanel.ts";
-import { assignUiLayer } from "./cocos-ui-layer.ts";
-import {
-  buildTimelineEntriesFromUpdate,
-  describeMoveAttemptFeedback,
-  describeSessionActionOutcome,
-  formatSessionActionReason,
-  formatSessionSettlementReason,
-  isSessionSettlementReason
-} from "./cocos-ui-formatters.ts";
-import { buildHeroProgressNotice, type HeroProgressNotice } from "./cocos-hero-progression.ts";
-import { VeilHudPanel, type VeilHudRenderState } from "./VeilHudPanel.ts";
-import { VeilLobbyPanel } from "./VeilLobbyPanel.ts";
-import {
-  startCocosMatchmakingStatusPolling,
-  type CocosMatchmakingPollController
-} from "./cocos-matchmaking.ts";
-import {
-  buildMatchmakingStatusView,
-  type MatchmakingStatusView
-} from "./cocos-matchmaking-status.ts";
-import {
-  buildCocosAccountLifecyclePanelView,
-  type CocosAccountLifecycleDeliveryMode,
-  type CocosAccountLifecycleDraft,
-  type CocosAccountLifecycleKind,
-  type CocosAccountLifecyclePanelView
-} from "./cocos-account-lifecycle.ts";
-import {
-  buildCocosAccountRegistrationPanelView,
-  type CocosAccountRegistrationPanelView,
-  type CocosWechatMinorProtectionSelection
-} from "./cocos-account-registration.ts";
-import { VeilMapBoard } from "./VeilMapBoard.ts";
-import { buildMapFeedbackEntriesFromUpdate, buildObjectPulseEntriesFromUpdate } from "./cocos-map-visuals.ts";
-import { getPlaceholderSpriteAssetUsageSummary } from "./cocos-placeholder-sprites.ts";
-import {
-  detectCocosRuntimePlatform,
-  readCocosRuntimeLaunchSearch,
-  resolveCocosRuntimeCapabilities,
-  type CocosRuntimeCapabilities,
-  type CocosRuntimePlatform
-} from "./cocos-runtime-platform.ts";
-import {
-  bindCocosRuntimeMemoryWarning,
-  formatCocosRuntimeMemoryStatus,
-  readCocosRuntimeMemorySnapshot,
-  triggerCocosRuntimeGc
-} from "./cocos-runtime-memory.ts";
-import {
-  buildCocosProfileNotice,
-  collectProfileNoticeEventIds,
-  shouldRefreshGameplayAccountProfileForEvents
-} from "./cocos-achievements.ts";
-import {
-  readLaunchReferrerId,
-  shareBattleResultForRuntime,
-  shouldOfferBattleResultShare,
-  type WechatSharePayload
-} from "./cocos-share-card.ts";
-import {
-  buildCocosWechatSharePayload,
-  syncCocosWechatShareBridge,
-  type CocosWechatShareRuntimeLike
-} from "./cocos-wechat-share.ts";
-import {
-  readCocosWechatFriendCloudEntries,
-  syncCocosWechatFriendCloudStorage
-} from "./cocos-wechat-social.ts";
-import {
-  clearStoredCocosAuthSession,
-  readStoredCocosAuthSession,
-  resolveCocosLaunchIdentity,
-  type CocosAuthProvider
-} from "./cocos-session-launch.ts";
-import { type ShopProduct } from "./cocos-shop-panel.ts";
-import { resolveCocosClientVersion } from "./cocos-client-version.ts";
-import {
-  type CocosDailyDungeonSummary,
-  type CocosSeasonProgress
-} from "./cocos-progression-panel.ts";
-import { VeilTimelinePanel } from "./VeilTimelinePanel.ts";
-import { VeilProgressionPanel } from "./VeilProgressionPanel.ts";
-import { VeilEquipmentPanel } from "./VeilEquipmentPanel.ts";
-import { VeilCampaignPanel } from "./VeilCampaignPanel.ts";
-import { VeilTutorialOverlay, type TutorialOverlayView } from "./VeilTutorialOverlay.ts";
-import { formatEquipmentActionReason, formatEquipmentSlotLabel } from "./cocos-hero-equipment.ts";
-import { type CocosBattleFeedbackView } from "./cocos-battle-feedback.ts";
-import {
-  createCocosBattlePresentationController,
-  type CocosBattlePresentationState
-} from "./cocos-battle-presentation-controller.ts";
-import { createCocosAudioRuntime } from "./cocos-audio-runtime.ts";
-import {
+  cocosPresentationConfig,
+  createCocosAudioAssetBridge,
+  createCocosAudioRuntime,
   setAssetLoadFailureReporter,
-  subscribeAssetLoadFailures,
-  type AssetLoadFailureEvent
-} from "./cocos-asset-load-resilience.ts";
-import { createCocosAudioAssetBridge } from "./cocos-audio-resources.ts";
-import {
-  applySettingsUpdate,
-  CocosSettingsPanel,
-  createDefaultCocosSettingsView,
-  readPersistedCocosSettings,
-  resolveCocosPrivacyPolicyUrl,
-  writePersistedCocosSettings,
-  type CocosSettingsPanelUpdate,
-  type CocosSettingsPanelView
-} from "./cocos-settings-panel.ts";
-import { cocosPresentationConfig } from "./cocos-presentation-config.ts";
-import { getPixelSpriteLoadStatus, loadPixelSpriteAssets } from "./cocos-pixel-sprites.ts";
-import { type CocosCampaignDialogueState } from "./cocos-campaign-panel.ts";
-import {
-  buildPrimaryClientTelemetryFromUpdate,
-  createPrimaryClientTelemetryEvent,
-  type ClientAnalyticsContext
-} from "./cocos-primary-client-telemetry.ts";
-import {
-  createClientPerfTelemetryMonitorState,
-  evaluateClientPerfTelemetry,
-  readClientPerfRuntimeMetadata,
-  recordClientPerfFrame,
-  type ClientPerfRuntimeMetadata
-} from "./cocos-client-perf-telemetry.ts";
-import { describeAccountAuthFailure, type PrimaryClientTelemetryEvent, type RuntimeDiagnosticsConnectionStatus, type StructuredErrorCode, validateAccountLifecycleConfirm, validateAccountLifecycleRequest, validateAccountPassword, validatePrivacyConsentAccepted } from "@veil/shared/platform";
-import type { TutorialProgressAction } from "@veil/shared/protocol";
-import {
-  createCocosWechatPaymentOrder,
-  requestCocosWechatPayment,
-  verifyCocosWechatPayment,
-  type CocosWechatPaymentRuntimeLike
-} from "./cocos-wechat-payment.ts";
-import { normalizeTutorialStep } from "@veil/shared/progression";
+  subscribeAssetLoadFailures
+} from "./root/deps.ts";
+import { assignVeilRootDefaultState } from "./root/default-state.ts";
+import { veilRootAccountSettingsMethods } from "./root/account-settings.ts";
+import { veilRootBattleSessionUpdateMethods } from "./root/battle-session-update.ts";
+import { veilRootLobbyMatchmakingMethods } from "./root/lobby-matchmaking.ts";
+import { veilRootLobbyProgressionMethods } from "./root/lobby-progression.ts";
+import { veilRootPrefetchSchedulerMethods } from "./root/prefetch-scheduler.ts";
+import { veilRootReportReviewMethods } from "./root/report-review.ts";
+import { disposeCurrentSessionForRoot, connectSessionForRoot, refreshSnapshotForRoot } from "./root/session-lifecycle.ts";
+export {
+  resetVeilRootRuntimeForTests,
+  resolveVeilRootRuntime,
+  setVeilRootRuntimeForTests
+} from "./root/runtime.ts";
+import { veilRootRuntimeIntegrationMethods } from "./root/runtime-integration.ts";
+import { veilRootViewLayoutMethods } from "./root/view-layout.ts";
+import { veilRootWorldBattleActionsMethods } from "./root/world-battle-actions.ts";
 
 const { ccclass, property } = _decorator;
 
-import {
-  ACCOUNT_REVIEW_PAGE_SIZE,
-  ACCOUNT_REVIEW_PANEL_NODE_NAME,
-  BATTLE_FEEDBACK_DURATION_MS,
-  BATTLE_NODE_NAME,
-  bumpLobbyAccountEpochForRoot,
-  bumpSessionEpochForRoot,
-  CAMPAIGN_PANEL_NODE_NAME,
-  cloneSessionUpdate,
-  collapseAdjacentEntries,
-  connectSessionForRoot,
-  createSessionOptionsForRoot,
-  buildBattleSettlementRecoveryStateForRoot,
-  buildHudPresentationStateForRoot,
-  buildHudSessionIndicatorsForRoot,
-  claimGameplayDailyDungeonRunForRoot,
-  claimGameplaySeasonTierForRoot,
-  completeGameplayCampaignMissionForRoot,
-  DEFAULT_MAP_HEIGHT_TILES,
-  DEFAULT_MAP_WIDTH_TILES,
-  describeCampaignErrorForRoot,
-  describeSessionErrorForRoot,
-  disposeCurrentSessionForRoot,
-  EQUIPMENT_PANEL_NODE_NAME,
-  FORCE_UPGRADE_MESSAGE,
-  formatHeroStatBonus,
-  formatResourceKindLabel,
-  formatUpgradeCostLabel,
-  handleForcedUpgradeForRoot,
-  HUD_NODE_NAME,
-  hydrateLaunchIdentityForRoot,
-  isActiveLobbyAccountEpochForRoot,
-  isActiveSessionEpochForRoot,
-  LOBBY_NODE_NAME,
-  MAP_NODE_NAME,
-  openLobbyPvePanelForRoot,
-  purchaseGameplaySeasonPremiumForRoot,
-  refreshSnapshotForRoot,
-  refreshActiveSeasonalEventForRoot,
-  refreshDailyDungeonPanelForRoot,
-  refreshGameplayCampaignForRoot,
-  refreshSeasonProgressForRoot,
-  renderViewForRoot,
-  renderGameplayAccountReviewPanelForRoot,
-  renderGameplayCampaignPanelForRoot,
-  renderGameplayEquipmentPanelForRoot,
-  resolveSelectedGameplayCampaignMissionForRoot,
-  resetSessionViewportForRoot,
-  resolveVeilRootRuntime,
-  SETTINGS_BUTTON_NODE_NAME,
-  SETTINGS_PANEL_NODE_NAME,
-  selectGameplayCampaignMissionForRoot,
-  snapshotSeasonProgressFromProfileForRoot,
-  startGameplayCampaignDialogueForRoot,
-  startGameplayCampaignMissionForRoot,
-  syncGameplayCampaignSelectionForRoot,
-  TIMELINE_NODE_NAME,
-  toggleGameplayAccountReviewPanelForRoot,
-  toggleGameplayBattlePassPanelForRoot,
-  toggleGameplayCampaignPanelForRoot,
-  toggleGameplayDailyDungeonPanelForRoot,
-  toggleGameplayEquipmentPanelForRoot,
-  toggleGameplaySeasonalEventPanelForRoot,
-  TUTORIAL_OVERLAY_NODE_NAME,
-  advanceTutorialFlowForRoot,
-  advanceGameplayCampaignDialogueForRoot,
-  attemptGameplayDailyDungeonFloorForRoot,
-  type BattleSettlementSnapshot,
-  bindGlobalErrorBoundaryForRoot,
-  buildTutorialOverlayViewForRoot,
-  completeTutorialAndFocusCampaignForRoot,
-  createClientAnalyticsContextForRoot,
-  createTelemetryContextForRoot,
-  emitPrimaryClientTelemetryForRoot,
-  type GlobalErrorBoundaryEvent,
-  ensureAnalyticsSessionIdForRoot,
-  type TutorialCampaignGuidance,
-  handleAssetLoadFailureForRoot,
-  handleTutorialPrimaryActionForRoot,
-  maybeEmitExperimentExposureAnalyticsForRoot,
-  maybeEmitQuestCompleteAnalyticsForRoot,
-  maybeEmitShopOpenAnalyticsForRoot,
-  maybeReportSessionRuntimeErrorForRoot,
-  reportClientRuntimeErrorForRoot,
-  resolveTutorialCampaignGuidanceForRoot,
-  resolveTutorialGuidanceMissionForRoot,
-  type VeilRootRuntime,
-  skipTutorialFlowForRoot,
-  submitTutorialProgressForRoot,
-  trackAssetLoadFailureAnalyticsForRoot,
-  trackClientAnalyticsEventForRoot,
-  trackPurchaseInitiatedForRoot
-} from "./root/index.ts";
-
 @ccclass("ProjectVeilRoot")
 export class VeilRoot extends Component {
+  [key: string]: any;
+
   @property
   roomId = "test-room";
 
@@ -324,153 +56,19 @@ export class VeilRoot extends Component {
   @property
   fogPulseIntervalSeconds = 0.8;
 
-  private hudPanel: VeilHudPanel | null = null;
-  private mapBoard: VeilMapBoard | null = null;
-  private battlePanel: VeilBattlePanel | null = null;
-  private timelinePanel: VeilTimelinePanel | null = null;
-  private lobbyPanel: VeilLobbyPanel | null = null;
-  private battleTransition: VeilBattleTransition | null = null;
-  private session: VeilCocosSession | null = null;
-  private lastUpdate: SessionUpdate | null = null;
-  private logLines: string[] = ["Cocos 主客户端已就绪。"];
-  private timelineEntries: string[] = [];
-  private moveInFlight = false;
-  private battleActionInFlight = false;
-  private predictionStatus = "";
-  private inputDebug = "input waiting";
-  private pendingPrediction: SessionUpdate | null = null;
-  private selectedBattleTargetId: string | null = null;
-  private selectedInteractionBuildingId: string | null = null;
-  private battleFeedback: (CocosBattleFeedbackView & { expiresAt: number }) | null = null;
-  private fogPulsePhase = 0;
-  private hudActionBinding = false;
-  private sessionEpoch = 0;
-  private authToken: string | null = null;
-  private authMode: "guest" | "account" = "guest";
-  private authProvider: CocosAuthProvider = "guest";
-  private loginId = "";
-  private privacyConsentAccepted = false;
-  private sessionSource: "remote" | "local" | "manual" | "none" = "none";
-  private levelUpNotice: (HeroProgressNotice & { expiresAt: number }) | null = null;
-  private achievementNotice: ({ title: string; detail: string; expiresAt: number } & { eventId: string }) | null = null;
-  private showLobby = false;
-  private lobbyRooms: CocosLobbyRoomSummary[] = [];
-  private lobbyStatus = "请选择一个房间，或手动输入新的房间 ID。";
-  private lobbyAnnouncements: CocosLaunchAnnouncement[] = [];
-  private lobbyMaintenanceMode: CocosMaintenanceModeSnapshot | null = null;
-  private upgradeRequired = false;
-  private lobbyLoading = false;
-  private lobbyEntering = false;
-  private matchmakingStatus: MatchmakingStatusResponse = { status: "idle" };
-  private matchmakingPollController: CocosMatchmakingPollController | null = null;
-  private matchmakingTimeoutHandle: ReturnType<typeof setTimeout> | null = null;
-  private matchmakingTimeoutMs = 120_000;
-  private matchmakingView: MatchmakingStatusView = buildMatchmakingStatusView({ status: "idle" });
-  private matchmakingJoinInFlight = false;
-  private lobbyLeaderboardEntries: LeaderboardEntry[] = [];
-  private lobbyLeaderboardStatus: "idle" | "loading" | "ready" | "error" = "idle";
-  private lobbyLeaderboardError: string | null = null;
-  private lobbyAccountProfile: CocosPlayerAccountProfile = createFallbackCocosPlayerAccountProfile("player-1", "test-room");
-  private tutorialOverlay: VeilTutorialOverlay | null = null;
-  private tutorialProgressInFlight = false;
-  private lobbyShopProducts: ShopProduct[] = [];
-  private lobbyShopLoading = false;
-  private lobbyShopStatus = "可用商品会在这里显示。";
-  private pendingShopProductId: string | null = null;
-  private seasonProgress: CocosSeasonProgress | null = null;
-  private seasonProgressStatus = "赛季进度待同步。";
-  private dailyDungeonSummary: CocosDailyDungeonSummary | null = null;
-  private dailyDungeonStatus = "每日地城待同步。";
-  private dailyDungeonLoading = false;
-  private pendingDailyDungeonFloor: number | null = null;
-  private pendingDailyDungeonClaimRunId: string | null = null;
-  private activeSeasonalEvent: CocosSeasonalEvent | null = null;
-  private seasonalEventStatus = "赛季活动待同步。";
-  private gameplaySeasonalEventPanelOpen = false;
-  private pendingSeasonalEventBattleIds = new Set<string>();
-  private pendingSeasonClaimTier: number | null = null;
-  private seasonPremiumPurchaseInFlight = false;
-  private dailyQuestClaimingId: string | null = null;
-  private mailboxClaimingMessageId: string | null = null;
-  private mailboxClaimAllInFlight = false;
-  private lobbyAccountReviewState: CocosAccountReviewState = createCocosAccountReviewState(this.lobbyAccountProfile);
-  private lobbyAccountEpoch = 0;
-  private gameplayAccountRefreshInFlight = false;
-  private gameplayAccountReviewPanel: VeilProgressionPanel | null = null;
-  private gameplayAccountReviewPanelOpen = false;
-  private gameplayBattlePassPanelOpen = false;
-  private gameplayDailyDungeonPanelOpen = false;
-  private gameplayEquipmentPanel: VeilEquipmentPanel | null = null;
-  private gameplayEquipmentPanelOpen = false;
-  private gameplayCampaignPanel: VeilCampaignPanel | null = null;
-  private gameplayCampaignPanelOpen = false;
-  private gameplayCampaign: CocosCampaignSummary | null = null;
-  private gameplayCampaignSelectedMissionId: string | null = null;
-  private gameplayCampaignActiveMissionId: string | null = null;
-  private gameplayCampaignDialogue: CocosCampaignDialogueState | null = null;
-  private gameplayCampaignStatus = "战役面板待同步。";
-  private gameplayCampaignLoading = false;
-  private gameplayCampaignPendingAction: "start" | "complete" | null = null;
-  private settingsPanel: CocosSettingsPanel | null = null;
-  private settingsView: CocosSettingsPanelView = createDefaultCocosSettingsView();
-  private supportTicketSubmittingCategory: "bug" | "payment" | "account" | null = null;
-  private activeAccountFlow: CocosAccountLifecycleKind | null = null;
-  private registrationDisplayName = "";
-  private registrationToken = "";
-  private registrationPassword = "";
-  private registrationDeliveryMode: CocosAccountLifecycleDeliveryMode = "idle";
-  private registrationExpiresAt = "";
-  private wechatMinorProtectionSelection: CocosWechatMinorProtectionSelection = "unknown";
-  private recoveryToken = "";
-  private recoveryPassword = "";
-  private recoveryDeliveryMode: CocosAccountLifecycleDeliveryMode = "idle";
-  private recoveryExpiresAt = "";
-  private runtimePlatform: CocosRuntimePlatform = "unknown";
-  private runtimeCapabilities: CocosRuntimeCapabilities = resolveCocosRuntimeCapabilities("unknown");
-  private loginRuntimeConfig: CocosLoginRuntimeConfig = resolveCocosLoginRuntimeConfig();
-  private loginProviders: CocosLoginProviderDescriptor[] = [];
-  private audioRuntime = createCocosAudioRuntime(cocosPresentationConfig.audio);
-  private pendingPixelSpriteGroups = new Set<"boot" | "battle">();
-  private seenProfileNoticeEventIds = new Set<string>();
-  private wechatShareStatus = "分享功能仅在微信小游戏可用。";
-  private wechatShareAvailable = false;
-  private runtimeMemoryNotice = "";
-  private diagnosticsConnectionStatus: RuntimeDiagnosticsConnectionStatus = "connecting";
-  private lastRoomUpdateSource: string | null = null;
-  private lastRoomUpdateReason: string | null = null;
-  private lastRoomUpdateAtMs: number | null = null;
-  private primaryClientTelemetry: PrimaryClientTelemetryEvent[] = [];
-  private analyticsSessionId: string | null = null;
-  private emittedExperimentExposureKeys = new Set<string>();
-  private emittedShopOpenSessionId: string | null = null;
-  private clientPerfTelemetry = createClientPerfTelemetryMonitorState();
-  private clientPerfRuntimeMetadata: ClientPerfRuntimeMetadata = {
-    deviceModel: "unknown",
-    wechatVersion: "unknown"
-  };
-  private stopRuntimeMemoryWarnings: (() => void) | null = null;
-  private stopAssetLoadFailureSubscription: (() => void) | null = null;
-  private battlePresentation = createCocosBattlePresentationController();
-  private lastBattleSettlementSnapshot: BattleSettlementSnapshot | null = null;
-  private reportDialogOpen = false;
-  private reportSubmitting = false;
-  private reportStatusMessage: string | null = null;
-  private surrenderDialogOpen = false;
-  private surrenderSubmitting = false;
-  private surrenderStatusMessage: string | null = null;
-  private launchReferrerId: string | null = null;
-  private lastReferralClaimKey: string | null = null;
-  private lastAssetFailureNoticeKey: string | null = null;
-  private stopGlobalErrorBoundary: (() => void) | null = null;
+  constructor() {
+    super();
+    assignVeilRootDefaultState(this);
+  }
 
   onLoad(): void {
     this.audioRuntime.dispose();
     this.stopAssetLoadFailureSubscription?.();
     this.stopGlobalErrorBoundary?.();
-    setAssetLoadFailureReporter((event) => {
+    setAssetLoadFailureReporter((event: unknown) => {
       this.trackAssetLoadFailureAnalytics(event);
     });
-    this.stopAssetLoadFailureSubscription = subscribeAssetLoadFailures((event) => {
+    this.stopAssetLoadFailureSubscription = subscribeAssetLoadFailures((event: unknown) => {
       this.handleAssetLoadFailure(event);
     });
     this.audioRuntime = createCocosAudioRuntime(cocosPresentationConfig.audio, {
@@ -538,4770 +136,24 @@ export class VeilRoot extends Component {
   async refreshSnapshot(): Promise<void> {
     await refreshSnapshotForRoot(this as unknown as Record<string, any>);
   }
-
-  async advanceDay(): Promise<void> {
-    if (this.moveInFlight || this.battleActionInFlight) {
-      return;
-    }
-
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.pushLog("战斗中无法推进天数。");
-      this.predictionStatus = "战斗中无法推进天数。";
-      this.renderView();
-      return;
-    }
-
-    this.predictionStatus = "正在推进到下一天...";
-    this.moveInFlight = true;
-    this.renderView();
-
-    try {
-      await this.applySessionUpdate(await this.session.endDay());
-      this.pushLog("已推进到下一天。");
-    } catch (error) {
-      this.maybeReportSessionRuntimeError(error, "end_day");
-      const failureMessage = this.describeSessionError(error, "推进天数失败。");
-      if (error instanceof Error && error.message === "upgrade_required") {
-        await this.handleForcedUpgrade(failureMessage);
-        return;
-      }
-      this.pushLog(failureMessage);
-      this.predictionStatus = failureMessage;
-    } finally {
-      this.moveInFlight = false;
-    }
-
-    this.renderView();
-  }
-
-  async learnHeroSkill(skillId: string): Promise<void> {
-    if (this.moveInFlight || this.battleActionInFlight) {
-      return;
-    }
-
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    const hero = this.activeHero();
-    if (!hero) {
-      this.pushLog("当前快照里没有可控制的英雄。");
-      this.renderView();
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.pushLog("战斗中无法调整技能树。");
-      this.predictionStatus = "战斗中无法调整技能树。";
-      this.renderView();
-      return;
-    }
-
-    this.moveInFlight = true;
-    this.predictionStatus = `正在学习技能 ${skillId}...`;
-    this.pushLog(`正在为 ${hero.name} 学习技能 ${skillId}...`);
-    this.renderView();
-
-    try {
-      const update = await this.session.learnSkill(hero.id, skillId);
-      await this.applySessionUpdate(update);
-      this.pushSessionActionOutcome(update, {
-        successMessage: "技能学习已结算。",
-        rejectedLabel: "技能学习"
-      });
-    } catch (error) {
-      const failureMessage = this.describeSessionError(error, "技能学习失败。");
-      if (error instanceof Error && error.message === "upgrade_required") {
-        await this.handleForcedUpgrade(failureMessage);
-        return;
-      }
-      this.pushLog(failureMessage);
-      this.predictionStatus = failureMessage;
-    } finally {
-      this.moveInFlight = false;
-    }
-
-    this.renderView();
-  }
-
-  async equipHeroItem(slot: EquipmentType, equipmentId: string): Promise<void> {
-    if (this.moveInFlight || this.battleActionInFlight) {
-      return;
-    }
-
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    const hero = this.activeHero();
-    if (!hero) {
-      this.pushLog("当前快照里没有可控制的英雄。");
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(), {
-          category: "inventory",
-          checkpoint: "equipment.equip.rejected",
-          status: "blocked",
-          detail: "Equip request ignored because no controlled hero is present.",
-          reason: "no_controlled_hero"
-        })
-      );
-      this.renderView();
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.pushLog("战斗中无法调整装备。");
-      this.predictionStatus = "战斗中无法调整装备。";
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(hero.id), {
-          category: "inventory",
-          checkpoint: "equipment.equip.rejected",
-          status: "blocked",
-          detail: "Equip request rejected because the client is currently in battle.",
-          reason: "in_battle"
-        })
-      );
-      this.renderView();
-      return;
-    }
-
-    const itemName = getEquipmentDefinition(equipmentId)?.name ?? equipmentId;
-    this.moveInFlight = true;
-    this.predictionStatus = `正在装备 ${itemName}...`;
-    this.pushLog(`正在为 ${hero.name} 装备 ${itemName}...`);
-    this.applyPrediction(
-      {
-        type: "hero.equip",
-        heroId: hero.id,
-        slot,
-        equipmentId
-      },
-      `预演装备 ${itemName}`
-    );
-    this.renderView();
-
-    try {
-      const update = await this.session.equipHeroItem(hero.id, slot, equipmentId);
-      await this.applySessionUpdate(update);
-      this.pushSessionActionOutcome(update, {
-        successMessage: "装备已结算。",
-        rejectedLabel: "装备调整"
-      });
-    } catch (error) {
-      const reason = error instanceof Error ? error.message : "equip_failed";
-      const detail = error instanceof Error ? formatEquipmentActionReason(error.message) : "装备失败。";
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(hero.id), {
-          category: "inventory",
-          checkpoint: "equipment.equip.rejected",
-          status: "failure",
-          detail,
-          reason,
-          slot,
-          ...(equipmentId ? { equipmentId } : {})
-        })
-      );
-      this.rollbackPrediction(error instanceof Error ? formatEquipmentActionReason(error.message) : "装备失败。");
-    } finally {
-      this.moveInFlight = false;
-    }
-
-    this.renderView();
-  }
-
-  async unequipHeroItem(slot: EquipmentType): Promise<void> {
-    if (this.moveInFlight || this.battleActionInFlight) {
-      return;
-    }
-
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    const hero = this.activeHero();
-    if (!hero) {
-      this.pushLog("当前快照里没有可控制的英雄。");
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(), {
-          category: "inventory",
-          checkpoint: "equipment.unequip.rejected",
-          status: "blocked",
-          detail: "Unequip request ignored because no controlled hero is present.",
-          reason: "no_controlled_hero"
-        })
-      );
-      this.renderView();
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.pushLog("战斗中无法调整装备。");
-      this.predictionStatus = "战斗中无法调整装备。";
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(hero.id), {
-          category: "inventory",
-          checkpoint: "equipment.unequip.rejected",
-          status: "blocked",
-          detail: "Unequip request rejected because the client is currently in battle.",
-          reason: "in_battle"
-        })
-      );
-      this.renderView();
-      return;
-    }
-
-    const currentItemId =
-      slot === "weapon"
-        ? hero.loadout.equipment.weaponId
-        : slot === "armor"
-          ? hero.loadout.equipment.armorId
-          : hero.loadout.equipment.accessoryId;
-    const itemName = currentItemId ? getEquipmentDefinition(currentItemId)?.name ?? currentItemId : formatEquipmentSlotLabel(slot);
-    this.moveInFlight = true;
-    this.predictionStatus = `正在卸下 ${itemName}...`;
-    this.pushLog(`正在为 ${hero.name} 卸下 ${itemName}...`);
-    this.applyPrediction(
-      {
-        type: "hero.unequip",
-        heroId: hero.id,
-        slot
-      },
-      `预演卸下 ${itemName}`
-    );
-    this.renderView();
-
-    try {
-      const update = await this.session.unequipHeroItem(hero.id, slot);
-      await this.applySessionUpdate(update);
-      this.pushSessionActionOutcome(update, {
-        successMessage: "卸装已结算。",
-        rejectedLabel: "卸装"
-      });
-    } catch (error) {
-      const reason = error instanceof Error ? error.message : "unequip_failed";
-      const detail = error instanceof Error ? formatEquipmentActionReason(error.message) : "卸装失败。";
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(hero.id), {
-          category: "inventory",
-          checkpoint: "equipment.unequip.rejected",
-          status: "failure",
-          detail,
-          reason,
-          slot,
-          ...(currentItemId ? { equipmentId: currentItemId } : {})
-        })
-      );
-      this.rollbackPrediction(error instanceof Error ? formatEquipmentActionReason(error.message) : "卸装失败。");
-    } finally {
-      this.moveInFlight = false;
-    }
-
-    this.renderView();
-  }
-
-  private ensureViewNodes(): void {
-    assignUiLayer(this.node);
-
-    if (!this.node.getComponent(Canvas)) {
-      this.node.addComponent(Canvas);
-    }
-
-    const rootTransform = this.node.getComponent(UITransform) ?? this.node.addComponent(UITransform);
-    const visibleSize = view.getVisibleSize();
-    rootTransform.setContentSize(visibleSize.width, visibleSize.height);
-    const { effectiveTileSize, hudWidth, rightWidth, mapWidth, hudHeight, battleHeight, timelineHeight } =
-      this.computeLayoutMetrics();
-
-    let hudNode = this.node.getChildByName(HUD_NODE_NAME);
-    if (!hudNode) {
-      hudNode = new Node(HUD_NODE_NAME);
-      hudNode.parent = this.node;
-    }
-    assignUiLayer(hudNode);
-
-    const hudTransform = hudNode.getComponent(UITransform) ?? hudNode.addComponent(UITransform);
-    hudTransform.setContentSize(hudWidth, hudHeight);
-    this.hudPanel = hudNode.getComponent(VeilHudPanel) ?? hudNode.addComponent(VeilHudPanel);
-    this.hudPanel.configure({
-      onNewRun: () => {
-        void this.startNewRun();
-      },
-      onRefresh: () => {
-        void this.refreshSnapshot();
-      },
-      onToggleSettings: () => {
-        this.toggleSettingsPanel();
-      },
-      onToggleCampaign: () => {
-        void this.toggleGameplayCampaignPanel();
-      },
-      onToggleInventory: () => {
-        this.toggleGameplayEquipmentPanel();
-      },
-      onToggleAchievements: () => {
-        void this.openGameplayBattleReportCenter();
-      },
-      onToggleDailyDungeon: () => {
-        void this.toggleGameplayDailyDungeonPanel();
-      },
-      onToggleProgression: () => {
-        void this.toggleGameplayBattlePassPanel();
-      },
-      onToggleSeasonalEvent: () => {
-        void this.toggleGameplaySeasonalEventPanel();
-      },
-      onToggleReport: () => {
-        this.toggleReportDialog();
-      },
-      onToggleSurrender: () => {
-        this.toggleSurrenderDialog();
-      },
-      onShareBattleResult: () => {
-        void this.handleBattleResultShare();
-      },
-      onSubmitReport: (reason) => {
-        void this.submitPlayerReport(reason);
-      },
-      onCancelReport: () => {
-        this.closeReportDialog();
-      },
-      onConfirmSurrender: () => {
-        void this.confirmSurrender();
-      },
-      onCancelSurrender: () => {
-        this.closeSurrenderDialog();
-      },
-      onLearnSkill: (skillId) => {
-        void this.learnHeroSkill(skillId);
-      },
-      onEquipItem: (slot, equipmentId) => {
-        void this.equipHeroItem(slot, equipmentId);
-      },
-      onUnequipItem: (slot) => {
-        void this.unequipHeroItem(slot);
-      },
-      onEndDay: () => {
-        void this.advanceDay();
-      },
-      onReturnLobby: () => {
-        void this.returnToLobby();
-      }
-    });
-
-    let lobbyNode = this.node.getChildByName(LOBBY_NODE_NAME);
-    if (!lobbyNode) {
-      lobbyNode = new Node(LOBBY_NODE_NAME);
-      lobbyNode.parent = this.node;
-    }
-    assignUiLayer(lobbyNode);
-    const lobbyTransform = lobbyNode.getComponent(UITransform) ?? lobbyNode.addComponent(UITransform);
-    lobbyTransform.setContentSize(Math.max(360, visibleSize.width - 48), Math.max(620, visibleSize.height - 52));
-    this.lobbyPanel = lobbyNode.getComponent(VeilLobbyPanel) ?? lobbyNode.addComponent(VeilLobbyPanel);
-    this.lobbyPanel.configure({
-      onEditPlayerId: () => {
-        this.promptForLobbyField("playerId");
-      },
-      onEditDisplayName: () => {
-        this.promptForLobbyField("displayName");
-      },
-      onEditRoomId: () => {
-        this.promptForLobbyField("roomId");
-      },
-      onEditLoginId: () => {
-        this.promptForLobbyField("loginId");
-      },
-      onTogglePrivacyConsent: () => {
-        this.togglePrivacyConsent();
-      },
-      onRefresh: () => {
-        void this.syncLobbyBootstrap();
-      },
-      onEnterRoom: () => {
-        void this.enterLobbyRoom();
-      },
-      onEnterMatchmaking: () => {
-        void this.enterLobbyMatchmaking();
-      },
-      onCancelMatchmaking: () => {
-        void this.cancelLobbyMatchmaking();
-      },
-      onLoginAccount: () => {
-        void this.loginLobbyAccount();
-      },
-      onRegisterAccount: () => {
-        this.openLobbyAccountFlow("registration");
-      },
-      onRecoverAccount: () => {
-        this.openLobbyAccountFlow("recovery");
-      },
-      onEditAccountFlowField: (field) => {
-        this.promptForAccountFlowField(field);
-      },
-      onRequestAccountFlow: () => {
-        void this.requestActiveAccountFlow();
-      },
-      onConfirmAccountFlow: () => {
-        void this.confirmActiveAccountFlow();
-      },
-      onToggleAccountMinorProtection: () => {
-        this.toggleWechatMinorProtectionSelection();
-      },
-      onBindWechatAccount: () => {
-        void this.bindWechatIdentityFromLobbyAccountFlow();
-      },
-      onCancelAccountFlow: () => {
-        this.closeLobbyAccountFlow();
-      },
-      onOpenCampaign: () => {
-        void this.openLobbyPvePanel("campaign");
-      },
-      onOpenDailyDungeon: () => {
-        void this.openLobbyPvePanel("daily-dungeon");
-      },
-      onOpenBattlePass: () => {
-        void this.openLobbyPvePanel("battle-pass");
-      },
-      onOpenConfigCenter: () => {
-        this.openConfigCenter();
-      },
-      onLogout: () => {
-        this.logoutAuthSession();
-      },
-      onJoinRoom: (roomId) => {
-        void this.enterLobbyRoom(roomId);
-      },
-      onToggleAccountReview: (open) => {
-        if (open) {
-          void this.refreshActiveAccountReviewSection();
-        }
-      },
-      onSelectAccountReviewSection: (section) => {
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "section.selected",
-          section
-        });
-        this.renderView();
-        void this.refreshActiveAccountReviewSection();
-      },
-      onSelectAccountReviewPage: (section, page) => {
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "section.selected",
-          section
-        });
-        this.renderView();
-        void this.refreshAccountReviewPage(section, page);
-      },
-      onRetryAccountReviewSection: (section) => {
-        void this.refreshActiveAccountReviewSection(section);
-      },
-      onSelectBattleReplayReview: (replayId) => {
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "battle-replay.selected",
-          replayId
-        });
-        this.renderView();
-      },
-      onLearnLobbySkill: (skillId) => {
-        void this.learnHeroSkill(skillId);
-      },
-      onPurchaseShopProduct: (productId) => {
-        void this.purchaseLobbyShopProduct(productId);
-      },
-      onClaimDailyQuest: (questId) => {
-        void this.claimLobbyDailyQuest(questId);
-      },
-      onClaimMailboxMessage: (messageId) => {
-        void this.claimLobbyMailboxMessage(messageId);
-      },
-      onClaimAllMailbox: () => {
-        void this.claimAllLobbyMailboxMessages();
-      }
-    });
-
-    let tutorialOverlayNode = this.node.getChildByName(TUTORIAL_OVERLAY_NODE_NAME);
-    if (!tutorialOverlayNode) {
-      tutorialOverlayNode = new Node(TUTORIAL_OVERLAY_NODE_NAME);
-      tutorialOverlayNode.parent = this.node;
-    }
-    assignUiLayer(tutorialOverlayNode);
-    const tutorialOverlayTransform = tutorialOverlayNode.getComponent(UITransform) ?? tutorialOverlayNode.addComponent(UITransform);
-    tutorialOverlayTransform.setContentSize(visibleSize.width, visibleSize.height);
-    this.tutorialOverlay =
-      tutorialOverlayNode.getComponent(VeilTutorialOverlay) ?? tutorialOverlayNode.addComponent(VeilTutorialOverlay);
-    this.tutorialOverlay.configure({
-      onPrimaryAction: () => {
-        void this.handleTutorialPrimaryAction();
-      },
-      onSecondaryAction: () => {
-        void this.skipTutorialFlow();
-      }
-    });
-
-    let mapRoot = this.node.getChildByName(MAP_NODE_NAME);
-    if (!mapRoot) {
-      mapRoot = new Node(MAP_NODE_NAME);
-      mapRoot.parent = this.node;
-    }
-    assignUiLayer(mapRoot);
-
-    const mapTransform = mapRoot.getComponent(UITransform) ?? mapRoot.addComponent(UITransform);
-    mapTransform.setContentSize(mapWidth, Math.max(480, visibleSize.height - 48));
-
-    this.mapBoard = mapRoot.getComponent(VeilMapBoard) ?? mapRoot.addComponent(VeilMapBoard);
-    this.mapBoard.setFogPulsePhase(this.fogPulsePhase);
-    this.mapBoard.configure({
-      tileSize: effectiveTileSize,
-      onTileSelected: (tile) => {
-        void this.moveHeroToTile(tile);
-      },
-      onInputDebug: (message) => {
-        this.inputDebug = message;
-        this.renderView();
-      }
-    });
-
-    let battleNode = this.node.getChildByName(BATTLE_NODE_NAME);
-    if (!battleNode) {
-      battleNode = new Node(BATTLE_NODE_NAME);
-      battleNode.parent = this.node;
-    }
-    assignUiLayer(battleNode);
-
-    const battleTransform = battleNode.getComponent(UITransform) ?? battleNode.addComponent(UITransform);
-    battleTransform.setContentSize(rightWidth, battleHeight);
-    this.battlePanel = battleNode.getComponent(VeilBattlePanel) ?? battleNode.addComponent(VeilBattlePanel);
-    this.battlePanel.configure({
-      onSelectTarget: (unitId) => {
-        this.selectedBattleTargetId = unitId;
-        this.renderView();
-      },
-      onAction: (action) => {
-        void this.actInBattle(action);
-      }
-    });
-
-    let timelineNode = this.node.getChildByName(TIMELINE_NODE_NAME);
-    if (!timelineNode) {
-      timelineNode = new Node(TIMELINE_NODE_NAME);
-      timelineNode.parent = this.node;
-    }
-    assignUiLayer(timelineNode);
-
-    const timelineTransform = timelineNode.getComponent(UITransform) ?? timelineNode.addComponent(UITransform);
-    timelineTransform.setContentSize(rightWidth, timelineHeight);
-    this.timelinePanel = timelineNode.getComponent(VeilTimelinePanel) ?? timelineNode.addComponent(VeilTimelinePanel);
-
-    let accountReviewPanelNode = this.node.getChildByName(ACCOUNT_REVIEW_PANEL_NODE_NAME);
-    if (!accountReviewPanelNode) {
-      accountReviewPanelNode = new Node(ACCOUNT_REVIEW_PANEL_NODE_NAME);
-      accountReviewPanelNode.parent = this.node;
-    }
-    assignUiLayer(accountReviewPanelNode);
-    const accountReviewTransform = accountReviewPanelNode.getComponent(UITransform) ?? accountReviewPanelNode.addComponent(UITransform);
-    accountReviewTransform.setContentSize(Math.max(320, Math.min(420, visibleSize.width - 56)), Math.max(360, visibleSize.height - 96));
-    this.gameplayAccountReviewPanel =
-      accountReviewPanelNode.getComponent(VeilProgressionPanel) ?? accountReviewPanelNode.addComponent(VeilProgressionPanel);
-    this.gameplayAccountReviewPanel.configure({
-      onClose: () => {
-        if (this.gameplayDailyDungeonPanelOpen) {
-          void this.toggleGameplayDailyDungeonPanel(false);
-          return;
-        }
-        if (this.gameplayBattlePassPanelOpen) {
-          void this.toggleGameplayBattlePassPanel(false);
-          return;
-        }
-        if (this.gameplaySeasonalEventPanelOpen) {
-          void this.toggleGameplaySeasonalEventPanel(false);
-          return;
-        }
-        void this.toggleGameplayAccountReviewPanel(false);
-      },
-      onSelectSection: (section) => {
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "section.selected",
-          section
-        });
-        this.renderView();
-        void this.refreshActiveAccountReviewSection();
-      },
-      onSelectPage: (section, page) => {
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "section.selected",
-          section
-        });
-        this.renderView();
-        void this.refreshAccountReviewPage(section, page);
-      },
-      onRetrySection: (section) => {
-        void this.refreshActiveAccountReviewSection(section);
-      },
-      onClaimTier: (tier) => {
-        void this.claimGameplaySeasonTier(tier);
-      },
-      onPurchasePremium: () => {
-        void this.purchaseGameplaySeasonPremium();
-      },
-      onAttemptDailyDungeonFloor: (floor) => {
-        void this.attemptGameplayDailyDungeonFloor(floor);
-      },
-      onClaimDailyDungeonRun: (runId) => {
-        void this.claimGameplayDailyDungeonRun(runId);
-      },
-      onRefreshDailyDungeon: () => {
-        void this.refreshDailyDungeonPanel();
-      }
-    });
-
-    let equipmentPanelNode = this.node.getChildByName(EQUIPMENT_PANEL_NODE_NAME);
-    if (!equipmentPanelNode) {
-      equipmentPanelNode = new Node(EQUIPMENT_PANEL_NODE_NAME);
-      equipmentPanelNode.parent = this.node;
-    }
-    assignUiLayer(equipmentPanelNode);
-    const equipmentPanelTransform = equipmentPanelNode.getComponent(UITransform) ?? equipmentPanelNode.addComponent(UITransform);
-    equipmentPanelTransform.setContentSize(Math.max(360, Math.min(460, visibleSize.width - 56)), Math.max(420, visibleSize.height - 96));
-    this.gameplayEquipmentPanel =
-      equipmentPanelNode.getComponent(VeilEquipmentPanel) ?? equipmentPanelNode.addComponent(VeilEquipmentPanel);
-    this.gameplayEquipmentPanel.configure({
-      onClose: () => {
-        this.toggleGameplayEquipmentPanel(false);
-      },
-      onEquipItem: (slot, equipmentId) => {
-        void this.equipHeroItem(slot, equipmentId);
-      },
-      onUnequipItem: (slot) => {
-        void this.unequipHeroItem(slot);
-      }
-    });
-
-    let campaignPanelNode = this.node.getChildByName(CAMPAIGN_PANEL_NODE_NAME);
-    if (!campaignPanelNode) {
-      campaignPanelNode = new Node(CAMPAIGN_PANEL_NODE_NAME);
-      campaignPanelNode.parent = this.node;
-    }
-    assignUiLayer(campaignPanelNode);
-    const campaignPanelTransform = campaignPanelNode.getComponent(UITransform) ?? campaignPanelNode.addComponent(UITransform);
-    campaignPanelTransform.setContentSize(Math.max(380, Math.min(500, visibleSize.width - 56)), Math.max(480, visibleSize.height - 96));
-    this.gameplayCampaignPanel =
-      campaignPanelNode.getComponent(VeilCampaignPanel) ?? campaignPanelNode.addComponent(VeilCampaignPanel);
-    this.gameplayCampaignPanel.configure({
-      onClose: () => {
-        void this.toggleGameplayCampaignPanel(false);
-      },
-      onRefresh: () => {
-        void this.refreshGameplayCampaign();
-      },
-      onSelectPrevious: () => {
-        this.selectGameplayCampaignMission("previous");
-      },
-      onSelectNext: () => {
-        this.selectGameplayCampaignMission("next");
-      },
-      onFocusNextAvailable: () => {
-        this.selectGameplayCampaignMission("next-available");
-      },
-      onStartMission: () => {
-        void this.startGameplayCampaignMission();
-      },
-      onAdvanceDialogue: () => {
-        this.advanceGameplayCampaignDialogue();
-      },
-      onCompleteMission: () => {
-        void this.completeGameplayCampaignMission();
-      }
-    });
-
-    let settingsPanelNode = this.node.getChildByName(SETTINGS_PANEL_NODE_NAME);
-    if (!settingsPanelNode) {
-      settingsPanelNode = new Node(SETTINGS_PANEL_NODE_NAME);
-      settingsPanelNode.parent = this.node;
-    }
-    assignUiLayer(settingsPanelNode);
-    const settingsPanelTransform = settingsPanelNode.getComponent(UITransform) ?? settingsPanelNode.addComponent(UITransform);
-    settingsPanelTransform.setContentSize(Math.max(360, Math.min(460, visibleSize.width - 64)), Math.max(440, visibleSize.height - 96));
-    this.settingsPanel = settingsPanelNode.getComponent(CocosSettingsPanel) ?? settingsPanelNode.addComponent(CocosSettingsPanel);
-    this.settingsPanel.configure({
-      onClose: () => {
-        this.toggleSettingsPanel(false);
-      },
-      onUpdate: (update) => {
-        this.updateSettings(update);
-      },
-      onLogout: () => {
-        void this.handleSettingsLogout();
-      },
-      onDeleteAccount: () => {
-        void this.handleSettingsDeleteAccount();
-      },
-      onWithdrawConsent: () => {
-        void this.handleSettingsWithdrawConsent();
-      },
-      onOpenPrivacyPolicy: () => {
-        this.openSettingsPrivacyPolicy();
-      },
-      onSubmitSupportTicket: (category) => {
-        void this.handleSettingsSupportTicket(category);
-      }
-    });
-
-    let settingsButtonNode = this.node.getChildByName(SETTINGS_BUTTON_NODE_NAME);
-    if (!settingsButtonNode) {
-      settingsButtonNode = new Node(SETTINGS_BUTTON_NODE_NAME);
-      settingsButtonNode.parent = this.node;
-    }
-    assignUiLayer(settingsButtonNode);
-    const settingsButtonTransform = settingsButtonNode.getComponent(UITransform) ?? settingsButtonNode.addComponent(UITransform);
-    settingsButtonTransform.setContentSize(58, 58);
-
-    this.battleTransition = this.node.getComponent(VeilBattleTransition) ?? this.node.addComponent(VeilBattleTransition);
-    this.updateLayout();
-  }
-
-  private ensureUiCameraVisibility(): void {
-    const sceneRoot = this.node.parent;
-    if (!sceneRoot) {
-      return;
-    }
-
-    const uiLayer = Layers.Enum.UI_2D;
-    const cameraNode = sceneRoot.getChildByName("Main Camera");
-    const camera = cameraNode?.getComponent(Camera) ?? null;
-    if (camera) {
-      camera.visibility |= uiLayer;
-      camera.orthoHeight = Math.max(320, view.getVisibleSize().height / 2);
-      camera.near = 0.1;
-      camera.far = 4000;
-    }
-    if (cameraNode) {
-      cameraNode.setPosition(0, 0, 1000);
-      cameraNode.setRotationFromEuler(0, 0, 0);
-    }
-  }
-
-  private ensureHudActionBinding(): void {
-    if (this.hudActionBinding) {
-      return;
-    }
-
-    input.on(Input.EventType.TOUCH_END, this.handleHudActionInput, this);
-    input.on(Input.EventType.MOUSE_UP, this.handleHudActionInput, this);
-    this.hudActionBinding = true;
-  }
-
-  private renderView(): void {
-    renderViewForRoot(this as unknown as Record<string, any>);
-  }
-
-  private renderGameplayEquipmentPanel(): void {
-    renderGameplayEquipmentPanelForRoot(this as unknown as Record<string, any>);
-  }
-
-  private renderGameplayCampaignPanel(): void {
-    renderGameplayCampaignPanelForRoot(this as unknown as Record<string, any>);
-  }
-
-  private renderGameplayAccountReviewPanel(): void {
-    renderGameplayAccountReviewPanelForRoot(this as unknown as Record<string, any>);
-  }
-
-  private formatLobbyVaultSummary(): string {
-    const resources = this.lobbyAccountProfile.globalResources;
-    const ownedCosmetics = this.lobbyAccountProfile.cosmeticInventory?.ownedIds.length ?? 0;
-    const equippedBorder = this.lobbyAccountProfile.equippedCosmetics?.profileBorderId ?? "未装备";
-    return `全局仓库 金币 ${resources.gold} / 木材 ${resources.wood} / 矿石 ${resources.ore} / 宝石 ${this.lobbyAccountProfile.gems ?? 0} / 外观 ${ownedCosmetics} 件 / 边框 ${equippedBorder}`;
-  }
-
-  private ensurePixelSpriteGroup(group: "boot" | "battle"): void {
-    const loadStatus = getPixelSpriteLoadStatus();
-    if (loadStatus.loadedGroups.includes(group) || this.pendingPixelSpriteGroups.has(group)) {
-      return;
-    }
-
-    this.pendingPixelSpriteGroups.add(group);
-    void loadPixelSpriteAssets(group)
-      .then(() => {
-        this.pendingPixelSpriteGroups.delete(group);
-        this.renderView();
-      })
-      .catch(() => {
-        this.pendingPixelSpriteGroups.delete(group);
-      });
-  }
-
-  private openConfigCenter(): void {
-    const configCenterUrl = resolveCocosConfigCenterUrl(this.remoteUrl);
-    if (this.runtimeCapabilities.configCenterAccess !== "external-window") {
-      this.lobbyStatus = `当前${this.runtimePlatform === "wechat-game" ? "微信小游戏" : "运行"}环境不支持直接打开配置台，请在 H5 调试壳访问 ${configCenterUrl}`;
-      this.renderView();
-      return;
-    }
-
-    const openRef = globalThis.open;
-    if (typeof openRef === "function") {
-      openRef(configCenterUrl, "_blank", "noopener,noreferrer");
-      this.lobbyStatus = "已在新窗口打开配置台。";
-    } else {
-      this.lobbyStatus = `当前运行环境无法直接打开配置台，请访问 ${configCenterUrl}`;
-    }
-    this.renderView();
-  }
-
-  private async syncLobbyBootstrap(): Promise<void> {
-    await this.refreshLobbyRoomList();
-    await this.refreshLobbyAccountProfile();
-  }
-
-  private async refreshLobbyAccountProfile(): Promise<void> {
-    const storage = this.readWebStorage();
-    const requestEpoch = this.bumpLobbyAccountEpoch();
-    this.lobbyLeaderboardStatus = "loading";
-    this.lobbyLeaderboardError = null;
-    this.lobbyShopLoading = true;
-    this.lobbyShopStatus = "正在同步商店商品...";
-    this.renderView();
-    const storedSession = readStoredCocosAuthSession(storage);
-    const activeSession = storedSession?.playerId === this.playerId ? storedSession : null;
-    const syncedSession = await resolveVeilRootRuntime().syncAuthSession(this.remoteUrl, {
-      storage,
-      session: activeSession
-    });
-    if (!this.isActiveLobbyAccountEpoch(requestEpoch)) {
-      return;
-    }
-
-    if (syncedSession) {
-      this.authToken = syncedSession.token ?? null;
-      this.authMode = syncedSession.authMode;
-      this.authProvider = syncedSession.provider ?? "guest";
-      this.loginId = syncedSession.loginId ?? "";
-      this.sessionSource = syncedSession.source;
-      this.playerId = syncedSession.playerId;
-      this.displayName = syncedSession.displayName;
-      await this.maybeClaimLaunchReferral(syncedSession);
-    } else if (this.sessionSource !== "manual") {
-      this.authToken = null;
-      this.authMode = "guest";
-      this.authProvider = "guest";
-      this.loginId = "";
-      this.sessionSource = "none";
-    }
-
-    const [profile, leaderboardResult, shopProductsResult, activeEventsResult, announcementsResult, maintenanceModeResult] = await Promise.all([
-      resolveVeilRootRuntime().loadAccountProfile(this.remoteUrl, this.playerId, this.roomId, {
-        storage,
-        authSession: syncedSession
-      }),
-      (this.runtimePlatform === "wechat-game"
-        ? (async () => {
-            const wxRuntime = (globalThis as { wx?: unknown }).wx ?? null;
-            const friendEntries = await readCocosWechatFriendCloudEntries(wxRuntime);
-            const friendIds = friendEntries.map((entry) => entry.playerId);
-            if (friendIds.length === 0) {
-              return { ok: true as const, entries: [] };
-            }
-
-            return {
-              ok: true as const,
-              entries: await resolveVeilRootRuntime().loadFriendLeaderboard(this.remoteUrl, friendIds, {
-                getAuthToken: () => this.authToken
-              })
-            };
-          })()
-        : resolveVeilRootRuntime().loadLeaderboard(this.remoteUrl, 50).then((entries) => ({ ok: true as const, entries })))
-        .catch((error: unknown) => ({ ok: false as const, error })),
-      resolveVeilRootRuntime()
-        .loadShopProducts(this.remoteUrl)
-        .then((products) => ({ ok: true as const, products }))
-        .catch((error: unknown) => ({ ok: false as const, error })),
-      syncedSession?.token
-        ? resolveVeilRootRuntime()
-            .loadActiveSeasonalEvents(this.remoteUrl, {
-              storage,
-              authSession: syncedSession
-            })
-            .then((events) => ({ ok: true as const, events }))
-            .catch((error: unknown) => ({ ok: false as const, error }))
-        : Promise.resolve({ ok: true as const, events: [] as CocosSeasonalEvent[] }),
-      resolveVeilRootRuntime().loadAnnouncements(this.remoteUrl)
-        .then((items) => ({ ok: true as const, items }))
-        .catch((error: unknown) => ({ ok: false as const, error })),
-      resolveVeilRootRuntime().loadMaintenanceMode(this.remoteUrl)
-        .then((snapshot) => ({ ok: true as const, snapshot }))
-        .catch((error: unknown) => ({ ok: false as const, error }))
-    ]);
-    if (!this.isActiveLobbyAccountEpoch(requestEpoch)) {
-      return;
-    }
-
-    this.commitAccountProfile(profile, false);
-    this.lobbyAnnouncements = announcementsResult.ok ? announcementsResult.items : [];
-    this.lobbyMaintenanceMode = maintenanceModeResult.ok ? maintenanceModeResult.snapshot : null;
-    if (this.runtimePlatform === "wechat-game") {
-      const wxRuntime = (globalThis as { wx?: unknown }).wx ?? null;
-      void syncCocosWechatFriendCloudStorage(wxRuntime, {
-        playerId: profile.playerId,
-        eloRating: profile.eloRating ?? 1000
-      });
-    }
-    if (leaderboardResult.ok) {
-      this.lobbyLeaderboardEntries = leaderboardResult.entries;
-      this.lobbyLeaderboardStatus = "ready";
-      this.lobbyLeaderboardError = null;
-    } else {
-      this.lobbyLeaderboardEntries = [];
-      this.lobbyLeaderboardStatus = "error";
-      this.lobbyLeaderboardError =
-        leaderboardResult.error instanceof Error ? leaderboardResult.error.message : "leaderboard_unavailable";
-    }
-    if (shopProductsResult.ok) {
-      this.lobbyShopProducts = shopProductsResult.products;
-      this.lobbyShopStatus =
-        shopProductsResult.products.length > 0
-          ? "点击商品卡片即可购买；微信商品会在小游戏环境拉起支付。"
-          : "当前没有上架商品。";
-      this.maybeEmitShopOpenAnalytics();
-    } else {
-      this.lobbyShopProducts = [];
-      this.lobbyShopStatus =
-        shopProductsResult.error instanceof Error ? shopProductsResult.error.message : "shop_unavailable";
-    }
-    if (activeEventsResult.ok) {
-      this.activeSeasonalEvent = activeEventsResult.events[0] ?? null;
-      this.seasonalEventStatus = this.activeSeasonalEvent
-        ? `已同步 ${this.activeSeasonalEvent.name} · 当前积分 ${this.activeSeasonalEvent.player.points}`
-        : "当前没有进行中的赛季活动。";
-    } else {
-      this.activeSeasonalEvent = null;
-      this.seasonalEventStatus =
-        activeEventsResult.error instanceof Error ? activeEventsResult.error.message : "seasonal_event_unavailable";
-    }
-    this.lobbyShopLoading = false;
-    if (profile.source === "remote") {
-      this.displayName = profile.displayName;
-      this.loginId = profile.loginId ?? this.loginId;
-    }
-    this.syncWechatShareBridge();
-    this.renderView();
-  }
-
-  private describeShopError(error: unknown): string {
-    if (!(error instanceof Error)) {
-      return "商品购买失败，请稍后重试。";
-    }
-
-    switch (error.message) {
-      case "cocos_request_failed:401:unauthorized":
-      case "cocos_request_failed:401:token_expired":
-        return "购买需要有效账号会话，请重新登录后再试。";
-      case "cocos_request_failed:409:insufficient_gems":
-        return "宝石不足，无法完成本次购买。";
-      case "cocos_request_failed:409:product_not_available":
-        return "该商品当前未上架。";
-      case "cocos_request_failed:409:equipment_inventory_full":
-        return "背包已满，暂时无法领取该装备。";
-      case "cocos_request_failed:409:cosmetic_not_owned":
-        return "尚未拥有该外观，无法装备。";
-      case "cocos_request_failed:409:cosmetic_not_found":
-        return "该外观已下架或不存在。";
-      case "cocos_request_failed:400:wechat_open_id_required":
-        return "微信支付需要先绑定小游戏身份。";
-      case "cocos_request_failed:503:wechat_pay_not_configured":
-        return "服务器尚未配置微信支付。";
-      default:
-        return error.message.startsWith("cocos_request_failed:")
-          ? "商品购买失败，请稍后重试。"
-          : error.message;
-    }
-  }
-
-  private async claimLobbyMailboxMessage(messageId: string): Promise<void> {
-    const storage = this.readWebStorage();
-    const authSession = readStoredCocosAuthSession(storage);
-    if (!authSession?.token) {
-      this.lobbyStatus = "系统邮箱领取需要先登录云端账号或游客会话。";
-      this.renderView();
-      return;
-    }
-
-    this.mailboxClaimingMessageId = messageId;
-    this.lobbyStatus = "正在领取邮件附件...";
-    this.renderView();
-    try {
-      const payload = await claimCocosMailboxMessage(this.remoteUrl, messageId, {
-        authSession,
-        storage
-      });
-      this.lobbyStatus =
-        payload.claimed
-          ? "邮件附件已领取，正在同步仓库状态。"
-          : payload.reason === "already_claimed"
-            ? "该邮件附件已经领取过。"
-            : payload.reason === "expired"
-              ? "该邮件已经过期。"
-              : "该邮件没有可领取附件。";
-      await this.refreshLobbyAccountProfile();
-    } catch (error) {
-      this.lobbyStatus = error instanceof Error ? error.message : "mailbox_claim_failed";
-    } finally {
-      this.mailboxClaimingMessageId = null;
-      this.renderView();
-    }
-  }
-
-  private async claimLobbyDailyQuest(questId: string): Promise<void> {
-    const storage = this.readWebStorage();
-    const authSession = readStoredCocosAuthSession(storage);
-    if (!authSession?.token) {
-      this.lobbyStatus = "每日任务领取需要先登录云端账号或游客会话。";
-      this.renderView();
-      return;
-    }
-
-    this.dailyQuestClaimingId = questId;
-    this.lobbyStatus = "正在领取每日任务奖励...";
-    this.renderView();
-    try {
-      const payload = await resolveVeilRootRuntime().claimDailyQuest(this.remoteUrl, questId, {
-        authSession,
-        storage
-      });
-      this.lobbyStatus =
-        payload.claimed
-          ? "每日任务奖励已领取，正在同步任务板。"
-          : payload.reason === "already_claimed"
-            ? "该每日任务奖励已经领取过。"
-            : payload.reason === "quest_incomplete"
-              ? "任务尚未完成，暂时无法领取奖励。"
-              : "每日任务领取失败，请稍后重试。";
-      await this.refreshLobbyAccountProfile();
-    } catch (error) {
-      this.lobbyStatus = error instanceof Error ? error.message : "daily_quest_claim_failed";
-    } finally {
-      this.dailyQuestClaimingId = null;
-      this.renderView();
-    }
-  }
-
-  private async claimAllLobbyMailboxMessages(): Promise<void> {
-    const storage = this.readWebStorage();
-    const authSession = readStoredCocosAuthSession(storage);
-    if (!authSession?.token) {
-      this.lobbyStatus = "系统邮箱领取需要先登录云端账号或游客会话。";
-      this.renderView();
-      return;
-    }
-
-    this.mailboxClaimAllInFlight = true;
-    this.lobbyStatus = "正在领取全部邮件附件...";
-    this.renderView();
-    try {
-      const payload = await claimAllCocosMailboxMessages(this.remoteUrl, {
-        authSession,
-        storage
-      });
-      this.lobbyStatus = payload.claimed ? "邮件附件已全部领取，正在同步仓库状态。" : "当前没有可领取的邮件附件。";
-      await this.refreshLobbyAccountProfile();
-    } catch (error) {
-      this.lobbyStatus = error instanceof Error ? error.message : "mailbox_claim_all_failed";
-    } finally {
-      this.mailboxClaimAllInFlight = false;
-      this.renderView();
-    }
-  }
-
-  private async purchaseLobbyShopProduct(productId: string): Promise<void> {
-    if (this.pendingShopProductId || this.lobbyEntering) {
-      return;
-    }
-
-    const product = this.lobbyShopProducts.find((entry) => entry.productId === productId);
-    if (!product) {
-      this.lobbyShopStatus = "未找到要购买的商品。";
-      this.renderView();
-      return;
-    }
-    if (!this.authToken) {
-      this.lobbyShopStatus = "购买需要有效会话，请先重新进入大厅。";
-      this.renderView();
-      return;
-    }
-
-    this.pendingShopProductId = productId;
-    const cosmeticId = product.grant.cosmeticIds?.[0];
-    const alreadyOwned = cosmeticId ? (this.lobbyAccountProfile.cosmeticInventory?.ownedIds ?? []).includes(cosmeticId) : false;
-    this.lobbyShopStatus =
-      product.type === "cosmetic" && alreadyOwned
-        ? `正在装备 ${product.name}...`
-        : product.wechatPriceFen
-          ? `正在创建微信订单 ${product.name}...`
-          : `正在购买 ${product.name}...`;
-    this.renderView();
-
-    try {
-      this.trackPurchaseInitiated(product, "lobby");
-      if (product.type === "cosmetic" && alreadyOwned && cosmeticId) {
-        await resolveVeilRootRuntime().equipShopCosmetic(this.remoteUrl, cosmeticId, {
-          getAuthToken: () => this.authToken
-        });
-        this.lobbyShopStatus = `${product.name} 已装备。`;
-        await this.refreshLobbyAccountProfile();
-      } else if (product.wechatPriceFen) {
-        const order = await createCocosWechatPaymentOrder(this.remoteUrl, productId, {
-          authToken: this.authToken
-        });
-        const paymentResult = await requestCocosWechatPayment(
-          (globalThis as { wx?: CocosWechatPaymentRuntimeLike | null }).wx,
-          order
-        );
-        const verification = await verifyCocosWechatPayment(this.remoteUrl, order.orderId, {
-          authToken: this.authToken
-        });
-        this.lobbyShopStatus =
-          verification.seasonPassPremium
-            ? `${product.name} 购买成功，赛季高级通行证已解锁。`
-            : `${product.name} 购买成功，当前宝石 ${verification.gemsBalance}。`;
-        if (paymentResult.available) {
-          await this.refreshLobbyAccountProfile();
-        }
-      } else {
-        const result = await resolveVeilRootRuntime().purchaseShopProduct(this.remoteUrl, productId, {
-          getAuthToken: () => this.authToken
-        });
-        this.lobbyShopStatus = `${product.name} 购买成功，当前宝石 ${result.gemsBalance}。`;
-        await this.refreshLobbyAccountProfile();
-      }
-    } catch (error) {
-      this.lobbyShopStatus = this.describeShopError(error);
-    } finally {
-      this.pendingShopProductId = null;
-      this.renderView();
-    }
-  }
-
-  private async refreshActiveAccountReviewSection(section = this.lobbyAccountReviewState.activeSection): Promise<void> {
-    if (section === "progression") {
-      await this.refreshProgressionReview();
-      return;
-    }
-
-    if (section === "achievements") {
-      await this.refreshAchievementReview();
-      return;
-    }
-
-    if (section === "event-history") {
-      await this.refreshAccountReviewPage("event-history", this.lobbyAccountReviewState.eventHistory.page);
-      return;
-    }
-
-    await this.refreshAccountReviewPage("battle-replays", this.lobbyAccountReviewState.battleReplays.page);
-  }
-
-  private async refreshProgressionReview(): Promise<void> {
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "section.loading",
-      section: "progression"
-    });
-    this.renderView();
-
-    try {
-      const snapshot = await resolveVeilRootRuntime().loadProgressionSnapshot(this.remoteUrl, this.playerId, 6, {
-        storage: this.readWebStorage(),
-        authSession: this.currentLobbyAuthSession(),
-        throwOnError: true
-      });
-      this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-        type: "progression.loaded",
-        snapshot
-      });
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(), {
-          category: "progression",
-          checkpoint: "review.loaded",
-          status: "success",
-          detail: `Progression review loaded with ${snapshot.recentEventLog.length} recent events.`,
-          itemCount: snapshot.recentEventLog.length
-        })
-      );
-    } catch (error) {
-      const message = this.describeAccountReviewLoadError(error);
-      this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-        type: "section.failed",
-        section: "progression",
-        message
-      });
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(), {
-          category: "progression",
-          checkpoint: "review.failed",
-          status: "failure",
-          detail: message,
-          reason: message
-        })
-      );
-    }
-
-    this.renderView();
-  }
-
-  private async refreshAchievementReview(): Promise<void> {
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "section.loading",
-      section: "achievements"
-    });
-    this.renderView();
-
-    try {
-      const items = await resolveVeilRootRuntime().loadAchievementProgress(this.remoteUrl, this.playerId, undefined, {
-        storage: this.readWebStorage(),
-        authSession: this.currentLobbyAuthSession(),
-        throwOnError: true
-      });
-      this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-        type: "achievements.loaded",
-        items
-      });
-    } catch (error) {
-      this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-        type: "section.failed",
-        section: "achievements",
-        message: this.describeAccountReviewLoadError(error)
-      });
-    }
-
-    this.renderView();
-  }
-
-  private async toggleGameplayAccountReviewPanel(forceOpen?: boolean): Promise<void> {
-    await toggleGameplayAccountReviewPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private async toggleGameplayBattlePassPanel(forceOpen?: boolean): Promise<void> {
-    await toggleGameplayBattlePassPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private async toggleGameplayDailyDungeonPanel(forceOpen?: boolean): Promise<void> {
-    await toggleGameplayDailyDungeonPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private async openLobbyPvePanel(target: "campaign" | "daily-dungeon" | "battle-pass"): Promise<void> {
-    await openLobbyPvePanelForRoot(this as unknown as Record<string, any>, target);
-  }
-
-  private async refreshDailyDungeonPanel(successStatus?: string): Promise<void> {
-    await refreshDailyDungeonPanelForRoot(this as unknown as Record<string, any>, successStatus);
-  }
-
-  private async toggleGameplaySeasonalEventPanel(forceOpen?: boolean): Promise<void> {
-    await toggleGameplaySeasonalEventPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private snapshotSeasonProgressFromProfile(): CocosSeasonProgress {
-    return snapshotSeasonProgressFromProfileForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async refreshSeasonProgress(): Promise<void> {
-    await refreshSeasonProgressForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async refreshActiveSeasonalEvent(): Promise<void> {
-    await refreshActiveSeasonalEventForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async submitBattleProgressForActiveEvents(update: SessionUpdate): Promise<void> {
-    if (this.sessionSource !== "remote" || !this.authToken) {
-      return;
-    }
-
-    const authSession = this.currentLobbyAuthSession();
-    if (!authSession?.token) {
-      return;
-    }
-
-    const resolvedBattles = update.events.filter(
-      (event): event is Extract<SessionUpdate["events"][number], { type: "battle.resolved" }> => event.type === "battle.resolved"
-    );
-    const ownResolvedBattle = resolvedBattles.find((event) => update.world.ownHeroes.some((hero) => hero.id === event.heroId)) ?? null;
-    if (!ownResolvedBattle) {
-      return;
-    }
-
-    const actionId = `${update.world.playerId}:${ownResolvedBattle.battleId}`;
-    if (this.pendingSeasonalEventBattleIds.has(actionId)) {
-      return;
-    }
-
-    this.pendingSeasonalEventBattleIds.add(actionId);
-    try {
-      const events =
-        this.activeSeasonalEvent ? [this.activeSeasonalEvent] : await resolveVeilRootRuntime().loadActiveSeasonalEvents(this.remoteUrl, {
-          storage: this.readWebStorage(),
-          authSession,
-          throwOnError: false
-        });
-      if (events.length === 0) {
-        if (!this.activeSeasonalEvent) {
-          this.seasonalEventStatus = "当前没有进行中的赛季活动。";
-          this.renderView();
-        }
-        return;
-      }
-
-      for (const seasonalEvent of events) {
-        const result = await resolveVeilRootRuntime().submitSeasonalEventProgress(
-          this.remoteUrl,
-          seasonalEvent.id,
-          {
-            actionId,
-            actionType: "battle_resolved",
-            battleId: ownResolvedBattle.battleId,
-            occurredAt: new Date().toISOString()
-          },
-          {
-            storage: this.readWebStorage(),
-            authSession
-          }
-        );
-        if (result.event) {
-          this.activeSeasonalEvent = result.event;
-        }
-        if (result.eventProgress) {
-          this.seasonalEventStatus = `${seasonalEvent.name} +${result.eventProgress.delta} 分 · 当前 ${result.eventProgress.points} 分`;
-        }
-      }
-      this.renderView();
-    } catch (error) {
-      this.seasonalEventStatus = error instanceof Error ? error.message : "seasonal_event_progress_failed";
-      this.renderView();
-    } finally {
-      this.pendingSeasonalEventBattleIds.delete(actionId);
-    }
-  }
-
-  private handleSeasonalEventProgressPush(message: {
-    payload: {
-      eventId: string;
-      points: number;
-      delta: number;
-      objectiveId: string;
-    };
-  }): void {
-    const eventId = message.payload.eventId.trim();
-    if (!eventId) {
-      return;
-    }
-
-    if (this.activeSeasonalEvent?.id === eventId) {
-      const previousRank =
-        this.activeSeasonalEvent.leaderboard.entries.find((entry) => entry.playerId === this.playerId)?.rank ?? null;
-      const currentEntries = this.activeSeasonalEvent.leaderboard.entries.filter((entry) => entry.playerId !== this.playerId);
-      currentEntries.push({
-        rank: previousRank ?? currentEntries.length + 1,
-        playerId: this.playerId,
-        displayName: this.displayName || this.playerId,
-        points: Math.max(0, Math.floor(message.payload.points)),
-        lastUpdatedAt: new Date().toISOString()
-      });
-      currentEntries.sort(
-        (left, right) =>
-          right.points - left.points || left.lastUpdatedAt.localeCompare(right.lastUpdatedAt) || left.playerId.localeCompare(right.playerId)
-      );
-
-      this.activeSeasonalEvent = {
-        ...this.activeSeasonalEvent,
-        player: {
-          ...this.activeSeasonalEvent.player,
-          points: Math.max(0, Math.floor(message.payload.points))
-        },
-        leaderboard: {
-          ...this.activeSeasonalEvent.leaderboard,
-          entries: currentEntries.map((entry, index) => ({
-            ...entry,
-            rank: index + 1
-          }))
-        }
-      };
-    }
-
-    this.seasonalEventStatus = `赛季活动推进：${message.payload.objectiveId} +${message.payload.delta} 分`;
-    if (this.gameplaySeasonalEventPanelOpen) {
-      void this.refreshActiveSeasonalEvent();
-    }
-    this.renderView();
-  }
-
-  private async claimGameplaySeasonTier(tier: number): Promise<void> {
-    await claimGameplaySeasonTierForRoot(this as unknown as Record<string, any>, tier);
-  }
-
-  private async purchaseGameplaySeasonPremium(): Promise<void> {
-    await purchaseGameplaySeasonPremiumForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async attemptGameplayDailyDungeonFloor(floor: number): Promise<void> {
-    await attemptGameplayDailyDungeonFloorForRoot(this as unknown as Record<string, any>, floor);
-  }
-
-  private async claimGameplayDailyDungeonRun(runId: string): Promise<void> {
-    await claimGameplayDailyDungeonRunForRoot(this as unknown as Record<string, any>, runId);
-  }
-
-  private toggleGameplayEquipmentPanel(forceOpen?: boolean): void {
-    toggleGameplayEquipmentPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private async toggleGameplayCampaignPanel(forceOpen?: boolean): Promise<void> {
-    await toggleGameplayCampaignPanelForRoot(this as unknown as Record<string, any>, forceOpen);
-  }
-
-  private resolveSelectedGameplayCampaignMission() {
-    return resolveSelectedGameplayCampaignMissionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private syncGameplayCampaignSelection(preferredMissionId?: string | null): void {
-    syncGameplayCampaignSelectionForRoot(this as unknown as Record<string, any>, preferredMissionId);
-  }
-
-  private selectGameplayCampaignMission(direction: "previous" | "next" | "next-available"): void {
-    selectGameplayCampaignMissionForRoot(this as unknown as Record<string, any>, direction);
-  }
-
-  private async refreshGameplayCampaign(preferredMissionId?: string | null): Promise<void> {
-    await refreshGameplayCampaignForRoot(this as unknown as Record<string, any>, preferredMissionId);
-  }
-
-  private startGameplayCampaignDialogue(missionId: string, sequence: "intro" | "outro"): void {
-    startGameplayCampaignDialogueForRoot(this as unknown as Record<string, any>, missionId, sequence);
-  }
-
-  private advanceGameplayCampaignDialogue(): void {
-    advanceGameplayCampaignDialogueForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async startGameplayCampaignMission(): Promise<void> {
-    await startGameplayCampaignMissionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async completeGameplayCampaignMission(): Promise<void> {
-    await completeGameplayCampaignMissionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private describeCampaignError(error: unknown): string {
-    return describeCampaignErrorForRoot(error);
-  }
-
-  private syncGameplayCampaignBattleOutcome(update: SessionUpdate): void {
-    if (this.gameplayCampaignPendingAction !== null || this.gameplayCampaignDialogue || !this.gameplayCampaignActiveMissionId) {
-      return;
-    }
-
-    const resolution = update.events.find(
-      (event): event is Extract<SessionUpdate["events"][number], { type: "battle.resolved" }> =>
-        event.type === "battle.resolved" && update.world.ownHeroes.some((hero) => hero.id === event.heroId)
-    );
-    if (!resolution) {
-      return;
-    }
-
-    const activeMission =
-      this.gameplayCampaign?.missions.find((mission) => mission.id === this.gameplayCampaignActiveMissionId) ?? null;
-    if (!activeMission) {
-      return;
-    }
-
-    this.gameplayCampaignPanelOpen = true;
-    this.gameplayCampaignSelectedMissionId = activeMission.id;
-    const victory = resolution.result === "attacker_victory" || resolution.result === "defender_victory";
-    if (!victory) {
-      this.gameplayCampaignActiveMissionId = null;
-      this.gameplayCampaignStatus = `${activeMission.name} 未完成，可重新发起挑战。`;
-      this.renderView();
-      return;
-    }
-
-    this.gameplayCampaignStatus = `${activeMission.name} 战斗胜利，正在同步战役结算...`;
-    this.renderView();
-    void this.completeGameplayCampaignMission();
-  }
-
-  private resolveReportTarget(): { playerId: string; name: string } | null {
-    const target = this.lastUpdate?.world.visibleHeroes.find((hero) => hero.playerId !== this.playerId) ?? null;
-    return target ? { playerId: target.playerId, name: `${target.name} · ${target.playerId}` } : null;
-  }
-
-  private resolveSurrenderTarget(): { playerId: string; name: string } | null {
-    const target = this.lastUpdate?.world.visibleHeroes.find((hero) => hero.playerId !== this.playerId) ?? null;
-    return target ? { playerId: target.playerId, name: `${target.name} · ${target.playerId}` } : null;
-  }
-
-  private isSurrenderAvailable(): boolean {
-    return Boolean(this.activeHero() && this.resolveSurrenderTarget() && !this.lastUpdate?.battle);
-  }
-
-  private toggleReportDialog(): void {
-    if (this.reportSubmitting) {
-      return;
-    }
-
-    const target = this.resolveReportTarget();
-    if (!target) {
-      this.reportDialogOpen = false;
-      this.reportStatusMessage = "当前没有可举报的对手。";
-      this.predictionStatus = this.reportStatusMessage;
-      this.renderView();
-      return;
-    }
-
-    this.reportDialogOpen = !this.reportDialogOpen;
-    this.reportStatusMessage = this.reportDialogOpen ? `目标 ${target.name} · ${target.playerId}` : null;
-    this.renderView();
-  }
-
-  private closeReportDialog(): void {
-    if (this.reportSubmitting) {
-      return;
-    }
-
-    this.reportDialogOpen = false;
-    this.reportStatusMessage = null;
-    this.renderView();
-  }
-
-  private toggleSurrenderDialog(): void {
-    if (this.surrenderSubmitting) {
-      return;
-    }
-
-    if (!this.isSurrenderAvailable()) {
-      this.surrenderDialogOpen = false;
-      this.surrenderStatusMessage = "当前不满足认输条件。";
-      this.predictionStatus = this.surrenderStatusMessage;
-      this.renderView();
-      return;
-    }
-
-    const target = this.resolveSurrenderTarget();
-    this.surrenderDialogOpen = !this.surrenderDialogOpen;
-    this.surrenderStatusMessage = this.surrenderDialogOpen ? `认输后将判负给 ${target?.name ?? "当前对手"}。` : null;
-    this.renderView();
-  }
-
-  private closeSurrenderDialog(): void {
-    if (this.surrenderSubmitting) {
-      return;
-    }
-
-    this.surrenderDialogOpen = false;
-    this.surrenderStatusMessage = null;
-    this.renderView();
-  }
-
-  private async submitPlayerReport(reason: PlayerReportReason): Promise<void> {
-    const target = this.resolveReportTarget();
-    if (!this.session || !target) {
-      this.reportDialogOpen = false;
-      this.reportStatusMessage = "当前没有可举报的对手。";
-      this.renderView();
-      return;
-    }
-
-    this.reportSubmitting = true;
-    this.reportStatusMessage = `正在举报 ${target.name}...`;
-    this.renderView();
-
-    try {
-      await this.session.reportPlayer(target.playerId, reason);
-      this.reportDialogOpen = false;
-      this.reportStatusMessage = `已提交举报：${target.name}`;
-      this.predictionStatus = "举报已提交，等待管理员审核。";
-      this.pushLog(`已举报 ${target.name}：${reason}`);
-    } catch (error) {
-      this.reportStatusMessage = error instanceof Error
-        ? error.message === "duplicate_player_report"
-          ? "同一场对局中已举报过该玩家。"
-          : error.message === "report_target_unavailable"
-            ? "目标玩家已不在当前对局中。"
-            : error.message === "reporting_unavailable"
-              ? "当前服务器未启用举报存储。"
-              : error.message === "report_submit_failed"
-                ? "举报提交失败。"
-            : "举报提交失败。"
-        : "举报提交失败。";
-      this.predictionStatus = this.reportStatusMessage;
-    } finally {
-      this.reportSubmitting = false;
-      this.renderView();
-    }
-  }
-
-  private async confirmSurrender(): Promise<void> {
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    const hero = this.activeHero();
-    const target = this.resolveSurrenderTarget();
-    if (!hero || !target || this.lastUpdate?.battle) {
-      this.surrenderDialogOpen = false;
-      this.surrenderStatusMessage = "当前不满足认输条件。";
-      this.predictionStatus = this.surrenderStatusMessage;
-      this.renderView();
-      return;
-    }
-
-    this.surrenderSubmitting = true;
-    this.surrenderStatusMessage = `正在向 ${target.name} 提交认输...`;
-    this.renderView();
-
-    try {
-      const update = await this.session.surrender(hero.id);
-      await this.applySessionUpdate(update);
-      if (update.reason && isSessionSettlementReason(update.reason)) {
-        const message = formatSessionSettlementReason(update.reason, false);
-        this.predictionStatus = message;
-        this.pushLog(message);
-      }
-      this.surrenderDialogOpen = false;
-      this.surrenderStatusMessage = "认输已提交。";
-    } catch (error) {
-      const failureMessage = this.describeSessionError(error, "认输失败。");
-      if (error instanceof Error && error.message === "upgrade_required") {
-        await this.handleForcedUpgrade(failureMessage);
-        return;
-      }
-      this.surrenderStatusMessage = failureMessage;
-      this.predictionStatus = failureMessage;
-      this.pushLog(failureMessage);
-    } finally {
-      this.surrenderSubmitting = false;
-      this.renderView();
-    }
-  }
-
-  private async openGameplayBattleReportCenter(): Promise<void> {
-    this.announceGameplayPanelSwitch("战报回看", "正在打开最近一战与完整战报页，方便决定下一局路线。");
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "section.selected",
-      section: "battle-replays"
-    });
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "battle-replay.selected",
-      replayId: this.lobbyAccountProfile.battleReportCenter?.latestReportId ?? this.lobbyAccountReviewState.selectedBattleReplayId
-    });
-    await this.toggleGameplayAccountReviewPanel(true);
-  }
-
-  private announceGameplayPanelSwitch(title: string, detail: string): void {
-    this.predictionStatus = `已切到${title}：${detail}`;
-    this.pushLog(`已切到${title}：${detail}`);
-  }
-
-  private async refreshAccountReviewPage(
-    section: "battle-replays" | "event-history",
-    page: number
-  ): Promise<void> {
-    const safePage = Math.max(0, Math.floor(page));
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "section.loading",
-      section
-    });
-    this.renderView();
-
-    try {
-      if (section === "event-history") {
-        const history = await resolveVeilRootRuntime().loadEventHistory(
-          this.remoteUrl,
-          this.playerId,
-          {
-            limit: ACCOUNT_REVIEW_PAGE_SIZE,
-            offset: safePage * ACCOUNT_REVIEW_PAGE_SIZE
-          },
-          {
-            storage: this.readWebStorage(),
-            authSession: this.currentLobbyAuthSession(),
-            throwOnError: true
-          }
-        );
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "event-history.loaded",
-          items: history.items,
-          page: Math.floor(history.offset / Math.max(1, history.limit)),
-          pageSize: history.limit,
-          total: history.total,
-          hasMore: history.hasMore
-        });
-      } else {
-        const history = await resolveVeilRootRuntime().loadBattleReplayHistoryPage(
-          this.remoteUrl,
-          this.playerId,
-          {
-            limit: ACCOUNT_REVIEW_PAGE_SIZE,
-            offset: safePage * ACCOUNT_REVIEW_PAGE_SIZE
-          },
-          {
-            storage: this.readWebStorage(),
-            authSession: this.currentLobbyAuthSession(),
-            throwOnError: true
-          }
-        );
-        this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-          type: "battle-replays.loaded",
-          items: history.items,
-          page: Math.floor(history.offset / Math.max(1, history.limit)),
-          pageSize: history.limit,
-          hasMore: history.hasMore
-        });
-      }
-    } catch (error) {
-      this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-        type: "section.failed",
-        section,
-        message: this.describeAccountReviewLoadError(error)
-      });
-    }
-
-    this.renderView();
-  }
-
-  private currentLobbyAuthSession(): {
-    token: string;
-    playerId: string;
-    displayName: string;
-    authMode: "guest" | "account";
-    loginId?: string;
-    source: "remote";
-  } | null {
-    if (!this.authToken) {
-      return null;
-    }
-
-    return {
-      token: this.authToken,
-      playerId: this.playerId,
-      displayName: this.displayName || this.playerId,
-      authMode: this.authMode,
-      ...(this.loginId ? { loginId: this.loginId } : {}),
-      source: "remote"
-    };
-  }
-
-  private describeAccountReviewLoadError(error: unknown): string {
-    return error instanceof Error && error.message.trim() ? error.message : "网络暂不可用，请稍后重试。";
-  }
-
-  private computeLayoutMetrics(): {
-    effectiveTileSize: number;
-    hudWidth: number;
-    rightWidth: number;
-    mapWidth: number;
-    hudHeight: number;
-    battleHeight: number;
-    timelineHeight: number;
-  } {
-    const visibleSize = view.getVisibleSize();
-    const hudWidth = Math.max(228, Math.min(264, Math.floor(visibleSize.width * 0.215)));
-    const rightWidth = Math.max(244, Math.min(276, Math.floor(visibleSize.width * 0.205)));
-    const effectiveTileSize = this.computeEffectiveTileSize(hudWidth, rightWidth);
-    const mapWidth = this.currentMapPixelWidth(effectiveTileSize);
-    const hudHeight = Math.max(318, visibleSize.height - 52);
-    const battleHeight = Math.max(132, Math.floor((visibleSize.height - 72) * 0.23));
-    const timelineHeight = Math.max(226, visibleSize.height - battleHeight - 74);
-
-    return {
-      effectiveTileSize,
-      hudWidth,
-      rightWidth,
-      mapWidth,
-      hudHeight,
-      battleHeight,
-      timelineHeight
-    };
-  }
-
-  private computeEffectiveTileSize(hudWidth: number, rightWidth: number): number {
-    const visibleSize = view.getVisibleSize();
-    const margin = 24;
-    const widthTiles = this.lastUpdate?.world.map.width ?? DEFAULT_MAP_WIDTH_TILES;
-    const heightTiles = this.lastUpdate?.world.map.height ?? DEFAULT_MAP_HEIGHT_TILES;
-    const availableWidth = Math.max(240, visibleSize.width - hudWidth - rightWidth - margin * 4);
-    const availableHeight = Math.max(320, visibleSize.height - margin * 2);
-    const widthBound = Math.floor(availableWidth / widthTiles);
-    const heightBound = Math.floor(availableHeight / heightTiles);
-    return Math.max(36, Math.min(this.tileSize, widthBound, heightBound));
-  }
-
-  private currentMapPixelWidth(tileSize = this.tileSize): number {
-    const widthTiles = this.lastUpdate?.world.map.width ?? DEFAULT_MAP_WIDTH_TILES;
-    return widthTiles * tileSize;
-  }
-
-  private currentMapPixelHeight(tileSize = this.tileSize): number {
-    const heightTiles = this.lastUpdate?.world.map.height ?? DEFAULT_MAP_HEIGHT_TILES;
-    return heightTiles * tileSize;
-  }
-
-  private updateLayout(): void {
-    const visibleSize = view.getVisibleSize();
-    const margin = 24;
-    const { effectiveTileSize, hudWidth, rightWidth, mapWidth, hudHeight, battleHeight, timelineHeight } =
-      this.computeLayoutMetrics();
-    const mapHeight = Math.max(this.currentMapPixelHeight(effectiveTileSize), visibleSize.height - margin * 2);
-
-    const hudNode = this.node.getChildByName(HUD_NODE_NAME);
-    const mapNode = this.node.getChildByName(MAP_NODE_NAME);
-    const battleNode = this.node.getChildByName(BATTLE_NODE_NAME);
-    const timelineNode = this.node.getChildByName(TIMELINE_NODE_NAME);
-    const lobbyNode = this.node.getChildByName(LOBBY_NODE_NAME);
-    const accountReviewPanelNode = this.node.getChildByName(ACCOUNT_REVIEW_PANEL_NODE_NAME);
-    const equipmentPanelNode = this.node.getChildByName(EQUIPMENT_PANEL_NODE_NAME);
-    const campaignPanelNode = this.node.getChildByName(CAMPAIGN_PANEL_NODE_NAME);
-    const settingsPanelNode = this.node.getChildByName(SETTINGS_PANEL_NODE_NAME);
-    const settingsButtonNode = this.node.getChildByName(SETTINGS_BUTTON_NODE_NAME);
-
-    this.mapBoard?.configure({
-      tileSize: effectiveTileSize,
-      onTileSelected: (tile) => {
-        void this.moveHeroToTile(tile);
-      },
-      onInputDebug: (message) => {
-        this.inputDebug = message;
-        this.renderView();
-      }
-    });
-
-    if (hudNode) {
-      const hudTransform = hudNode.getComponent(UITransform) ?? hudNode.addComponent(UITransform);
-      hudTransform.setContentSize(hudWidth, hudHeight);
-      hudNode.setPosition(-visibleSize.width / 2 + margin + hudWidth / 2, 56, 0);
-      this.hudPanel?.configure({
-        onNewRun: () => {
-          void this.startNewRun();
-        },
-        onRefresh: () => {
-          void this.refreshSnapshot();
-        },
-        onToggleSettings: () => {
-          this.toggleSettingsPanel();
-        },
-        onToggleCampaign: () => {
-          void this.toggleGameplayCampaignPanel();
-        },
-        onToggleInventory: () => {
-          this.toggleGameplayEquipmentPanel();
-        },
-        onToggleAchievements: () => {
-          void this.openGameplayBattleReportCenter();
-        },
-        onToggleDailyDungeon: () => {
-          void this.toggleGameplayDailyDungeonPanel();
-        },
-        onToggleProgression: () => {
-          void this.toggleGameplayBattlePassPanel();
-        },
-        onToggleReport: () => {
-          this.toggleReportDialog();
-        },
-        onToggleSurrender: () => {
-          this.toggleSurrenderDialog();
-        },
-        onSubmitReport: (reason) => {
-          void this.submitPlayerReport(reason);
-        },
-        onCancelReport: () => {
-          this.closeReportDialog();
-        },
-        onConfirmSurrender: () => {
-          void this.confirmSurrender();
-        },
-        onCancelSurrender: () => {
-          this.closeSurrenderDialog();
-        },
-        onLearnSkill: (skillId) => {
-          void this.learnHeroSkill(skillId);
-        },
-        onEquipItem: (slot, equipmentId) => {
-          void this.equipHeroItem(slot, equipmentId);
-        },
-        onUnequipItem: (slot) => {
-          void this.unequipHeroItem(slot);
-        },
-        onEndDay: () => {
-          void this.advanceDay();
-        },
-        onReturnLobby: () => {
-          void this.returnToLobby();
-        },
-        onInteractionAction: (actionId) => {
-          const tile = this.selectedInteractionTile();
-          if (!tile?.building) {
-            return;
-          }
-          if (actionId === "recruit" || actionId === "visit" || actionId === "claim" || actionId === "upgrade") {
-            void this.executeBuildingInteraction(tile, actionId);
-          }
-        }
-      });
-    }
-
-    if (mapNode) {
-      const mapTransform = mapNode.getComponent(UITransform) ?? mapNode.addComponent(UITransform);
-      mapTransform.setContentSize(mapWidth, mapHeight);
-      const mapLeft = -visibleSize.width / 2 + margin + hudWidth + margin;
-      mapNode.setPosition(mapLeft + mapWidth / 2, 0, 0);
-    }
-
-    if (battleNode) {
-      const battleTransform = battleNode.getComponent(UITransform) ?? battleNode.addComponent(UITransform);
-      battleTransform.setContentSize(rightWidth, battleHeight);
-      battleNode.setPosition(
-        visibleSize.width / 2 - margin - rightWidth / 2,
-        visibleSize.height / 2 - margin - battleHeight / 2 + 2,
-        0
-      );
-    }
-
-    if (timelineNode) {
-      const timelineTransform = timelineNode.getComponent(UITransform) ?? timelineNode.addComponent(UITransform);
-      timelineTransform.setContentSize(rightWidth, timelineHeight);
-      timelineNode.setPosition(
-        visibleSize.width / 2 - margin - rightWidth / 2,
-        -visibleSize.height / 2 + margin + timelineHeight / 2 + 8,
-        0
-      );
-    }
-
-    if (lobbyNode) {
-      const lobbyTransform = lobbyNode.getComponent(UITransform) ?? lobbyNode.addComponent(UITransform);
-      lobbyTransform.setContentSize(Math.max(360, Math.min(860, visibleSize.width - 40)), Math.max(520, visibleSize.height - 48));
-      lobbyNode.setPosition(0, 0, 0);
-    }
-
-    if (accountReviewPanelNode) {
-      const accountReviewTransform =
-        accountReviewPanelNode.getComponent(UITransform) ?? accountReviewPanelNode.addComponent(UITransform);
-      accountReviewTransform.setContentSize(Math.max(320, Math.min(420, visibleSize.width - 56)), Math.max(360, visibleSize.height - 96));
-      accountReviewPanelNode.setPosition(0, 0, 4);
-    }
-
-    if (equipmentPanelNode) {
-      const equipmentPanelTransform =
-        equipmentPanelNode.getComponent(UITransform) ?? equipmentPanelNode.addComponent(UITransform);
-      equipmentPanelTransform.setContentSize(Math.max(360, Math.min(460, visibleSize.width - 56)), Math.max(420, visibleSize.height - 96));
-      equipmentPanelNode.setPosition(0, 0, 4);
-    }
-
-    if (campaignPanelNode) {
-      const campaignPanelTransform =
-        campaignPanelNode.getComponent(UITransform) ?? campaignPanelNode.addComponent(UITransform);
-      campaignPanelTransform.setContentSize(Math.max(380, Math.min(500, visibleSize.width - 56)), Math.max(480, visibleSize.height - 96));
-      campaignPanelNode.setPosition(0, 0, 5);
-    }
-
-    if (settingsPanelNode) {
-      const settingsPanelTransform =
-        settingsPanelNode.getComponent(UITransform) ?? settingsPanelNode.addComponent(UITransform);
-      settingsPanelTransform.setContentSize(Math.max(360, Math.min(460, visibleSize.width - 64)), Math.max(440, visibleSize.height - 96));
-      settingsPanelNode.setPosition(0, 0, 6);
-    }
-
-    if (settingsButtonNode) {
-      const buttonTransform = settingsButtonNode.getComponent(UITransform) ?? settingsButtonNode.addComponent(UITransform);
-      buttonTransform.setContentSize(58, 58);
-      settingsButtonNode.setPosition(visibleSize.width / 2 - margin - 34, visibleSize.height / 2 - margin - 34, 7);
-      this.renderSettingsButton();
-    }
-  }
-
-  private handleHudActionInput(...args: unknown[]): void {
-    this.audioRuntime.unlock();
-    const event = args[0] as EventTouch | EventMouse | undefined;
-    if (!event) {
-      return;
-    }
-
-    const visibleSize = view.getVisibleSize();
-    const centeredX = event.getUILocation().x - visibleSize.width / 2;
-    const centeredY = event.getUILocation().y - visibleSize.height / 2;
-    const settingsButtonNode = this.node.getChildByName(SETTINGS_BUTTON_NODE_NAME);
-    if (this.pointInRootNode(centeredX, centeredY, settingsButtonNode)) {
-      this.toggleSettingsPanel();
-      this.inputDebug = "button settings-fab";
-      return;
-    }
-
-    const settingsPanelNode = this.node.getChildByName(SETTINGS_PANEL_NODE_NAME);
-    const settingsPanelTransform = settingsPanelNode?.getComponent(UITransform) ?? null;
-    if (this.settingsView.open && settingsPanelNode && settingsPanelTransform) {
-      const settingsLocalX = centeredX - settingsPanelNode.position.x;
-      const settingsLocalY = centeredY - settingsPanelNode.position.y;
-      if (
-        settingsLocalX >= -settingsPanelTransform.width / 2
-        && settingsLocalX <= settingsPanelTransform.width / 2
-        && settingsLocalY >= -settingsPanelTransform.height / 2
-        && settingsLocalY <= settingsPanelTransform.height / 2
-      ) {
-        const action = this.settingsPanel?.dispatchPointerUp(settingsLocalX, settingsLocalY) ?? null;
-        if (action) {
-          this.inputDebug = `button ${action}`;
-        }
-        return;
-      }
-    }
-
-    if (this.showLobby) {
-      return;
-    }
-
-    const hudNode = this.node.getChildByName(HUD_NODE_NAME);
-    const hudTransform = hudNode?.getComponent(UITransform) ?? null;
-    if (!hudNode || !hudTransform) {
-      return;
-    }
-
-    const hudLocalX = centeredX - hudNode.position.x;
-    const hudLocalY = centeredY - hudNode.position.y;
-    if (
-      hudLocalX < -hudTransform.width / 2 ||
-      hudLocalX > hudTransform.width / 2 ||
-      hudLocalY < -hudTransform.height / 2 ||
-      hudLocalY > hudTransform.height / 2
-    ) {
-      return;
-    }
-
-    const action = this.hudPanel?.dispatchPointerUp(hudLocalX, hudLocalY) ?? null;
-    if (!action) {
-      return;
-    }
-
-    this.inputDebug = `button ${action}`;
-  }
-
-  private scheduleFogPulseTick(): void {
-    this.scheduleOnce(() => {
-      if (!this.fogPulseEnabled) {
-        return;
-      }
-
-      this.fogPulsePhase = (this.fogPulsePhase + 1) % 2;
-      this.mapBoard?.setFogPulsePhase(this.fogPulsePhase);
-      if (this.lastUpdate) {
-        this.mapBoard?.render(this.lastUpdate);
-      }
-      this.scheduleFogPulseTick();
-    }, Math.max(0.2, this.fogPulseIntervalSeconds));
-  }
-
-  private pushLog(line: string): void {
-    this.logLines.unshift(line);
-    this.logLines = this.logLines.slice(0, 8);
-  }
-
-  private emitPrimaryClientTelemetry(event: PrimaryClientTelemetryEvent | PrimaryClientTelemetryEvent[] | null): void {
-    emitPrimaryClientTelemetryForRoot(this as unknown as Record<string, any>, event);
-  }
-
-  private ensureAnalyticsSessionId(): string {
-    return ensureAnalyticsSessionIdForRoot(this as unknown as Record<string, any>);
-  }
-
-  private createClientAnalyticsContext(roomId = this.roomId): ClientAnalyticsContext {
-    return createClientAnalyticsContextForRoot(this as unknown as Record<string, any>, roomId);
-  }
-
-  private trackClientAnalyticsEvent<Name extends
-    | "session_start"
-    | "battle_start"
-    | "battle_end"
-    | "mission_started"
-    | "quest_complete"
-    | "tutorial_step"
-    | "experiment_exposure"
-    | "shop_open"
-    | "purchase_initiated"
-    | "purchase_attempt"
-    | "asset_load_failed"
-    | "client_perf_degraded"
-    | "client_runtime_error"
-  >(
-    name: Name,
-    payload: Record<string, unknown>,
-    roomId = this.roomId
-  ): void {
-    trackClientAnalyticsEventForRoot(this as unknown as Record<string, any>, name, payload, roomId);
-  }
-
-  private trackAssetLoadFailureAnalytics(event: AssetLoadFailureEvent): void {
-    trackAssetLoadFailureAnalyticsForRoot(this as unknown as Record<string, any>, event);
-  }
-
-  private handleAssetLoadFailure(event: AssetLoadFailureEvent): void {
-    handleAssetLoadFailureForRoot(this as unknown as Record<string, any>, event);
-  }
-
-  private reportClientRuntimeError(input: {
-    errorCode: StructuredErrorCode | "session_disconnect" | "client_error_boundary_triggered";
-    severity: "error" | "fatal";
-    stage: string;
-    recoverable: boolean;
-    message: string;
-    detail?: string;
-    roomId?: string | null;
-  }): void {
-    reportClientRuntimeErrorForRoot(this as unknown as Record<string, any>, input);
-  }
-
-  private maybeReportSessionRuntimeError(error: unknown, stage: string): void {
-    maybeReportSessionRuntimeErrorForRoot(this as unknown as Record<string, any>, error, stage);
-  }
-
-  private bindGlobalErrorBoundary(): (() => void) | null {
-    return bindGlobalErrorBoundaryForRoot(this as unknown as Record<string, any>);
-  }
-
-  private createTelemetryContext(heroId?: string | null): { roomId: string; playerId: string; heroId?: string } {
-    return createTelemetryContextForRoot(this as unknown as Record<string, any>, heroId);
-  }
-
-  private maybeEmitShopOpenAnalytics(): void {
-    maybeEmitShopOpenAnalyticsForRoot(this as unknown as Record<string, any>);
-  }
-
-  private maybeEmitExperimentExposureAnalytics(profile: CocosPlayerAccountProfile): void {
-    maybeEmitExperimentExposureAnalyticsForRoot(this as unknown as Record<string, any>, profile);
-  }
-
-  private maybeEmitQuestCompleteAnalytics(previousProfile: CocosPlayerAccountProfile, profile: CocosPlayerAccountProfile): void {
-    maybeEmitQuestCompleteAnalyticsForRoot(this as unknown as Record<string, any>, previousProfile, profile);
-  }
-
-  private trackPurchaseInitiated(product: ShopProduct, surface: "lobby" | "battle_pass"): void {
-    trackPurchaseInitiatedForRoot(this as unknown as Record<string, any>, product, surface);
-  }
-
-  private setBattleFeedback(feedback: CocosBattleFeedbackView | null, durationMs = BATTLE_FEEDBACK_DURATION_MS): void {
-    if (!feedback) {
-      return;
-    }
-
-    this.battleFeedback = {
-      ...feedback,
-      expiresAt: Date.now() + durationMs
-    };
-  }
-
-  private activeHero(): HeroView | null {
-    return this.lastUpdate?.world.ownHeroes[0] ?? null;
-  }
-
-  private controlledBattleCamp(): "attacker" | "defender" | null {
-    const battle = this.lastUpdate?.battle;
-    const heroId = this.activeHero()?.id;
-    if (!battle || !heroId) {
-      return null;
-    }
-
-    if (battle.worldHeroId === heroId) {
-      return "attacker";
-    }
-
-    if (battle.defenderHeroId === heroId) {
-      return "defender";
-    }
-
-    return null;
-  }
-
-  private opposingBattleCamp(camp: "attacker" | "defender" | null): "attacker" | "defender" | null {
-    if (!camp) {
-      return null;
-    }
-
-    return camp === "attacker" ? "defender" : "attacker";
-  }
-
-  private syncSelectedBattleTarget(): void {
-    const battle = this.lastUpdate?.battle;
-    const enemyCamp = this.opposingBattleCamp(this.controlledBattleCamp());
-    if (!battle || !enemyCamp) {
-      this.selectedBattleTargetId = null;
-      return;
-    }
-
-    const targets = Object.values(battle.units).filter((unit) => unit.camp === enemyCamp && unit.count > 0);
-    if (targets.length === 0) {
-      this.selectedBattleTargetId = null;
-      return;
-    }
-
-    if (!this.selectedBattleTargetId || !targets.some((target) => target.id === this.selectedBattleTargetId)) {
-      this.selectedBattleTargetId = targets[0]?.id ?? null;
-    }
-  }
-
-  private selectedInteractionTile(): PlayerTileView | null {
-    const buildingId = this.selectedInteractionBuildingId;
-    if (!buildingId) {
-      return null;
-    }
-
-    return this.lastUpdate?.world.map.tiles.find((tile) => tile.building?.id === buildingId) ?? null;
-  }
-
-  private clearSelectedInteractionBuilding(): void {
-    this.selectedInteractionBuildingId = null;
-  }
-
-
-  private buildHudInteractionState(): VeilHudRenderState["interaction"] {
-    const hero = this.activeHero();
-    const tile = this.selectedInteractionTile();
-    const building = tile?.building;
-    if (!hero || !tile || !building) {
-      return null;
-    }
-
-    const heroDistance = Math.abs(hero.position.x - tile.position.x) + Math.abs(hero.position.y - tile.position.y);
-    if (heroDistance > 1) {
-      return null;
-    }
-
-    const actions: NonNullable<VeilHudRenderState["interaction"]>["actions"] = [];
-    const tierLabel = `等级 ${building.tier}${building.maxTier ? `/${building.maxTier}` : ""}`;
-    const trackId = building.kind === "recruitment_post" ? "castle" : building.kind === "resource_mine" ? "mine" : null;
-    const maxTier = building.maxTier ?? (trackId === "castle" ? 3 : trackId === "mine" ? 2 : building.tier);
-    const upgradeStep =
-      building.kind === "recruitment_post" || building.kind === "resource_mine"
-        ? getBuildingUpgradeConfig()[trackId!].find((step) => step.fromTier === building.tier) ?? null
-        : null;
-
-    if (heroDistance === 0) {
-      if (building.kind === "recruitment_post") {
-        actions.push({ id: "recruit", label: "招募部队" });
-      } else if (building.kind === "attribute_shrine" || building.kind === "watchtower") {
-        actions.push({ id: "visit", label: "访问建筑" });
-      } else if (building.kind === "resource_mine") {
-        actions.push({ id: "claim", label: "采集矿场" });
-      }
-    }
-
-    if ((building.kind === "recruitment_post" || building.kind === "resource_mine") && building.ownerPlayerId === hero.playerId) {
-      if (building.tier >= maxTier) {
-        return {
-          title: building.label,
-          detail: `${tierLabel} · 已满级`,
-          actions
-        };
-      }
-
-      if (upgradeStep) {
-        actions.push({ id: "upgrade", label: `升级建筑 · ${building.tier}→${upgradeStep.toTier}` });
-        return {
-          title: building.label,
-          detail: `${tierLabel} · 升级花费 ${formatUpgradeCostLabel(upgradeStep.cost)}`,
-          actions
-        };
-      }
-    }
-
-    return {
-      title: building.label,
-      detail:
-        building.kind === "resource_mine"
-          ? `${tierLabel} · ${formatResourceKindLabel(building.resourceKind)} +${building.income}`
-          : building.kind === "recruitment_post"
-            ? `${tierLabel} · 可招募 ${building.availableCount} 单位`
-            : tierLabel,
-      actions
-    };
-  }
-
-  private async executeBuildingInteraction(tile: PlayerTileView, actionId: "recruit" | "visit" | "claim" | "upgrade"): Promise<void> {
-    const hero = this.activeHero();
-    const building = tile.building;
-    if (!hero || !building || !this.session) {
-      return;
-    }
-
-    this.moveInFlight = true;
-    const predictionAction =
-      actionId === "recruit"
-        ? { type: "hero.recruit", heroId: hero.id, buildingId: building.id } as const
-        : actionId === "visit"
-          ? { type: "hero.visit", heroId: hero.id, buildingId: building.id } as const
-          : actionId === "claim"
-            ? { type: "hero.claimMine", heroId: hero.id, buildingId: building.id } as const
-            : { type: "hero.upgradeBuilding", heroId: hero.id, buildingId: building.id } as const;
-    const predictionLabel =
-      actionId === "recruit"
-        ? `预演招募 ${building.kind === "recruitment_post" ? building.availableCount : 0} 单位`
-        : actionId === "visit"
-          ? building.kind === "attribute_shrine"
-            ? `预演获得 ${formatHeroStatBonus(building.bonus)}`
-            : building.kind === "watchtower"
-              ? `预演提高视野 ${building.visionBonus}`
-              : "预演访问建筑"
-          : actionId === "claim"
-            ? building.kind === "resource_mine"
-              ? `预演占领矿场，改为每日产出 ${building.income} ${formatResourceKindLabel(building.resourceKind)}`
-              : "预演矿场采集"
-            : "预演建筑升级";
-    this.applyPrediction(predictionAction, predictionLabel);
-    this.renderView();
-
-    try {
-      this.mapBoard?.playHeroAnimation("attack");
-      const update =
-        actionId === "recruit"
-          ? await this.session.recruit(hero.id, building.id)
-          : actionId === "visit"
-            ? await this.session.visitBuilding(hero.id, building.id)
-            : actionId === "claim"
-              ? await this.session.claimMine(hero.id, building.id)
-              : await this.session.upgradeBuilding(hero.id, building.id);
-      this.clearSelectedInteractionBuilding();
-      await this.applySessionUpdate(update);
-      this.pushSessionActionOutcome(update, {
-        successMessage:
-          actionId === "recruit"
-            ? "招募已结算。"
-            : actionId === "visit"
-              ? building.kind === "watchtower"
-                ? "瞭望塔访问已结算。"
-                : "建筑访问已结算。"
-              : actionId === "claim"
-                ? "矿场占领已结算。"
-                : "建筑升级已结算。",
-        rejectedLabel:
-          actionId === "recruit"
-            ? "招募"
-            : actionId === "visit"
-              ? "访问"
-              : actionId === "claim"
-                ? "矿场占领"
-                : "建筑升级"
-      });
-    } catch (error) {
-      this.rollbackPrediction(error instanceof Error ? error.message : `${actionId}失败。`);
-    } finally {
-      this.moveInFlight = false;
-      this.renderView();
-    }
-  }
-
-  private async moveHeroToTile(tile: PlayerTileView): Promise<void> {
-    if (this.moveInFlight) {
-      return;
-    }
-
-    if (!this.session) {
-      await this.connect();
-      return;
-    }
-
-    const hero = this.activeHero();
-    if (!hero) {
-      this.pushLog("当前快照里没有可控制的英雄。");
-      this.renderView();
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.pushLog("当前正在战斗，暂时无法移动。");
-      this.renderView();
-      return;
-    }
-
-    const reachableTiles = await this.ensureReachableTiles(hero.id);
-    const clickedCurrentTile = hero.position.x === tile.position.x && hero.position.y === tile.position.y;
-    if (!clickedCurrentTile && tile.building) {
-      const interactionDistance = Math.abs(hero.position.x - tile.position.x) + Math.abs(hero.position.y - tile.position.y);
-      if (interactionDistance <= 1) {
-        this.selectedInteractionBuildingId = tile.building.id;
-        this.pushLog(`已选中 ${tile.building.label}，请在 HUD 中确认操作。`);
-        this.mapBoard?.pulseObject(tile.position, 1.2, 0.24);
-        this.renderView();
-        return;
-      }
-    }
-
-    if (clickedCurrentTile) {
-      if (!tile.resource && !tile.building) {
-        this.pushLog("英雄已经站在这里了。");
-        this.mapBoard?.pulseTile(tile.position, 1.04, 0.14);
-        this.mapBoard?.showTileFeedback(tile.position, "原地", 0.45);
-        this.renderView();
-        return;
-      }
-
-      if (tile.building) {
-        this.selectedInteractionBuildingId = tile.building.id;
-        this.pushLog(`已选中 ${tile.building.label}，请在 HUD 中确认操作。`);
-        this.mapBoard?.pulseObject(tile.position, 1.2, 0.24);
-        this.renderView();
-        return;
-      }
-
-      const resource = tile.resource;
-      if (!resource) {
-        this.clearSelectedInteractionBuilding();
-        this.pushLog("当前格子没有可采集资源。");
-        this.renderView();
-        return;
-      }
-
-      this.moveInFlight = true;
-      const resourceLabel = resource.kind === "gold" ? "金币" : resource.kind === "wood" ? "木材" : resource.kind === "ore" ? "矿石" : resource.kind;
-      this.pushLog(`正在采集 ${resourceLabel} +${resource.amount}`);
-      this.mapBoard?.pulseTile(tile.position, 1.12, 0.22);
-      this.mapBoard?.pulseObject(tile.position, 1.2, 0.24);
-      this.applyPrediction(
-        {
-          type: "hero.collect",
-          heroId: hero.id,
-          position: tile.position
-        },
-        `预演采集 ${resourceLabel} +${resource.amount}`
-      );
-      this.renderView();
-
-      try {
-        this.mapBoard?.playHeroAnimation("attack");
-        const update = await this.session.collect(hero.id, tile.position);
-        await this.applySessionUpdate(update);
-        this.pushSessionActionOutcome(update, {
-          successMessage: "采集已结算。",
-          rejectedLabel: "采集"
-        });
-      } catch (error) {
-        this.rollbackPrediction(error instanceof Error ? error.message : "采集失败。");
-      } finally {
-        this.moveInFlight = false;
-        this.renderView();
-      }
-      return;
-    }
-
-    this.clearSelectedInteractionBuilding();
-    if (hero.move.remaining <= 0) {
-      this.pushLog(`${hero.name} 今天已经没有移动力了。`);
-      this.predictionStatus = "今天已经没有移动点了。";
-      this.mapBoard?.pulseTile(hero.position, 1.06, 0.18);
-      this.mapBoard?.showTileFeedback(hero.position, "耗尽", 0.7);
-      this.renderView();
-      return;
-    }
-
-    const target = reachableTiles.find((node) => node.x === tile.position.x && node.y === tile.position.y) ?? null;
-    if (!target) {
-      const movePrediction = this.lastUpdate
-        ? predictSharedPlayerWorldAction(this.lastUpdate.world, {
-            type: "hero.move",
-            heroId: hero.id,
-            destination: tile.position
-          })
-        : null;
-      const moveFeedback = describeMoveAttemptFeedback(tile.position, movePrediction?.reason);
-      this.pushLog(moveFeedback.message);
-      if (movePrediction?.reason === "not_enough_move_points") {
-        this.predictionStatus = moveFeedback.message;
-      }
-      this.mapBoard?.pulseTile(tile.position, 1.08, 0.18);
-      this.mapBoard?.showTileFeedback(tile.position, moveFeedback.tileFeedback, 0.6);
-      this.renderView();
-      return;
-    }
-
-    this.moveInFlight = true;
-    this.pushLog(`正在移动 ${hero.name} -> (${target.x}, ${target.y})`);
-    this.mapBoard?.pulseTile(target, tile.occupant?.kind ? 1.1 : 1.06, 0.18);
-    if (tile.resource || tile.occupant) {
-      this.mapBoard?.pulseObject(target, tile.occupant?.kind ? 1.18 : 1.14, 0.22);
-    }
-    this.applyPrediction(
-      {
-        type: "hero.move",
-        heroId: hero.id,
-        destination: target
-      },
-      tile.occupant?.kind === "neutral" || tile.occupant?.kind === "hero"
-        ? "正在预演遭遇..."
-        : "正在预演移动..."
-    );
-    this.renderView();
-
-    try {
-      this.mapBoard?.playHeroAnimation("move");
-      const update = await this.session.moveHero(hero.id, target);
-      await this.applySessionUpdate(update);
-      this.pushSessionActionOutcome(update, {
-        successMessage: "移动已结算。",
-        rejectedLabel: "移动"
-      });
-    } catch (error) {
-      this.rollbackPrediction(error instanceof Error ? error.message : "移动失败。");
-    } finally {
-      this.moveInFlight = false;
-      this.renderView();
-    }
-  }
-
-  private async ensureReachableTiles(heroId: string): Promise<Vec2[]> {
-    if (this.lastUpdate?.reachableTiles.length) {
-      return this.lastUpdate.reachableTiles;
-    }
-
-    if (!this.session) {
-      return [];
-    }
-
-    const reachableTiles = await this.session.listReachable(heroId);
-    if (this.lastUpdate) {
-      this.lastUpdate = {
-        ...this.lastUpdate,
-        reachableTiles
-      };
-    }
-
-    return reachableTiles;
-  }
-
-  private async startNewRun(): Promise<void> {
-    if (this.moveInFlight || this.battleActionInFlight) {
-      return;
-    }
-
-    const previousSession = this.session;
-    const previousSessionEpoch = this.sessionEpoch;
-    const previousRoomId = this.roomId;
-    const previousSeed = this.seed;
-    const nextRoomId = `run-${Date.now().toString(36).slice(-6)}`;
-    const nextSeed = this.seed + 1;
-    let freshSession: VeilCocosSession | null = null;
-    const nextSessionEpoch = this.bumpSessionEpoch();
-
-    this.pendingPrediction = null;
-    this.selectedBattleTargetId = null;
-    this.moveInFlight = false;
-    this.battleActionInFlight = false;
-    this.predictionStatus = "正在开启新一局...";
-    this.inputDebug = "input waiting";
-    this.timelineEntries = [];
-    this.primaryClientTelemetry = [];
-    this.logLines = [`正在创建新房间 ${nextRoomId} ...`];
-    this.renderView();
-
-    try {
-      freshSession = await resolveVeilRootRuntime().createSession(
-        nextRoomId,
-        this.playerId,
-        nextSeed,
-        this.createSessionOptions(nextSessionEpoch)
-      );
-      if (!this.isActiveSessionEpoch(nextSessionEpoch)) {
-        await freshSession.dispose().catch(() => undefined);
-        return;
-      }
-
-      const freshUpdate = await freshSession.snapshot();
-      if (!this.isActiveSessionEpoch(nextSessionEpoch)) {
-        await freshSession.dispose().catch(() => undefined);
-        return;
-      }
-
-      this.session = freshSession;
-      this.roomId = nextRoomId;
-      this.seed = nextSeed;
-      this.syncBrowserRoomQuery(nextRoomId);
-      this.pushLog(`已进入新房间 ${nextRoomId}。`);
-      await this.applySessionUpdate(freshUpdate);
-
-      if (previousSession) {
-        await previousSession.dispose().catch(() => undefined);
-      }
-      return;
-    } catch (error) {
-      if (freshSession) {
-        await freshSession.dispose().catch(() => undefined);
-      }
-      this.sessionEpoch = previousSessionEpoch;
-      this.session = previousSession;
-      this.roomId = previousRoomId;
-      this.seed = previousSeed;
-      const failureMessage = this.describeSessionError(error, "开启新一局失败。");
-      if (error instanceof Error && error.message === "upgrade_required") {
-        await this.handleForcedUpgrade(failureMessage);
-        return;
-      }
-      this.pushLog(failureMessage);
-      this.predictionStatus = failureMessage;
-      this.renderView();
-    }
-  }
-
-  private async refreshLobbyRoomList(): Promise<void> {
-    if (this.lobbyLoading || this.lobbyEntering) {
-      return;
-    }
-
-    this.lobbyLoading = true;
-    this.lobbyStatus = "正在刷新可加入房间...";
-    this.renderView();
-
-    try {
-      const [rooms, announcements, maintenanceMode] = await Promise.all([
-        resolveVeilRootRuntime().loadLobbyRooms(this.remoteUrl),
-        resolveVeilRootRuntime().loadAnnouncements(this.remoteUrl).catch(() => []),
-        resolveVeilRootRuntime().loadMaintenanceMode(this.remoteUrl).catch(() => null)
-      ]);
-      this.lobbyRooms = rooms;
-      this.lobbyAnnouncements = announcements;
-      this.lobbyMaintenanceMode = maintenanceMode;
-      this.lobbyStatus = maintenanceMode?.active
-        ? `${maintenanceMode.title} · ${maintenanceMode.message}`
-        : rooms.length > 0
-          ? `发现 ${rooms.length} 个活跃房间，可直接加入或继续创建新房间。`
-          : "当前没有活跃房间，输入房间 ID 后点击“进入房间”即可创建新实例。";
-    } catch {
-      this.lobbyRooms = [];
-      this.lobbyStatus = "Lobby 服务暂不可达；仍可直接输入房间 ID，进入时会自动尝试远端房间并在失败后回退本地模式。";
-    } finally {
-      this.lobbyLoading = false;
-      this.renderView();
-    }
-  }
-
-  private async enterLobbyRoom(roomIdOverride?: string): Promise<void> {
-    if (this.lobbyEntering) {
-      return;
-    }
-
-    if (this.isMatchmakingActive()) {
-      this.lobbyStatus = "正在匹配中，请先取消当前队列。";
-      this.renderView();
-      return;
-    }
-
-    if (!this.ensurePrivacyConsentAccepted()) {
-      return;
-    }
-
-    const storage = this.readWebStorage();
-    const preferences = saveCocosLobbyPreferences(this.playerId, roomIdOverride ?? this.roomId, undefined, storage);
-    const displayName = rememberPreferredCocosDisplayName(preferences.playerId, this.displayName || preferences.playerId, storage);
-    this.playerId = preferences.playerId;
-    this.roomId = preferences.roomId;
-    this.displayName = displayName;
-    this.lobbyEntering = true;
-    this.lobbyStatus =
-      this.authMode === "account" && this.authToken
-        ? `正在使用账号 ${this.loginId || this.playerId} 进入房间 ${preferences.roomId}...`
-        : `正在登录游客账号并进入房间 ${preferences.roomId}...`;
-    this.renderView();
-
-    try {
-      let authSession: Awaited<ReturnType<typeof loginCocosGuestAuthSession>>;
-      if (this.authMode === "account" && this.authToken) {
-        const syncedSession = await resolveVeilRootRuntime().syncAuthSession(this.remoteUrl, {
-          storage,
-          session: readStoredCocosAuthSession(storage)
-        });
-        if (!syncedSession) {
-          throw new Error("cocos_request_failed:401");
-        }
-        authSession = syncedSession;
-      } else {
-        authSession = await resolveVeilRootRuntime().loginGuestAuthSession(this.remoteUrl, preferences.playerId, displayName, {
-          storage,
-          privacyConsentAccepted: this.privacyConsentAccepted
-        });
-      }
-      this.authToken = authSession.token ?? null;
-      this.playerId = authSession.playerId;
-      this.displayName = authSession.displayName;
-      this.authMode = authSession.authMode;
-      this.authProvider = authSession.provider ?? "guest";
-      this.loginId = authSession.loginId ?? "";
-      this.sessionSource = authSession.source;
-      await this.maybeClaimLaunchReferral(authSession);
-      saveCocosLobbyPreferences(authSession.playerId, preferences.roomId, undefined, storage);
-      this.resetSessionViewport(`正在进入房间 ${preferences.roomId} ...`);
-      this.showLobby = false;
-      this.syncBrowserRoomQuery(preferences.roomId);
-      this.syncWechatShareBridge();
-      this.lobbyStatus =
-        authSession.authMode === "account"
-          ? `账号 ${authSession.loginId ?? authSession.playerId} 登录成功，正在进入房间 ${preferences.roomId}...`
-          : authSession.source === "remote"
-            ? `游客登录成功，正在进入房间 ${preferences.roomId}...`
-            : `登录服务暂不可达，正在以本地游客档进入房间 ${preferences.roomId}...`;
-      this.renderView();
-      await this.connect();
-
-      if (this.upgradeRequired) {
-        this.renderView();
-        return;
-      }
-
-      if (!this.session && !this.lastUpdate) {
-        this.showLobby = true;
-        this.lobbyStatus = "进入房间失败，请稍后重试或刷新房间列表。";
-        this.renderView();
-        return;
-      }
-
-      this.commitAccountProfile(
-        createFallbackCocosPlayerAccountProfile(this.playerId, this.roomId, this.displayName),
-        false
-      );
-      this.renderView();
-    } catch (error) {
-      this.showLobby = true;
-      if (error instanceof Error && error.message === "cocos_request_failed:401") {
-        const storage = this.readWebStorage();
-        if (storage) {
-          clearStoredCocosAuthSession(storage);
-        }
-        this.authToken = null;
-        this.authMode = "guest";
-        this.authProvider = "guest";
-        this.loginId = "";
-        this.sessionSource = "none";
-      }
-      this.lobbyStatus =
-        error instanceof Error && error.message === "cocos_request_failed:401"
-          ? "账号会话已失效，请重新登录后再进入房间。"
-          : this.describeCocosAccountFlowError(error, "enter_room_failed");
-      this.renderView();
-    } finally {
-      this.lobbyEntering = false;
-    }
-  }
-
-  private isMatchmakingActive(): boolean {
-    return this.matchmakingStatus.status === "queued" || this.matchmakingJoinInFlight;
-  }
-
-  private updateMatchmakingStatus(status: MatchmakingStatusResponse, lobbyStatus?: string): void {
-    this.matchmakingStatus = status;
-    this.matchmakingView = buildMatchmakingStatusView(status);
-    if (lobbyStatus) {
-      this.lobbyStatus = lobbyStatus;
-    }
-  }
-
-  private stopMatchmakingPolling(): void {
-    this.matchmakingPollController?.stop();
-    this.matchmakingPollController = null;
-    if (this.matchmakingTimeoutHandle) {
-      clearTimeout(this.matchmakingTimeoutHandle);
-      this.matchmakingTimeoutHandle = null;
-    }
-  }
-
-  private startMatchmakingPolling(): void {
-    this.stopMatchmakingPolling();
-    this.matchmakingPollController = resolveVeilRootRuntime().startMatchmakingPolling(
-      this.remoteUrl,
-      (status) => {
-        void this.handleMatchmakingStatusUpdate(status);
-      },
-      {
-        pollIntervalMs: 3000,
-        stopOnMatched: true,
-        authSession: this.authToken
-          ? {
-              token: this.authToken,
-              playerId: this.playerId,
-              displayName: this.displayName || this.playerId,
-              authMode: this.authMode,
-              ...(this.loginId ? { loginId: this.loginId } : {}),
-              source: "remote"
-            }
-          : null
-      }
-    );
-    this.matchmakingTimeoutHandle = setTimeout(() => {
-      void this.handleMatchmakingTimeout();
-    }, this.matchmakingTimeoutMs);
-  }
-
-  private async ensureMatchmakingAuthSession(): Promise<void> {
-    const storage = this.readWebStorage();
-    if (this.authMode === "account" && this.authToken) {
-      const syncedSession = await resolveVeilRootRuntime().syncAuthSession(this.remoteUrl, {
-        storage,
-        session: readStoredCocosAuthSession(storage)
-      });
-      if (!syncedSession) {
-        throw new Error("cocos_request_failed:401");
-      }
-      this.authToken = syncedSession.token ?? null;
-      this.playerId = syncedSession.playerId;
-      this.displayName = syncedSession.displayName;
-      this.authMode = syncedSession.authMode;
-      this.authProvider = syncedSession.provider ?? "account-password";
-      this.loginId = syncedSession.loginId ?? "";
-      this.sessionSource = syncedSession.source;
-      await this.maybeClaimLaunchReferral(syncedSession);
-      return;
-    }
-
-    const authSession = await resolveVeilRootRuntime().loginGuestAuthSession(
-      this.remoteUrl,
-      this.playerId,
-      this.displayName || this.playerId,
-      {
-        storage,
-        privacyConsentAccepted: this.privacyConsentAccepted
-      }
-    );
-    this.authToken = authSession.token ?? null;
-    this.playerId = authSession.playerId;
-    this.displayName = authSession.displayName;
-    this.authMode = authSession.authMode;
-    this.authProvider = authSession.provider ?? "guest";
-    this.loginId = authSession.loginId ?? "";
-    this.sessionSource = authSession.source;
-    await this.maybeClaimLaunchReferral(authSession);
-  }
-
-  private async enterLobbyMatchmaking(): Promise<void> {
-    if (this.lobbyEntering || this.isMatchmakingActive()) {
-      return;
-    }
-
-    if (!this.ensurePrivacyConsentAccepted()) {
-      return;
-    }
-
-    this.lobbyEntering = true;
-    this.updateMatchmakingStatus({ status: "idle" }, "正在进入 PVP 匹配队列...");
-    this.renderView();
-
-    try {
-      await this.ensureMatchmakingAuthSession();
-      const rating = this.lobbyAccountProfile.eloRating ?? 1000;
-      const status = await resolveVeilRootRuntime().enqueueMatchmaking(this.remoteUrl, this.playerId, rating, {
-        getDisplayName: () => this.displayName || this.playerId,
-        getAuthToken: () => this.authToken
-      });
-      this.updateMatchmakingStatus(status, this.describeMatchmakingStatus(status));
-      void requestCocosWechatSubscribeConsent();
-      this.startMatchmakingPolling();
-    } catch (error) {
-      this.updateMatchmakingStatus({ status: "idle" });
-      this.lobbyStatus = this.describeMatchmakingError(error);
-    } finally {
-      this.lobbyEntering = false;
-      this.renderView();
-    }
-  }
-
-  private async cancelLobbyMatchmaking(): Promise<void> {
-    if (!this.isMatchmakingActive() || this.lobbyEntering) {
-      return;
-    }
-
-    this.lobbyEntering = true;
-    this.lobbyStatus = "正在取消 PVP 匹配...";
-    this.renderView();
-
-    try {
-      await resolveVeilRootRuntime().cancelMatchmaking(this.remoteUrl, this.playerId, {
-        getDisplayName: () => this.displayName || this.playerId,
-        getAuthToken: () => this.authToken
-      });
-      this.stopMatchmakingPolling();
-      this.updateMatchmakingStatus({ status: "idle" }, "已取消当前匹配队列。");
-    } catch (error) {
-      this.lobbyStatus = this.describeMatchmakingError(error);
-    } finally {
-      this.lobbyEntering = false;
-      this.renderView();
-    }
-  }
-
-  private async handleMatchmakingStatusUpdate(status: MatchmakingStatusResponse): Promise<void> {
-    if (status.status === "idle") {
-      this.stopMatchmakingPolling();
-    }
-    this.updateMatchmakingStatus(status, this.describeMatchmakingStatus(status));
-    this.renderView();
-
-    if (status.status === "matched" && !this.matchmakingJoinInFlight) {
-      await this.enterMatchedRoom(status);
-    }
-  }
-
-  private async handleMatchmakingTimeout(): Promise<void> {
-    if (!this.isMatchmakingActive()) {
-      return;
-    }
-
-    this.stopMatchmakingPolling();
-    try {
-      await resolveVeilRootRuntime().cancelMatchmaking(this.remoteUrl, this.playerId, {
-        getDisplayName: () => this.displayName || this.playerId,
-        getAuthToken: () => this.authToken
-      });
-    } catch {
-      // Keep the timeout surfaced locally even if remote dequeue fails.
-    }
-    this.updateMatchmakingStatus({ status: "idle" }, "匹配超时，请稍后重试。");
-    this.renderView();
-  }
-
-  private async enterMatchedRoom(status: Extract<MatchmakingStatusResponse, { status: "matched" }>): Promise<void> {
-    this.matchmakingJoinInFlight = true;
-    this.stopMatchmakingPolling();
-    this.lobbyStatus = `匹配成功，正在进入房间 ${status.roomId}...`;
-    this.renderView();
-
-    try {
-      this.roomId = status.roomId;
-      this.seed = status.seedOverride;
-      this.resetSessionViewport(`正在进入匹配房间 ${status.roomId} ...`);
-      this.showLobby = false;
-      this.syncBrowserRoomQuery(status.roomId);
-      this.syncWechatShareBridge();
-      await this.connect();
-      if (!this.session && !this.lastUpdate) {
-        throw new Error("enter_room_failed");
-      }
-      this.updateMatchmakingStatus({ status: "idle" }, `已进入匹配房间 ${status.roomId}。`);
-    } catch (error) {
-      this.showLobby = true;
-      this.updateMatchmakingStatus({ status: "idle" }, this.describeMatchmakingError(error));
-    } finally {
-      this.matchmakingJoinInFlight = false;
-      this.renderView();
-    }
-  }
-
-  private describeMatchmakingStatus(status: MatchmakingStatusResponse): string {
-    const view = buildMatchmakingStatusView(status);
-    if (status.status === "queued") {
-      return `${view.statusLabel} ${view.queuePositionLabel}，${view.waitEstimateLabel}`;
-    }
-    if (status.status === "matched") {
-      return view.matchedLabel ? `${view.statusLabel} · ${view.matchedLabel}` : view.statusLabel;
-    }
-    return view.statusLabel;
-  }
-
-  private describeMatchmakingError(error: unknown): string {
-    if (error instanceof Error && error.message === "cocos_request_failed:401") {
-      return "匹配会话已失效，请重新登录后再试。";
-    }
-    if (error instanceof Error) {
-      return error.message;
-    }
-    return "matchmaking_failed";
-  }
-
-  private async loginLobbyAccount(): Promise<void> {
-    if (this.lobbyEntering) {
-      return;
-    }
-
-    const primaryProvider = this.primaryLoginProvider();
-    if (!primaryProvider.available) {
-      this.lobbyStatus = primaryProvider.message;
-      this.renderView();
-      return;
-    }
-
-    if (primaryProvider.id === "wechat-mini-game") {
-      await this.loginLobbyWechatMiniGame();
-      return;
-    }
-
-    const promptRef = globalThis.prompt;
-    if (typeof promptRef !== "function") {
-      this.lobbyStatus = "当前运行环境不支持弹出式输入，请先在浏览器调试壳完成账号登录，或复用已缓存会话。";
-      this.renderView();
-      return;
-    }
-
-    const nextLoginId = promptRef("输入登录 ID", this.loginId || "")?.trim();
-    if (nextLoginId === undefined) {
-      return;
-    }
-    const loginIdError = validateAccountLifecycleRequest("registration", nextLoginId);
-    if (loginIdError) {
-      this.lobbyStatus = loginIdError.message;
-      this.renderView();
-      return;
-    }
-
-    const password = promptRef("输入账号口令", "");
-    if (password == null) {
-      return;
-    }
-    const passwordError = validateAccountPassword(password, "password", "账号口令");
-    if (passwordError) {
-      this.lobbyStatus = passwordError.message;
-      this.renderView();
-      return;
-    }
-
-    if (!this.ensurePrivacyConsentAccepted()) {
-      return;
-    }
-
-    const storage = this.readWebStorage();
-    this.lobbyEntering = true;
-    this.lobbyStatus = `正在使用账号 ${nextLoginId.toLowerCase()} 登录并进入房间 ${this.roomId}...`;
-    this.renderView();
-
-    try {
-      const authSession = await loginWithCocosProvider(
-        this.remoteUrl,
-        {
-          provider: "account-password",
-          loginId: nextLoginId,
-          password
-        },
-        {
-          storage,
-          privacyConsentAccepted: this.privacyConsentAccepted
-        }
-      );
-      this.authToken = authSession.token ?? null;
-      this.playerId = authSession.playerId;
-      this.displayName = authSession.displayName;
-      this.authMode = authSession.authMode;
-      this.authProvider = authSession.provider ?? "account-password";
-      this.loginId = authSession.loginId ?? nextLoginId.toLowerCase();
-      this.sessionSource = authSession.source;
-      this.syncWechatShareBridge();
-      this.lobbyStatus = `账号 ${this.loginId} 登录成功，正在同步全局仓库并进入房间 ${this.roomId}...`;
-      saveCocosLobbyPreferences(authSession.playerId, this.roomId, undefined, storage);
-      this.renderView();
-      await this.refreshLobbyAccountProfile();
-      this.lobbyEntering = false;
-      await this.enterLobbyRoom(this.roomId);
-    } catch (error) {
-      this.lobbyEntering = false;
-      this.lobbyStatus = this.describeCocosAccountFlowError(error, "account_login_failed");
-      this.renderView();
-    }
-  }
-
-  private parseCocosRequestFailure(error: unknown): { status: number; code: string } | null {
-    if (!(error instanceof Error)) {
-      return null;
-    }
-
-    const matched = /^cocos_request_failed:(\d+):(.+)$/.exec(error.message);
-    if (!matched) {
-      return null;
-    }
-
-    return {
-      status: Number(matched[1]),
-      code: matched[2] ?? "unknown"
-    };
-  }
-
-  private describeCocosAccountFlowError(
-    error: unknown,
-    fallback: string,
-    options: {
-      invalidTokenCode?: string;
-    } = {}
-  ): string {
-    const failure = this.parseCocosRequestFailure(error);
-    if (!failure) {
-      return error instanceof Error ? error.message : fallback;
-    }
-
-    const message = describeAccountAuthFailure(failure, options);
-    if (message) {
-      return message;
-    }
-
-    return error instanceof Error ? error.message : fallback;
-  }
-
-  private ensurePrivacyConsentAccepted(): boolean {
-    const privacyConsentError = validatePrivacyConsentAccepted(this.privacyConsentAccepted);
-    if (!privacyConsentError) {
-      return true;
-    }
-
-    this.lobbyStatus = privacyConsentError.message;
-    this.renderView();
-    return false;
-  }
-
-  private async registerLobbyAccount(): Promise<void> {
-    this.openLobbyAccountFlow("registration");
-  }
-
-  private async recoverLobbyAccountPassword(): Promise<void> {
-    this.openLobbyAccountFlow("recovery");
-  }
-
-  private async loginLobbyWechatMiniGame(): Promise<void> {
-    if (!this.ensurePrivacyConsentAccepted()) {
-      return;
-    }
-
-    const storage = this.readWebStorage();
-    this.lobbyEntering = true;
-    this.lobbyStatus = "正在调用 wx.login() 并交换小游戏会话...";
-    this.renderView();
-
-    try {
-      const authSession = await loginCocosWechatAuthSession(this.remoteUrl, this.playerId, this.displayName || this.playerId, {
-        storage,
-        wx: (globalThis as { wx?: { login?: ((options: unknown) => void) | undefined } }).wx ?? null,
-        exchangePath: this.loginRuntimeConfig.wechatMiniGame.exchangePath,
-        ...(this.loginRuntimeConfig.wechatMiniGame.mockCode ? { mockCode: this.loginRuntimeConfig.wechatMiniGame.mockCode } : {}),
-        ...(this.authToken ? { authToken: this.authToken } : {}),
-        ...(this.privacyConsentAccepted ? { privacyConsentAccepted: true } : {}),
-        ...(this.activeAccountFlow === "registration" && this.wechatMinorProtectionSelection !== "unknown"
-          ? { minorProtection: { isAdult: this.wechatMinorProtectionSelection === "adult" } }
-          : {})
-      });
-      this.authToken = authSession.token ?? null;
-      this.playerId = authSession.playerId;
-      this.displayName = authSession.displayName;
-      this.authMode = authSession.authMode;
-      this.authProvider = authSession.provider ?? "wechat-mini-game";
-      this.loginId = authSession.loginId ?? "";
-      this.sessionSource = authSession.source;
-      if (this.activeAccountFlow === "registration") {
-        this.activeAccountFlow = null;
-        this.wechatMinorProtectionSelection = "unknown";
-      }
-      this.syncWechatShareBridge();
-      this.lobbyStatus = "微信小游戏登录已连通，正在同步会话并进入房间...";
-      saveCocosLobbyPreferences(authSession.playerId, this.roomId, undefined, storage);
-      this.renderView();
-      await this.refreshLobbyAccountProfile();
-      this.lobbyEntering = false;
-      await this.enterLobbyRoom(this.roomId);
-    } catch (error) {
-      this.lobbyEntering = false;
-      this.lobbyStatus =
-        error instanceof Error && error.message === "wechat_login_unavailable"
-          ? "当前小游戏壳未暴露 wx.login()，且也未配置 mock code。"
-          : error instanceof Error && error.message === "cocos_request_failed:501"
-            ? "服务端已预留小游戏登录交换接口，但当前未启用。"
-            : error instanceof Error && error.message === "cocos_request_failed:401"
-              ? "小游戏登录 code 校验失败，请刷新后重试。"
-              : error instanceof Error
-                ? error.message
-                : "wechat_login_failed";
-      this.renderView();
-    }
-  }
-
-  private primaryLoginProvider(): CocosLoginProviderDescriptor {
-    return (
-      this.loginProviders.find((provider) => provider.id === "wechat-mini-game" && provider.available) ??
-      this.loginProviders.find((provider) => provider.id === "account-password") ?? {
-        id: "account-password",
-        label: this.authMode === "account" ? "账号进入" : "账号登录并进入",
-        available: true,
-        message: ""
-      }
-    );
-  }
-
-  private describeLobbyLoginHint(): string {
-    const primaryProvider = this.primaryLoginProvider();
-    if (primaryProvider.id === "wechat-mini-game") {
-      return this.authProvider === "wechat-mini-game" ? "当前已使用小游戏登录脚手架会话" : primaryProvider.message;
-    }
-
-    return this.authMode === "account" ? "当前已处于正式账号模式" : "H5 绑定后的登录 ID 可以在这里直接进入";
-  }
-
-  private async returnToLobby(): Promise<void> {
-    if (this.showLobby) {
-      return;
-    }
-
-    const storage = this.readWebStorage();
-    saveCocosLobbyPreferences(this.playerId, this.roomId, undefined, storage);
-    this.displayName = rememberPreferredCocosDisplayName(this.playerId, this.displayName || this.playerId, storage);
-    this.stopMatchmakingPolling();
-    this.updateMatchmakingStatus({ status: "idle" });
-    await this.disposeCurrentSession();
-    this.resetSessionViewport("已返回 Cocos Lobby。");
-    this.gameplayAccountReviewPanelOpen = false;
-    this.gameplayBattlePassPanelOpen = false;
-    this.gameplayDailyDungeonPanelOpen = false;
-    this.gameplaySeasonalEventPanelOpen = false;
-    this.gameplayEquipmentPanelOpen = false;
-    this.gameplayCampaignPanelOpen = false;
-    this.gameplayCampaign = null;
-    this.gameplayCampaignSelectedMissionId = null;
-    this.gameplayCampaignActiveMissionId = null;
-    this.gameplayCampaignDialogue = null;
-    this.gameplayCampaignPendingAction = null;
-    this.gameplayCampaignStatus = "战役面板待同步。";
-    this.seasonProgress = null;
-    this.showLobby = true;
-    this.syncWechatShareBridge();
-    this.lobbyStatus = "已返回大厅，可继续选房或创建新实例。";
-    this.syncBrowserRoomQuery(null);
-    this.renderView();
-    await this.syncLobbyBootstrap();
-  }
-
-  private toggleSettingsPanel(open = !this.settingsView.open): void {
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      open,
-      deleteAccountPending: false,
-      withdrawConsentPending: false,
-      statusMessage: open ? null : this.settingsView.statusMessage
-    });
-    this.renderView();
-  }
-
-  private updateSettings(update: CocosSettingsPanelUpdate): void {
-    const resetPending = update.bgmVolume !== undefined || update.sfxVolume !== undefined || update.frameRateCap !== undefined;
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      ...update,
-      ...(resetPending ? { deleteAccountPending: false, withdrawConsentPending: false } : {})
-    });
-    this.persistSettings();
-    this.applyRuntimeSettings();
-    this.renderView();
-  }
-
-  private openSettingsPrivacyPolicy(): void {
-    const wxRuntime = (globalThis as {
-      wx?: {
-        openPrivacyContract?: (options?: { success?: () => void; fail?: (error?: unknown) => void }) => void;
-      } | null;
-    }).wx;
-    if (wxRuntime?.openPrivacyContract) {
-      wxRuntime.openPrivacyContract({
-        success: () => {
-          this.updateSettings({ statusMessage: "已打开微信隐私说明。" });
-        },
-        fail: () => {
-          this.updateSettings({ statusMessage: `隐私说明入口 ${this.settingsView.privacyPolicyUrl}` });
-        }
-      });
-      return;
-    }
-
-    const open = (globalThis as { open?: (url: string, target?: string) => void }).open;
-    if (typeof open === "function") {
-      open(this.settingsView.privacyPolicyUrl, "_blank");
-      this.updateSettings({ statusMessage: `已打开隐私说明 ${this.settingsView.privacyPolicyUrl}` });
-      return;
-    }
-
-    this.updateSettings({ statusMessage: `隐私说明 ${this.settingsView.privacyPolicyUrl}` });
-  }
-
-  private async handleSettingsLogout(): Promise<void> {
-    this.updateSettings({ statusMessage: "正在退出当前会话..." });
-    await this.logoutAuthSession();
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      open: false,
-      deleteAccountPending: false,
-      withdrawConsentPending: false,
-      statusMessage: "已退出当前会话。"
-    });
-    this.renderView();
-  }
-
-  private async handleSettingsDeleteAccount(): Promise<void> {
-    if (!this.settingsView.deleteAccountPending) {
-      this.updateSettings({
-        deleteAccountPending: true,
-        withdrawConsentPending: false,
-        statusMessage: "再次点击“删除账号”确认删除当前账号。"
-      });
-      return;
-    }
-
-    this.updateSettings({
-      statusMessage: "正在删除当前账号并撤销会话...",
-      deleteAccountPending: false
-    });
-
-    await this.deleteCurrentPlayerAccount();
-  }
-
-  private async handleSettingsWithdrawConsent(): Promise<void> {
-    if (!this.settingsView.withdrawConsentPending) {
-      this.updateSettings({
-        withdrawConsentPending: true,
-        deleteAccountPending: false,
-        statusMessage: "再次点击“撤回同意”以清除本地同意状态并退出当前会话。"
-      });
-      return;
-    }
-
-    this.privacyConsentAccepted = false;
-    this.updateSettings({
-      withdrawConsentPending: false,
-      statusMessage: "已撤回本地隐私同意，正在退出当前会话..."
-    });
-    await this.logoutAuthSession();
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      open: false,
-      statusMessage: "已撤回本地隐私同意；下次进入前请重新确认隐私说明。"
-    });
-    this.renderView();
-  }
-
-  private buildSupportTicketDraft(category: "bug" | "payment" | "account"): string {
-    const headline =
-      category === "bug"
-        ? "Cocos 客户端 BUG 反馈"
-        : category === "payment"
-          ? "支付与商品问题反馈"
-          : "账号与客服问题反馈";
-    const lines = [
-      headline,
-      `玩家：${this.displayName || this.playerId}`,
-      `玩家 ID：${this.playerId}`,
-      `登录方式：${this.authMode === "account" ? this.loginId || "account" : this.authMode}`,
-      `房间：${this.roomId}`,
-      `客户端：Cocos`,
-      "问题描述："
-    ];
-    return lines.join("\n");
-  }
-
-  private async handleSettingsSupportTicket(category: "bug" | "payment" | "account"): Promise<void> {
-    const authSession = this.currentLobbyAuthSession();
-    if (!authSession?.token) {
-      this.updateSettings({ statusMessage: "客服工单需要先登录云端账号或游客会话。" });
-      return;
-    }
-
-    this.supportTicketSubmittingCategory = category;
-    this.updateSettings({
-      supportSubmittingCategory: category,
-      statusMessage: "正在提交客服工单..."
-    });
-
-    try {
-      const result = await resolveVeilRootRuntime().submitSupportTicket(
-        this.remoteUrl,
-        {
-          category,
-          message: this.buildSupportTicketDraft(category),
-          priority: category === "payment" ? "high" : "normal"
-        },
-        {
-          authSession,
-          storage: this.readWebStorage()
-        }
-      );
-      const successMessage = `客服工单已提交：${result.ticket.ticketId}`;
-      this.lobbyStatus = successMessage;
-      this.updateSettings({ statusMessage: successMessage });
-    } catch (error) {
-      this.updateSettings({
-        statusMessage: error instanceof Error ? error.message : "客服工单提交失败。"
-      });
-    } finally {
-      this.supportTicketSubmittingCategory = null;
-      this.settingsView = applySettingsUpdate(this.settingsView, {
-        supportSubmittingCategory: null
-      });
-      this.renderView();
-    }
-  }
-
-  private async logoutAuthSession(): Promise<void> {
-    this.stopMatchmakingPolling();
-    this.updateMatchmakingStatus({ status: "idle" });
-    await resolveVeilRootRuntime().logoutAuthSession(this.remoteUrl, {
-      storage: this.readWebStorage()
-    });
-    this.authToken = null;
-    this.authMode = "guest";
-    this.authProvider = "guest";
-    this.loginId = "";
-    this.sessionSource = "none";
-    this.displayName = readPreferredCocosDisplayName(this.playerId, this.readWebStorage());
-    this.commitAccountProfile(createFallbackCocosPlayerAccountProfile(this.playerId, this.roomId, this.displayName), false);
-    this.syncWechatShareBridge();
-    this.lobbyStatus = "已退出当前会话，请重新选择游客身份或使用正式账号进入。";
-    this.renderView();
-  }
-
-  private async deleteCurrentPlayerAccount(): Promise<void> {
-    this.stopMatchmakingPolling();
-    this.updateMatchmakingStatus({ status: "idle" });
-    await resolveVeilRootRuntime().deletePlayerAccount(this.remoteUrl, {
-      storage: this.readWebStorage()
-    });
-    this.authToken = null;
-    this.authMode = "guest";
-    this.authProvider = "guest";
-    this.loginId = "";
-    this.sessionSource = "none";
-    this.privacyConsentAccepted = false;
-    this.displayName = readPreferredCocosDisplayName(this.playerId, this.readWebStorage());
-    this.commitAccountProfile(createFallbackCocosPlayerAccountProfile(this.playerId, this.roomId, this.displayName), false);
-    await this.disposeCurrentSession();
-    this.resetSessionViewport("账号已删除。");
-    this.showLobby = true;
-    this.lobbyStatus = "账号已删除，原会话已撤销。请重新确认隐私说明后再创建新档。";
-    this.syncBrowserRoomQuery(null);
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      open: false,
-      deleteAccountPending: false,
-      withdrawConsentPending: false,
-      statusMessage: "账号已删除。"
-    });
-    this.renderView();
-    await this.syncLobbyBootstrap();
-  }
-
-  private buildActiveAccountFlowPanelView(): CocosAccountLifecyclePanelView | CocosAccountRegistrationPanelView | null {
-    const draft = this.buildActiveAccountLifecycleDraft();
-    if (!draft) {
-      return null;
-    }
-
-    if (draft.kind === "registration") {
-      const registrationDraft: CocosAccountLifecycleDraft & { kind: "registration" } = {
-        ...draft,
-        kind: "registration"
-      };
-      const wechatProvider = this.loginProviders.find((provider) => provider.id === "wechat-mini-game");
-      const registeredAccount =
-        this.authMode === "account" || Boolean(this.lobbyAccountProfile.credentialBoundAt)
-          ? {
-              ...(this.lobbyAccountProfile.loginId ?? this.loginId
-                ? { loginId: this.lobbyAccountProfile.loginId ?? this.loginId }
-                : {}),
-              ...(this.lobbyAccountProfile.credentialBoundAt
-                ? { credentialBoundAt: this.lobbyAccountProfile.credentialBoundAt }
-                : {}),
-              provider: this.authProvider
-            }
-          : null;
-      return buildCocosAccountRegistrationPanelView({
-        draft: registrationDraft,
-        privacyConsentAccepted: this.privacyConsentAccepted,
-        submitState: this.lobbyEntering
-          ? this.activeAccountFlow === "registration" && this.authProvider === "wechat-mini-game"
-            ? "binding-wechat"
-            : this.registrationToken && this.registrationPassword
-              ? "confirming-registration"
-              : "requesting-token"
-          : this.authMode === "account" && !this.activeAccountFlow
-            ? "success"
-            : "idle",
-        statusMessage: this.lobbyStatus,
-        showValidationErrors: false,
-        ...(registeredAccount ? { registeredAccount } : {}),
-        wechat: {
-          supported: this.runtimePlatform === "wechat-game",
-          available: wechatProvider?.available === true,
-          bound: this.authProvider === "wechat-mini-game",
-          minorProtectionSelection: this.wechatMinorProtectionSelection
-        }
-      });
-    }
-
-    return buildCocosAccountLifecyclePanelView(draft);
-  }
-
-  private buildActiveAccountLifecycleDraft(): CocosAccountLifecycleDraft | null {
-    if (!this.activeAccountFlow) {
-      return null;
-    }
-
-    return this.activeAccountFlow === "registration"
-      ? {
-          kind: "registration",
-          loginId: this.loginId,
-          displayName: this.registrationDisplayName || this.displayName || this.loginId,
-          token: this.registrationToken,
-          password: this.registrationPassword,
-          deliveryMode: this.registrationDeliveryMode,
-          ...(this.registrationExpiresAt ? { expiresAt: this.registrationExpiresAt } : {})
-        }
-      : {
-          kind: "recovery",
-          loginId: this.loginId,
-          displayName: "",
-          token: this.recoveryToken,
-          password: this.recoveryPassword,
-          deliveryMode: this.recoveryDeliveryMode,
-          ...(this.recoveryExpiresAt ? { expiresAt: this.recoveryExpiresAt } : {})
-        };
-  }
-
-  private openLobbyAccountFlow(kind: CocosAccountLifecycleKind): void {
-    this.activeAccountFlow = kind;
-    this.loginId = this.loginId.trim().toLowerCase();
-    if (kind === "registration" && !this.registrationDisplayName.trim()) {
-      this.registrationDisplayName = this.displayName || this.loginId;
-      this.wechatMinorProtectionSelection = "unknown";
-    }
-    this.lobbyStatus =
-      kind === "registration"
-        ? "已打开正式注册面板。先申请注册令牌，再确认口令并进入房间。"
-        : "已打开密码找回面板。先申请找回令牌，再确认新口令并进入房间。";
-    this.renderView();
-  }
-
-  private closeLobbyAccountFlow(): void {
-    this.wechatMinorProtectionSelection = "unknown";
-    this.activeAccountFlow = null;
-    this.lobbyStatus = "已收起账号生命周期面板。";
-    this.renderView();
-  }
-
-  private toggleWechatMinorProtectionSelection(): void {
-    if (this.activeAccountFlow !== "registration") {
-      return;
-    }
-
-    this.wechatMinorProtectionSelection =
-      this.wechatMinorProtectionSelection === "unknown"
-        ? "adult"
-        : this.wechatMinorProtectionSelection === "adult"
-          ? "minor"
-          : "adult";
-    this.lobbyStatus =
-      this.wechatMinorProtectionSelection === "adult"
-        ? "已声明为成年人，绑定微信时将按成年人策略提交。"
-        : "已声明为未成年人，绑定微信时将按未成年人策略提交。";
-    this.renderView();
-  }
-
-  private async bindWechatIdentityFromLobbyAccountFlow(): Promise<void> {
-    if (this.activeAccountFlow !== "registration" || this.authMode !== "account") {
-      this.lobbyStatus = "需先完成正式注册或登录正式账号，才能绑定微信身份。";
-      this.renderView();
-      return;
-    }
-
-    if (this.wechatMinorProtectionSelection === "unknown") {
-      this.lobbyStatus = "绑定微信身份前，请先设置未成年人保护声明。";
-      this.renderView();
-      return;
-    }
-
-    await this.loginLobbyWechatMiniGame();
-  }
-
-  private togglePrivacyConsent(): void {
-    this.privacyConsentAccepted = !this.privacyConsentAccepted;
-    this.lobbyStatus = this.privacyConsentAccepted ? "已同意隐私说明。" : "已取消隐私说明勾选。";
-    this.renderView();
-  }
-
-  private promptForAccountFlowField(field: "loginId" | "displayName" | "token" | "password"): void {
-    const promptRef = globalThis.prompt;
-    if (typeof promptRef !== "function" || !this.activeAccountFlow) {
-      this.lobbyStatus = "当前运行环境不支持弹出式输入，请改用浏览器调试壳填写流程字段。";
-      this.renderView();
-      return;
-    }
-
-    if (field === "loginId") {
-      const nextValue = promptRef("输入登录 ID", this.loginId)?.trim();
-      if (nextValue === undefined) {
-        return;
-      }
-      this.loginId = nextValue.toLowerCase();
-      this.lobbyStatus = this.loginId ? `已更新登录 ID 草稿为 ${this.loginId}。` : "已清空登录 ID 草稿。";
-      this.renderView();
-      return;
-    }
-
-    if (field === "displayName") {
-      const nextValue = promptRef("输入注册昵称", this.registrationDisplayName || this.displayName || this.loginId);
-      if (nextValue === null) {
-        return;
-      }
-      this.registrationDisplayName = nextValue.trim();
-      this.lobbyStatus = this.registrationDisplayName ? "已更新注册昵称草稿。" : "已清空注册昵称草稿。";
-      this.renderView();
-      return;
-    }
-
-    if (field === "token") {
-      const nextValue = promptRef(
-        this.activeAccountFlow === "registration" ? "输入注册令牌" : "输入找回令牌",
-        this.activeAccountFlow === "registration" ? this.registrationToken : this.recoveryToken
-      )?.trim();
-      if (nextValue === undefined) {
-        return;
-      }
-      if (this.activeAccountFlow === "registration") {
-        this.registrationToken = nextValue;
-      } else {
-        this.recoveryToken = nextValue;
-      }
-      this.lobbyStatus = nextValue ? "已更新令牌草稿。" : "已清空令牌草稿。";
-      this.renderView();
-      return;
-    }
-
-    const nextValue = promptRef(
-      this.activeAccountFlow === "registration" ? "输入注册口令（至少 6 位）" : "输入新的账号口令（至少 6 位）",
-      ""
-    );
-    if (nextValue === null) {
-      return;
-    }
-    if (this.activeAccountFlow === "registration") {
-      this.registrationPassword = nextValue.trim();
-    } else {
-      this.recoveryPassword = nextValue.trim();
-    }
-    this.lobbyStatus = nextValue.trim() ? "已更新口令草稿。" : "已清空口令草稿。";
-    this.renderView();
-  }
-
-  private async requestActiveAccountFlow(): Promise<void> {
-    if (!this.activeAccountFlow || this.lobbyEntering) {
-      return;
-    }
-    const loginId = this.loginId.trim().toLowerCase();
-    const validationError = validateAccountLifecycleRequest(this.activeAccountFlow, loginId);
-    if (validationError) {
-      this.lobbyStatus = validationError.message;
-      this.renderView();
-      return;
-    }
-
-    this.loginId = loginId;
-    this.lobbyEntering = true;
-    this.lobbyStatus =
-      this.activeAccountFlow === "registration"
-        ? `正在为 ${loginId} 申请注册令牌...`
-        : `正在为 ${loginId} 申请密码找回令牌...`;
-    this.renderView();
-
-    try {
-      if (this.activeAccountFlow === "registration") {
-        const requested = await requestCocosAccountRegistration(
-          this.remoteUrl,
-          loginId,
-          this.registrationDisplayName || this.displayName || loginId
-        );
-        this.registrationToken = requested.registrationToken ?? this.registrationToken;
-        this.registrationExpiresAt = requested.expiresAt ?? "";
-        this.registrationDeliveryMode = requested.registrationToken ? "dev-token" : "external";
-        this.lobbyStatus = requested.registrationToken
-          ? `注册令牌已生成，可直接确认注册。令牌：${requested.registrationToken}${requested.expiresAt ? `；过期时间：${requested.expiresAt}` : ""}`
-          : `注册申请已受理，请从外部渠道获取令牌${requested.expiresAt ? `；过期时间：${requested.expiresAt}` : ""}。`;
-      } else {
-        const requested = await requestCocosPasswordRecovery(this.remoteUrl, loginId);
-        this.recoveryToken = requested.recoveryToken ?? this.recoveryToken;
-        this.recoveryExpiresAt = requested.expiresAt ?? "";
-        this.recoveryDeliveryMode = requested.recoveryToken ? "dev-token" : "external";
-        this.lobbyStatus = requested.recoveryToken
-          ? `找回令牌已生成，可直接确认重置。令牌：${requested.recoveryToken}${requested.expiresAt ? `；过期时间：${requested.expiresAt}` : ""}`
-          : `找回申请已受理，请从外部渠道获取令牌${requested.expiresAt ? `；过期时间：${requested.expiresAt}` : ""}。`;
-      }
-    } catch (error) {
-      this.lobbyStatus =
-        this.activeAccountFlow === "registration"
-          ? this.describeCocosAccountFlowError(error, "account_registration_request_failed")
-          : this.describeCocosAccountFlowError(error, "password_recovery_request_failed");
-    } finally {
-      this.lobbyEntering = false;
-      this.renderView();
-    }
-  }
-
-  private async confirmActiveAccountFlow(): Promise<void> {
-    if (!this.activeAccountFlow || this.lobbyEntering) {
-      return;
-    }
-    const loginId = this.loginId.trim().toLowerCase();
-    this.loginId = loginId;
-    const validationError = validateAccountLifecycleConfirm(this.activeAccountFlow, {
-      loginId,
-      token: this.activeAccountFlow === "registration" ? this.registrationToken : this.recoveryToken,
-      password: this.activeAccountFlow === "registration" ? this.registrationPassword : this.recoveryPassword,
-      privacyConsentAccepted: this.privacyConsentAccepted
-    });
-    if (validationError) {
-      this.lobbyStatus = validationError.message;
-      this.renderView();
-      return;
-    }
-
-    if (this.activeAccountFlow === "registration") {
-      await this.confirmLobbyAccountRegistration(loginId);
-      return;
-    }
-
-    await this.confirmLobbyAccountRecovery(loginId);
-  }
-
-  private async confirmLobbyAccountRegistration(loginId: string): Promise<void> {
-    this.lobbyEntering = true;
-    this.lobbyStatus = `正在确认正式注册 ${loginId} 并进入房间 ${this.roomId}...`;
-    this.renderView();
-
-    try {
-      const storage = this.readWebStorage();
-      const authSession = await confirmCocosAccountRegistration(
-        this.remoteUrl,
-        loginId,
-        this.registrationToken,
-        this.registrationPassword,
-        {
-          storage,
-          privacyConsentAccepted: this.privacyConsentAccepted
-        }
-      );
-      this.authToken = authSession.token ?? null;
-      this.playerId = authSession.playerId;
-      this.displayName = authSession.displayName;
-      this.authMode = authSession.authMode;
-      this.authProvider = authSession.provider ?? "account-password";
-      this.loginId = authSession.loginId ?? loginId;
-      this.sessionSource = authSession.source;
-      this.registrationToken = "";
-      this.registrationPassword = "";
-      this.registrationDeliveryMode = "idle";
-      this.registrationExpiresAt = "";
-      this.activeAccountFlow = null;
-      this.syncWechatShareBridge();
-      this.lobbyStatus = `正式账号注册成功，正在同步全局仓库并进入房间 ${this.roomId}...`;
-      saveCocosLobbyPreferences(authSession.playerId, this.roomId, undefined, storage);
-      this.renderView();
-      await this.refreshLobbyAccountProfile();
-      this.lobbyEntering = false;
-      await this.enterLobbyRoom(this.roomId);
-    } catch (error) {
-      this.lobbyEntering = false;
-      this.lobbyStatus = this.describeCocosAccountFlowError(error, "account_registration_failed", {
-        invalidTokenCode: "invalid_registration_token"
-      });
-      this.renderView();
-    }
-  }
-
-  private async confirmLobbyAccountRecovery(loginId: string): Promise<void> {
-    this.lobbyEntering = true;
-    this.lobbyStatus = `正在重置 ${loginId} 的口令并进入房间 ${this.roomId}...`;
-    this.renderView();
-
-    try {
-      await confirmCocosPasswordRecovery(this.remoteUrl, loginId, this.recoveryToken, this.recoveryPassword);
-      const storage = this.readWebStorage();
-      const authSession = await loginWithCocosProvider(
-        this.remoteUrl,
-        {
-          provider: "account-password",
-          loginId,
-          password: this.recoveryPassword
-        },
-        {
-          storage,
-          privacyConsentAccepted: this.privacyConsentAccepted
-        }
-      );
-      this.authToken = authSession.token ?? null;
-      this.playerId = authSession.playerId;
-      this.displayName = authSession.displayName;
-      this.authMode = authSession.authMode;
-      this.authProvider = authSession.provider ?? "account-password";
-      this.loginId = authSession.loginId ?? loginId;
-      this.sessionSource = authSession.source;
-      this.recoveryToken = "";
-      this.recoveryPassword = "";
-      this.recoveryDeliveryMode = "idle";
-      this.recoveryExpiresAt = "";
-      this.activeAccountFlow = null;
-      this.syncWechatShareBridge();
-      this.lobbyStatus = `口令重置成功，正在同步全局仓库并进入房间 ${this.roomId}...`;
-      saveCocosLobbyPreferences(authSession.playerId, this.roomId, undefined, storage);
-      this.renderView();
-      await this.refreshLobbyAccountProfile();
-      this.lobbyEntering = false;
-      await this.enterLobbyRoom(this.roomId);
-    } catch (error) {
-      this.lobbyEntering = false;
-      this.lobbyStatus = this.describeCocosAccountFlowError(error, "password_recovery_failed", {
-        invalidTokenCode: "invalid_recovery_token"
-      });
-      this.renderView();
-    }
-  }
-
-  private promptForLobbyField(field: "playerId" | "displayName" | "roomId" | "loginId"): void {
-    const promptRef = globalThis.prompt;
-    if (typeof promptRef !== "function") {
-      this.lobbyStatus = "当前运行环境不支持弹出式输入，请改用 URL 参数、已缓存会话或浏览器调试壳。";
-      this.renderView();
-      return;
-    }
-
-    const storage = this.readWebStorage();
-    if (field === "playerId") {
-      const previousSuggestedName = readPreferredCocosDisplayName(this.playerId, storage);
-      const nextValue = promptRef("输入游客 playerId", this.playerId)?.trim();
-      if (nextValue === undefined) {
-        return;
-      }
-
-      const nextPlayerId = nextValue || createCocosGuestPlayerId();
-      const storedSession = readStoredCocosAuthSession(storage);
-      this.playerId = nextPlayerId;
-      if (!this.displayName.trim() || this.displayName === previousSuggestedName) {
-        this.displayName =
-          storedSession?.playerId === nextPlayerId ? storedSession.displayName : readPreferredCocosDisplayName(nextPlayerId, storage);
-      }
-      this.authToken = storedSession?.playerId === nextPlayerId ? storedSession.token ?? null : null;
-      this.authMode = storedSession?.playerId === nextPlayerId ? storedSession.authMode : "guest";
-      this.authProvider = storedSession?.playerId === nextPlayerId ? storedSession.provider ?? "guest" : "guest";
-      this.loginId = storedSession?.playerId === nextPlayerId ? storedSession.loginId ?? "" : "";
-      this.sessionSource = storedSession?.playerId === nextPlayerId ? storedSession.source : "manual";
-      this.syncWechatShareBridge();
-      this.lobbyStatus = `已切换游客身份草稿为 ${nextPlayerId}。`;
-      this.renderView();
-      void this.refreshLobbyAccountProfile();
-      return;
-    }
-
-    if (field === "displayName") {
-      const nextValue = promptRef("输入展示昵称", this.displayName || this.playerId);
-      if (nextValue === null) {
-        return;
-      }
-
-      this.displayName = rememberPreferredCocosDisplayName(this.playerId, nextValue, storage);
-      this.syncWechatShareBridge();
-      this.lobbyStatus = "昵称草稿已更新。";
-      this.renderView();
-      void this.refreshLobbyAccountProfile();
-      return;
-    }
-
-    if (field === "loginId") {
-      const nextValue = promptRef("输入登录 ID", this.loginId)?.trim();
-      if (nextValue === undefined) {
-        return;
-      }
-
-      this.loginId = nextValue.toLowerCase();
-      this.lobbyStatus = this.loginId ? `已更新登录 ID 草稿为 ${this.loginId}。` : "已清空登录 ID 草稿。";
-      this.renderView();
-      return;
-    }
-
-    const nextValue = promptRef("输入房间 ID", this.roomId)?.trim();
-    if (nextValue === undefined || nextValue.length === 0) {
-      return;
-    }
-
-    this.roomId = nextValue;
-    this.syncWechatShareBridge();
-    this.lobbyStatus = `已将目标房间切换为 ${nextValue}。`;
-    this.renderView();
-    void this.refreshLobbyAccountProfile();
-  }
-
-  private async disposeCurrentSession(): Promise<void> {
-    await disposeCurrentSessionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private resetSessionViewport(logLine: string): void {
-    resetSessionViewportForRoot(this as unknown as Record<string, any>, logLine);
-  }
-
-  private async handleForcedUpgrade(failureMessage: string): Promise<void> {
-    await handleForcedUpgradeForRoot(this as unknown as Record<string, any>, failureMessage);
-  }
-
-  private describeSessionError(error: unknown, fallback: string): string {
-    return describeSessionErrorForRoot(error, fallback);
-  }
-
-  private pushSessionActionOutcome(
-    update: SessionUpdate,
-    options: {
-      successMessage: string;
-      rejectedLabel: string;
-    }
-  ): void {
-    const outcome = describeSessionActionOutcome(update, options);
-    this.pushLog(outcome.message);
-    if (!outcome.accepted) {
-      this.predictionStatus = outcome.message;
-      this.mapBoard?.playHeroAnimation("hit");
-    }
-  }
-
-  private bumpSessionEpoch(): number {
-    return bumpSessionEpochForRoot(this as unknown as Record<string, any>);
-  }
-
-  private bumpLobbyAccountEpoch(): number {
-    return bumpLobbyAccountEpochForRoot(this as unknown as Record<string, any>);
-  }
-
-  private isActiveSessionEpoch(epoch: number): boolean {
-    return isActiveSessionEpochForRoot(this as unknown as Record<string, any>, epoch);
-  }
-
-  private isActiveLobbyAccountEpoch(epoch: number): boolean {
-    return isActiveLobbyAccountEpochForRoot(this as unknown as Record<string, any>, epoch);
-  }
-
-  private createSessionOptions(epoch: number): VeilCocosSessionOptions {
-    return createSessionOptionsForRoot(this as unknown as Record<string, any>, epoch);
-  }
-
-  private hydrateRuntimePlatform(): void {
-    this.runtimePlatform = detectCocosRuntimePlatform(globalThis as {
-      location?: Location;
-      history?: History;
-      wx?: {
-        getLaunchOptionsSync?: () => { query?: Record<string, unknown> | null } | null | undefined;
-        login?: ((options: unknown) => void) | undefined;
-      };
-    });
-    this.runtimeCapabilities = resolveCocosRuntimeCapabilities(this.runtimePlatform);
-    this.loginRuntimeConfig = resolveCocosLoginRuntimeConfig(globalThis as never);
-    this.loginProviders = resolveCocosLoginProviders({
-      platform: this.runtimePlatform,
-      capabilities: this.runtimeCapabilities,
-      config: this.loginRuntimeConfig,
-      wx: (globalThis as { wx?: { login?: ((options: unknown) => void) | undefined } }).wx ?? null
-    });
-
-    if (this.runtimePlatform === "wechat-game") {
-      this.pushLog("已识别微信小游戏运行时，启动参数将改读 wx.getLaunchOptionsSync().query。");
-      const wechatProvider = this.loginProviders.find((provider) => provider.id === "wechat-mini-game");
-      if (wechatProvider) {
-        this.pushLog(`小游戏登录状态：${wechatProvider.message}`);
-      }
-    }
-  }
-
-  private hydrateClientPerfRuntimeMetadata(): void {
-    this.clientPerfRuntimeMetadata = readClientPerfRuntimeMetadata(globalThis as {
-      wx?: {
-        getSystemInfoSync?: (() => { model?: unknown; version?: unknown } | null | undefined) | undefined;
-      } | null;
-    });
-  }
-
-  private bindRuntimeMemoryWarnings(): void {
-    this.stopRuntimeMemoryWarnings?.();
-    this.stopRuntimeMemoryWarnings = bindCocosRuntimeMemoryWarning((event) => {
-      const gcTriggered = triggerCocosRuntimeGc();
-      this.runtimeMemoryNotice =
-        event.level != null
-          ? `收到内存告警 L${event.level}${gcTriggered ? "，已请求 GC" : ""}`
-          : `收到内存告警${gcTriggered ? "，已请求 GC" : ""}`;
-      this.pushLog(this.runtimeMemoryNotice);
-      this.renderView();
-    });
-  }
-
-  private describeRuntimeMemoryHealth(): string {
-    const snapshot = readCocosRuntimeMemorySnapshot();
-    const summary = formatCocosRuntimeMemoryStatus(snapshot, getPlaceholderSpriteAssetUsageSummary());
-    return this.runtimeMemoryNotice ? `${summary} · ${this.runtimeMemoryNotice}` : summary;
-  }
-
-  private trackClientPerfTelemetry(deltaTime: number): void {
-    const nowMs = Date.now();
-    recordClientPerfFrame(this.clientPerfTelemetry, deltaTime, nowMs);
-
-    const memorySnapshot = readCocosRuntimeMemorySnapshot();
-    const memoryUsageRatio =
-      memorySnapshot.heapUsedBytes != null && memorySnapshot.heapLimitBytes != null && memorySnapshot.heapLimitBytes > 0
-        ? memorySnapshot.heapUsedBytes / memorySnapshot.heapLimitBytes
-        : memorySnapshot.heapUsedBytes != null && memorySnapshot.heapTotalBytes != null && memorySnapshot.heapTotalBytes > 0
-          ? memorySnapshot.heapUsedBytes / memorySnapshot.heapTotalBytes
-          : null;
-
-    const payload = evaluateClientPerfTelemetry(this.clientPerfTelemetry, {
-      nowMs,
-      memoryUsageRatio,
-      metadata: this.clientPerfRuntimeMetadata
-    });
-    if (!payload) {
-      return;
-    }
-
-    this.trackClientAnalyticsEvent("client_perf_degraded", payload);
-  }
-
-  private hydrateLaunchIdentity(): void {
-    hydrateLaunchIdentityForRoot(this as unknown as Record<string, any>);
-  }
-
-  private latestShareableBattleReplay() {
-    const replay = this.lobbyAccountProfile?.recentBattleReplays?.[0] ?? null;
-    if (!shouldOfferBattleResultShare(replay)) {
-      return null;
-    }
-    if (!this.lastBattleSettlementSnapshot || this.lastBattleSettlementSnapshot.tone !== "victory") {
-      return null;
-    }
-    return replay;
-  }
-
-  private canShareLatestBattleResult(): boolean {
-    return Boolean(this.latestShareableBattleReplay());
-  }
-
-  private async handleBattleResultShare(): Promise<void> {
-    const replay = this.latestShareableBattleReplay();
-    if (!replay) {
-      this.predictionStatus = "当前没有可分享的 PVP 胜利战报。";
-      this.renderView();
-      return;
-    }
-
-    const wxRuntime = (globalThis as {
-      wx?: {
-        shareAppMessage?: (sharePayload: WechatSharePayload) => void;
-      } | null;
-    }).wx;
-    const shareResult = await shareBattleResultForRuntime(replay, this.displayName || this.playerId, {
-      runtimePlatform: this.runtimePlatform,
-      ...(wxRuntime !== undefined ? { wechatRuntime: wxRuntime } : {})
-    });
-    this.predictionStatus = shareResult.message;
-    this.renderView();
-  }
-
-  private async maybeClaimLaunchReferral(authSession: {
-    playerId: string;
-    displayName: string;
-    authMode: "guest" | "account";
-    provider?: string;
-    loginId?: string;
-    token?: string;
-    source: "remote" | "local";
-  }): Promise<void> {
-    const referrerId = this.launchReferrerId?.trim() ?? "";
-    if (!referrerId || authSession.source !== "remote" || !authSession.token) {
-      return;
-    }
-
-    const claimKey = `${referrerId}:${authSession.playerId}`;
-    if (this.lastReferralClaimKey === claimKey) {
-      return;
-    }
-
-    try {
-      const result = await resolveVeilRootRuntime().postPlayerReferral(
-        this.remoteUrl,
-        { referrerId },
-        {
-          storage: this.readWebStorage(),
-          authSession: {
-            token: authSession.token,
-            playerId: authSession.playerId,
-            displayName: authSession.displayName,
-            authMode: authSession.authMode,
-            ...(authSession.provider ? { provider: authSession.provider as never } : {}),
-            ...(authSession.loginId ? { loginId: authSession.loginId } : {}),
-            source: "remote"
-          }
-        }
-      );
-      this.lastReferralClaimKey = claimKey;
-      if (result.claimed) {
-        this.pushLog(`已完成推荐奖励绑定：邀请人 ${referrerId} 与新玩家 ${authSession.playerId} 各获得 20 宝石。`);
-      }
-    } catch (error) {
-      if (error instanceof Error && error.message === "cocos_request_failed:409:referral_already_claimed") {
-        this.lastReferralClaimKey = claimKey;
-        return;
-      }
-      throw error;
-    }
-  }
-
-  private readLaunchSearch(): string {
-    return readCocosRuntimeLaunchSearch(globalThis as {
-      location?: Pick<Location, "search"> | null;
-      wx?: { getLaunchOptionsSync?: () => { query?: Record<string, unknown> | null } | null | undefined };
-    });
-  }
-
-  private describeLobbyShareHint(): string {
-    return `分享：${this.wechatShareStatus}`;
-  }
-
-  private hydrateSettings(): void {
-    const wxRuntime = (globalThis as { wx?: unknown }).wx as { getStorageSync?: (key: string) => unknown } | null | undefined;
-    const persisted = readPersistedCocosSettings({
-      localStorage: this.readWebStorage(),
-      ...(wxRuntime ? { wx: wxRuntime } : {})
-    });
-    this.settingsView = applySettingsUpdate(this.settingsView, {
-      ...persisted,
-      privacyPolicyUrl: resolveCocosPrivacyPolicyUrl(globalThis.location)
-    });
-    this.applyRuntimeSettings();
-  }
-
-  private applyRuntimeSettings(): void {
-    this.audioRuntime.setBgmVolume(this.settingsView.bgmVolume);
-    this.audioRuntime.setSfxVolume(this.settingsView.sfxVolume);
-    const gameRuntime = (globalThis as {
-      game?: {
-        frameRate?: number;
-        setFrameRate?: (value: number) => void;
-      };
-    }).game;
-    if (typeof gameRuntime?.setFrameRate === "function") {
-      gameRuntime.setFrameRate(this.settingsView.frameRateCap);
-    } else {
-      const fallbackRuntime = (gameRuntime ?? (globalThis as { frameRate?: number })) as { frameRate?: number };
-      fallbackRuntime.frameRate = this.settingsView.frameRateCap;
-    }
-  }
-
-  private persistSettings(): void {
-    const wxRuntime = (globalThis as { wx?: unknown }).wx as { setStorageSync?: (key: string, value: string) => void } | null | undefined;
-    writePersistedCocosSettings(
-      {
-        bgmVolume: this.settingsView.bgmVolume,
-        sfxVolume: this.settingsView.sfxVolume,
-        frameRateCap: this.settingsView.frameRateCap
-      },
-      {
-        localStorage: this.readWebStorage(),
-        ...(wxRuntime ? { wx: wxRuntime } : {})
-      }
-    );
-  }
-
-  private buildTutorialOverlayView(): TutorialOverlayView | null {
-    return buildTutorialOverlayViewForRoot(this as unknown as Record<string, any>);
-  }
-
-  private resolveTutorialCampaignGuidance(): TutorialCampaignGuidance {
-    return resolveTutorialCampaignGuidanceForRoot(this as unknown as Record<string, any>);
-  }
-
-  private resolveTutorialGuidanceMission(): NonNullable<CocosCampaignSummary["missions"]>[number] | null {
-    return resolveTutorialGuidanceMissionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async submitTutorialProgress(action: TutorialProgressAction): Promise<void> {
-    await submitTutorialProgressForRoot(this as unknown as Record<string, any>, action);
-  }
-
-  private async advanceTutorialFlow(): Promise<void> {
-    await advanceTutorialFlowForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async handleTutorialPrimaryAction(): Promise<void> {
-    await handleTutorialPrimaryActionForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async completeTutorialAndFocusCampaign(): Promise<void> {
-    await completeTutorialAndFocusCampaignForRoot(this as unknown as Record<string, any>);
-  }
-
-  private async skipTutorialFlow(): Promise<void> {
-    await skipTutorialFlowForRoot(this as unknown as Record<string, any>);
-  }
-
-  private buildSettingsView(): CocosSettingsPanelView {
-    return applySettingsUpdate(this.settingsView, {
-      displayName: this.displayName || this.playerId,
-      loginId: this.loginId,
-      authMode: this.authMode,
-      privacyConsentAccepted: this.privacyConsentAccepted,
-      supportSubmittingCategory: this.supportTicketSubmittingCategory,
-      privacyPolicyUrl: resolveCocosPrivacyPolicyUrl(globalThis.location)
-    });
-  }
-
-  private renderSettingsOverlay(): void {
-    this.settingsView = this.buildSettingsView();
-    this.settingsPanel?.render(this.settingsView);
-    this.renderSettingsButton();
-  }
-
-  private renderSettingsButton(): void {
-    const buttonNode = this.node.getChildByName(SETTINGS_BUTTON_NODE_NAME);
-    if (!buttonNode) {
-      return;
-    }
-
-    assignUiLayer(buttonNode);
-    const transform = buttonNode.getComponent(UITransform) ?? buttonNode.addComponent(UITransform);
-    const width = transform.width || 58;
-    const height = transform.height || 58;
-    const graphics = buttonNode.getComponent(Graphics) ?? buttonNode.addComponent(Graphics);
-    graphics.clear();
-    graphics.fillColor = this.settingsView.open ? new Color(108, 88, 54, 244) : new Color(52, 68, 92, 236);
-    graphics.strokeColor = this.settingsView.open ? new Color(244, 225, 180, 164) : new Color(226, 236, 248, 96);
-    graphics.lineWidth = 2;
-    graphics.roundRect(-width / 2, -height / 2, width, height, 16);
-    graphics.fill();
-    graphics.stroke();
-    graphics.fillColor = new Color(255, 255, 255, 18);
-    graphics.roundRect(-width / 2 + 10, height / 2 - 14, width - 20, 4, 2);
-    graphics.fill();
-
-    const labelNode = buttonNode.getChildByName("Label") ?? new Node("Label");
-    labelNode.parent = buttonNode;
-    assignUiLayer(labelNode);
-    const labelTransform = labelNode.getComponent(UITransform) ?? labelNode.addComponent(UITransform);
-    labelTransform.setContentSize(width - 8, height - 8);
-    labelNode.setPosition(0, 0, 1);
-    const label = labelNode.getComponent(Label) ?? labelNode.addComponent(Label);
-    label.string = "⚙";
-    label.fontSize = 26;
-    label.lineHeight = 28;
-    label.horizontalAlign = 1;
-    label.verticalAlign = 1;
-    label.enableWrapText = false;
-    label.color = new Color(244, 247, 252, 255);
-  }
-
-  private syncWechatShareBridge(immediate = false) {
-    const payload = buildCocosWechatSharePayload({
-      roomId: this.roomId,
-      inviterPlayerId: this.playerId,
-      displayName: this.displayName || this.playerId,
-      scene: this.showLobby ? "lobby" : this.lastUpdate?.battle ? "battle" : "world",
-      day: this.lastUpdate?.world.meta.day ?? null,
-      battleLabel: this.lastUpdate?.battle ? "当前战斗" : null
-    });
-
-    if (this.runtimePlatform !== "wechat-game") {
-      this.wechatShareAvailable = false;
-      this.wechatShareStatus = "分享功能仅在微信小游戏可用。";
-      return {
-        available: false,
-        menuEnabled: false,
-        handlerRegistered: false,
-        canShareDirectly: false,
-        immediateShared: false,
-        payload,
-        message: this.wechatShareStatus
-      };
-    }
-
-    const result = syncCocosWechatShareBridge(
-      (globalThis as { wx?: CocosWechatShareRuntimeLike | null }).wx ?? null,
-      payload,
-      immediate ? { immediate: true } : undefined
-    );
-    this.wechatShareAvailable = result.available;
-    this.wechatShareStatus = result.message;
-    return result;
-  }
-
-  private readWebStorage(): Storage | null {
-    const webStorage = (sys as unknown as { localStorage?: Storage }).localStorage;
-    return webStorage ?? null;
-  }
-
-  private pointInRootNode(centeredX: number, centeredY: number, node: Node | null): boolean {
-    if (!node || !node.active) {
-      return false;
-    }
-
-    const transform = node.getComponent(UITransform) ?? null;
-    if (!transform) {
-      return false;
-    }
-
-    return (
-      centeredX >= node.position.x - transform.width / 2
-      && centeredX <= node.position.x + transform.width / 2
-      && centeredY >= node.position.y - transform.height / 2
-      && centeredY <= node.position.y + transform.height / 2
-    );
-  }
-
-  private syncBrowserRoomQuery(roomId: string | null): void {
-    if (!this.runtimeCapabilities.supportsBrowserHistory) {
-      return;
-    }
-
-    const historyRef = globalThis.history;
-    const locationRef = globalThis.location;
-    if (!historyRef?.replaceState || !locationRef?.href) {
-      return;
-    }
-
-    const nextUrl = new URL(locationRef.href);
-    if (roomId?.trim()) {
-      nextUrl.searchParams.set("roomId", roomId.trim());
-      nextUrl.searchParams.delete("playerId");
-      nextUrl.searchParams.delete("displayName");
-    } else {
-      nextUrl.search = "";
-    }
-
-    historyRef.replaceState(null, "", nextUrl.toString());
-  }
-
-  private handleConnectionEvent(event: ConnectionEvent): void {
-    this.diagnosticsConnectionStatus =
-      event === "reconnecting" ? "reconnecting" : event === "reconnected" ? "connected" : "reconnect_failed";
-    if (event === "reconnect_failed") {
-      this.reportClientRuntimeError({
-        errorCode: "session_disconnect",
-        severity: "error",
-        stage: "reconnect",
-        recoverable: true,
-        message: "Client reconnect failed; falling back to room snapshot recovery."
-      });
-    }
-    const activePvpBattle = Boolean(this.lastUpdate?.battle?.defenderHeroId);
-    const label =
-      event === "reconnecting"
-        ? activePvpBattle
-          ? "PVP 遭遇连接已中断，正在尝试重连..."
-          : "连接已中断，正在尝试重连..."
-        : event === "reconnected"
-          ? activePvpBattle
-            ? "PVP 遭遇连接已恢复。"
-            : "连接已恢复。"
-          : activePvpBattle
-            ? "PVP 遭遇重连失败，正在尝试恢复房间快照..."
-            : "重连失败，正在尝试恢复房间快照...";
-    if (this.showLobby) {
-      this.lobbyStatus = label;
-    }
-    this.pushLog(label);
-    this.renderView();
-  }
-
-  private async actInBattle(action: BattleAction): Promise<void> {
-    if (!this.session || this.battleActionInFlight) {
-      return;
-    }
-
-    this.battleActionInFlight = true;
-    const actionPresentation = this.battlePresentation.previewAction(action, this.lastUpdate?.battle ?? null);
-    const skillName =
-      action.type === "battle.skill"
-        ? this.lastUpdate?.battle?.units[action.unitId]?.skills?.find((skill) => skill.id === action.skillId)?.name ?? action.skillId
-        : null;
-    const actionLabel =
-      action.type === "battle.attack"
-        ? "攻击"
-        : action.type === "battle.wait"
-          ? "等待"
-          : action.type === "battle.defend"
-            ? "防御"
-            : skillName ?? "技能";
-    this.pushLog(`战斗指令：${actionLabel}`);
-    this.emitPrimaryClientTelemetry(
-      createPrimaryClientTelemetryEvent(this.createTelemetryContext(this.activeHero()?.id ?? null), {
-        category: "combat",
-        checkpoint: "command.submitted",
-        status: "info",
-        detail: `Battle command submitted: ${actionLabel}.`,
-        ...(this.lastUpdate?.battle?.id ? { battleId: this.lastUpdate.battle.id } : {})
-      })
-    );
-    this.setBattleFeedback(actionPresentation.feedback);
-    if (actionPresentation.cue) {
-      this.audioRuntime.playCue(actionPresentation.cue);
-    }
-    this.renderView();
-
-    try {
-      if (actionPresentation.animation !== "idle") {
-        this.mapBoard?.playHeroAnimation(actionPresentation.animation);
-      }
-
-      const update = await this.session.actInBattle(action);
-      await this.applySessionUpdate(update);
-      if (update.reason) {
-        this.pushSessionActionOutcome(update, {
-          successMessage: "战斗指令已结算。",
-          rejectedLabel: "战斗指令"
-        });
-      }
-    } catch (error) {
-      const reason = error instanceof Error ? error.message : "battle_action_failed";
-      const detail = error instanceof Error ? error.message : "战斗操作失败。";
-      this.pushLog(detail);
-      this.emitPrimaryClientTelemetry(
-        createPrimaryClientTelemetryEvent(this.createTelemetryContext(this.activeHero()?.id ?? null), {
-          category: "combat",
-          checkpoint: "command.rejected",
-          status: "failure",
-          detail,
-          reason,
-          ...(this.lastUpdate?.battle?.id ? { battleId: this.lastUpdate.battle.id } : {})
-        })
-      );
-      this.mapBoard?.playHeroAnimation("hit");
-    } finally {
-      this.battleActionInFlight = false;
-      this.renderView();
-    }
-  }
-
-  private async applySessionUpdate(update: SessionUpdate): Promise<void> {
-    const previousBattle = this.lastUpdate?.battle ?? null;
-    const heroId = this.activeHero()?.id ?? null;
-    const presentation = this.battlePresentation.applyUpdate(previousBattle, update, heroId);
-    this.captureBattleSettlementSnapshot(update, presentation.state);
-
-    this.pendingPrediction = null;
-    this.predictionStatus = "";
-    this.surrenderDialogOpen = false;
-    this.surrenderStatusMessage = null;
-    this.diagnosticsConnectionStatus = "connected";
-    this.lastRoomUpdateSource = "session";
-    this.lastRoomUpdateReason = update.reason ?? "snapshot";
-    this.lastRoomUpdateAtMs = Date.now();
-    this.lastUpdate = update;
-    this.maybeOpenGameplayEquipmentPanelForLoot(update);
-    const eventEntries = buildTimelineEntriesFromUpdate(update);
-    if (eventEntries.length > 0) {
-      this.timelineEntries = collapseAdjacentEntries([...eventEntries, ...this.timelineEntries]).slice(0, 12);
-    }
-    this.emitPrimaryClientTelemetry(
-      buildPrimaryClientTelemetryFromUpdate(update, this.createTelemetryContext(heroId))
-    );
-    for (const event of update.events) {
-      const ownsEventHero =
-        "heroId" in event && typeof event.heroId === "string"
-          ? update.world.ownHeroes.some((hero) => hero.id === event.heroId)
-          : false;
-
-      if (event.type === "battle.started" && ownsEventHero) {
-        this.trackClientAnalyticsEvent("battle_start", {
-          roomId: update.world.meta.roomId,
-          battleId: event.battleId,
-          encounterKind: event.encounterKind,
-          heroId: event.heroId
-        }, update.world.meta.roomId);
-      }
-
-      if (event.type === "battle.resolved" && ownsEventHero) {
-        this.trackClientAnalyticsEvent("battle_end", {
-          roomId: update.world.meta.roomId,
-          battleId: event.battleId,
-          result: event.result,
-          heroId: event.heroId,
-          battleKind: "battleKind" in event && (event.battleKind === "neutral" || event.battleKind === "hero")
-            ? event.battleKind
-            : previousBattle?.defenderHeroId
-              ? "hero"
-              : "neutral"
-        }, update.world.meta.roomId);
-      }
-    }
-    if (update.events.some((event) => event.type === "battle.resolved")) {
-      this.syncGameplayCampaignBattleOutcome(update);
-      void this.submitBattleProgressForActiveEvents(update);
-    }
-    if (shouldRefreshGameplayAccountProfileForEvents(update.events.map((event) => event.type))) {
-      void this.refreshGameplayAccountProfile();
-    }
-    this.syncSelectedBattleTarget();
-    this.renderView();
-    if (presentation.pauseDurationMs) {
-      await new Promise<void>((resolve) => {
-        globalThis.setTimeout(resolve, presentation.pauseDurationMs ?? 0);
-      });
-    }
-    this.playMapFeedbackForUpdate(update);
-    this.maybeShowHeroProgressNotice(update);
-    this.setBattleFeedback(presentation.feedback, presentation.feedbackDurationMs ?? BATTLE_FEEDBACK_DURATION_MS);
-    if (presentation.cue) {
-      this.audioRuntime.playCue(presentation.cue);
-    }
-    this.mapBoard?.playHeroAnimation(presentation.animation);
-
-    if (update.reason && isSessionSettlementReason(update.reason)) {
-      this.predictionStatus = formatSessionSettlementReason(update.reason, !this.surrenderSubmitting);
-    }
-
-    if (presentation.transition?.kind === "enter") {
-      await this.battleTransition?.playEnter(presentation.transition.copy);
-    } else if (presentation.transition?.kind === "exit") {
-      await this.battleTransition?.playExit(presentation.transition.copy);
-    }
-
-    this.syncWechatShareBridge();
-    this.renderView();
-  }
-
-  private maybeOpenGameplayEquipmentPanelForLoot(update: SessionUpdate): void {
-    if (this.showLobby || update.battle) {
-      return;
-    }
-
-    const controlledHeroIds = new Set(update.world.ownHeroes.map((hero) => hero.id));
-    const hasOwnedLoot = update.events.some(
-      (event) => event.type === "hero.equipmentFound" && controlledHeroIds.has(event.heroId)
-    );
-    if (!hasOwnedLoot) {
-      return;
-    }
-
-    this.gameplayEquipmentPanelOpen = true;
-    this.gameplayCampaignPanelOpen = false;
-  }
-
-  private async refreshGameplayAccountProfile(): Promise<void> {
-    if (this.gameplayAccountRefreshInFlight) {
-      return;
-    }
-
-    if (this.sessionSource !== "remote") {
-      return;
-    }
-
-    if (!this.remoteUrl?.trim()) {
-      return;
-    }
-
-    this.gameplayAccountRefreshInFlight = true;
-    try {
-      const authSession = this.authToken
-        ? {
-            token: this.authToken,
-            playerId: this.playerId,
-            displayName: this.displayName || this.playerId,
-            authMode: this.authMode,
-            ...(this.loginId ? { loginId: this.loginId } : {}),
-            source: "remote" as const
-          }
-        : null;
-      const [profile, activeEvents] = await Promise.all([
-        resolveVeilRootRuntime().loadAccountProfile(this.remoteUrl, this.playerId, this.roomId, {
-          storage: this.readWebStorage(),
-          authSession
-        }),
-        authSession?.token
-          ? resolveVeilRootRuntime().loadActiveSeasonalEvents(this.remoteUrl, {
-              storage: this.readWebStorage(),
-              authSession
-            })
-          : Promise.resolve([])
-      ]);
-      this.commitAccountProfile(profile, true);
-      this.activeSeasonalEvent = activeEvents[0] ?? null;
-      if (this.activeSeasonalEvent) {
-        this.seasonalEventStatus = `已同步 ${this.activeSeasonalEvent.name} · 当前积分 ${this.activeSeasonalEvent.player.points}`;
-      }
-      this.renderView();
-    } finally {
-      this.gameplayAccountRefreshInFlight = false;
-    }
-  }
-
-  private maybeShowHeroProgressNotice(update: SessionUpdate): void {
-    const heroId = update.world.ownHeroes[0]?.id ?? null;
-    const notice = buildHeroProgressNotice(update, heroId);
-    if (!notice) {
-      return;
-    }
-
-    this.levelUpNotice = {
-      ...notice,
-      expiresAt: Date.now() + 5000
-    };
-    this.pushLog(`${notice.title}。${notice.detail}`);
-    this.mapBoard?.playHeroAnimation("victory");
-    this.audioRuntime.playCue("level_up");
-  }
-
-  private commitAccountProfile(profile: CocosPlayerAccountProfile, allowAchievementNotice: boolean): void {
-    const previousProfile = this.lobbyAccountProfile;
-    if (profile.playerId !== this.lobbyAccountProfile.playerId) {
-      this.seenProfileNoticeEventIds.clear();
-    }
-
-    if (allowAchievementNotice) {
-      const notice = buildCocosProfileNotice(profile.recentEventLog, this.seenProfileNoticeEventIds);
-      if (notice) {
-        this.achievementNotice = {
-          ...notice,
-          expiresAt: Date.now() + 4000
-        };
-        this.pushLog(`${notice.title}：${notice.detail}`);
-      }
-    }
-
-    for (const eventId of collectProfileNoticeEventIds(profile.recentEventLog)) {
-      this.seenProfileNoticeEventIds.add(eventId);
-    }
-
-    this.lobbyAccountProfile = profile;
-    this.maybeEmitExperimentExposureAnalytics(profile);
-    this.maybeEmitQuestCompleteAnalytics(previousProfile, profile);
-    this.seasonProgress = this.snapshotSeasonProgressFromProfile();
-    if (
-      this.sessionSource === "remote"
-      && this.authMode === "account"
-      && this.authToken
-      && normalizeTutorialStep(profile.tutorialStep) !== null
-      && !this.gameplayCampaign
-      && !this.gameplayCampaignLoading
-    ) {
-      void this.refreshGameplayCampaign();
-    }
-    if (
-      this.showLobby
-      && this.sessionSource === "remote"
-      && this.authMode === "account"
-      && this.authToken
-      && !this.dailyDungeonSummary
-      && !this.dailyDungeonLoading
-    ) {
-      void this.refreshDailyDungeonPanel();
-    }
-    this.lobbyAccountReviewState = transitionCocosAccountReviewState(this.lobbyAccountReviewState, {
-      type: "account.synced",
-      account: profile
-    });
-  }
-
-  private playMapFeedbackForUpdate(update: SessionUpdate): void {
-    const heroId = update.world.ownHeroes[0]?.id;
-    if (!heroId) {
-      return;
-    }
-
-    for (const entry of buildMapFeedbackEntriesFromUpdate(update, heroId)) {
-      this.mapBoard?.showTileFeedback(entry.position, entry.text, entry.durationSeconds);
-    }
-
-    for (const entry of buildObjectPulseEntriesFromUpdate(update, heroId)) {
-      this.mapBoard?.pulseObject(entry.position, entry.scale, entry.durationSeconds);
-    }
-  }
-
-  private applyReplayedSessionUpdate(update: SessionUpdate): void {
-    this.pendingPrediction = null;
-    this.predictionStatus = "已回放缓存状态，等待房间同步...";
-    this.lastRoomUpdateSource = "replay";
-    this.lastRoomUpdateReason = "cached_snapshot";
-    this.lastRoomUpdateAtMs = Date.now();
-    this.lastUpdate = {
-      ...update,
-      events: [],
-      movementPlan: null
-    };
-    this.renderView();
-  }
-
-  private buildBattleSettlementRecoveryState(): {
-    title: string;
-    detail: string;
-    badge: string;
-    tone: CocosBattleFeedbackView["tone"];
-    summaryLines: string[];
-  } | null {
-    return buildBattleSettlementRecoveryStateForRoot(this as unknown as Record<string, any>);
-  }
-
-  private captureBattleSettlementSnapshot(
-    update: SessionUpdate,
-    presentationState: CocosBattlePresentationState
-  ): void {
-    if (update.battle) {
-      this.lastBattleSettlementSnapshot = null;
-      return;
-    }
-
-    if (presentationState.phase === "resolution") {
-      this.lastBattleSettlementSnapshot = {
-        label: presentationState.label,
-        detail: presentationState.detail,
-        badge: presentationState.badge,
-        tone: presentationState.tone,
-        summaryLines: presentationState.summaryLines
-      };
-    }
-  }
-
-  private buildHudSessionIndicators(): VeilHudRenderState["sessionIndicators"] {
-    return buildHudSessionIndicatorsForRoot(this as unknown as Record<string, any>);
-  }
-
-  private applyPrediction(action: CocosWorldAction, status: string): void {
-    if (!this.lastUpdate) {
-      return;
-    }
-
-    const prediction = predictPlayerWorldAction(this.lastUpdate.world, action);
-    if (prediction.reason) {
-      return;
-    }
-
-    if (!this.pendingPrediction) {
-      this.pendingPrediction = cloneSessionUpdate(this.lastUpdate);
-    }
-
-    this.lastUpdate = {
-      ...this.lastUpdate,
-      world: prediction.world,
-      movementPlan: prediction.movementPlan,
-      reachableTiles: prediction.reachableTiles
-    };
-    this.predictionStatus = status;
-  }
-
-  private rollbackPrediction(reason?: string): void {
-    if (this.pendingPrediction) {
-      this.lastUpdate = this.pendingPrediction;
-      this.pendingPrediction = null;
-    }
-
-    this.predictionStatus = "";
-    if (reason) {
-      this.pushLog(reason);
-      this.mapBoard?.playHeroAnimation("hit");
-    }
-
-    this.renderView();
-  }
-
-  private syncMusicScene(): void {
-    if (this.showLobby) {
-      this.audioRuntime.setScene(null);
-      return;
-    }
-
-    if (this.lastUpdate?.battle) {
-      this.audioRuntime.setScene("battle");
-      return;
-    }
-
-    this.audioRuntime.setScene(this.lastUpdate?.world ? "explore" : null);
-  }
-
-  private buildHudPresentationState(): VeilHudRenderState["presentation"] {
-    return buildHudPresentationStateForRoot(this as unknown as Record<string, any>);
-  }
-
 }
 
-export { setVeilRootRuntimeForTests, resetVeilRootRuntimeForTests } from "./root/index.ts";
+function installVeilRootMethods(...sources: object[]): void {
+  for (const source of sources) {
+    const { constructor: _constructor, ...descriptors } = Object.getOwnPropertyDescriptors(source);
+    void _constructor;
+    Object.defineProperties(VeilRoot.prototype, descriptors);
+  }
+}
+
+installVeilRootMethods(
+  veilRootViewLayoutMethods,
+  veilRootPrefetchSchedulerMethods,
+  veilRootLobbyProgressionMethods,
+  veilRootReportReviewMethods,
+  veilRootWorldBattleActionsMethods,
+  veilRootBattleSessionUpdateMethods,
+  veilRootLobbyMatchmakingMethods,
+  veilRootAccountSettingsMethods,
+  veilRootRuntimeIntegrationMethods
+);
