+++ apps/cocos-client/assets/scripts/root/index.ts
@@ -3,8 +3,19 @@ export * from "./types";
 export * from "./runtime";
 export * from "./formatters";
 export * from "./session-helpers";
+export * from "./deps";
+export * from "./default-state";
 export * from "./session-lifecycle";
 export * from "./panel-orchestration";
 export * from "./render-state-composer";
+export * from "./view-layout";
+export * from "./prefetch-scheduler";
+export * from "./lobby-progression";
+export * from "./report-review";
+export * from "./world-battle-actions";
+export * from "./battle-session-update";
+export * from "./lobby-matchmaking";
+export * from "./account-settings";
+export * from "./runtime-integration";
 export * from "./telemetry-hooks";
 export * from "./tutorial-orchestrator";
